Parsing and deparsing chord information

These functions can be used to extract and "translate," or otherwise modify, data representing tertian harmony information. The functions are:

Arguments

x

An atomic vector.

The x argument can be any (atomic) vectors

Key

The diatonic key used by the parser, deparser, and transposer.

Defaults to NULL, which is interpreted as C major.

Must be a diatonicSet or something coercable to diatonicSet; must be either length 1 or length(x).

parseArgs

An optional list of arguments to the chord parser.

Defaults to an empty list().

Must be a list of named arguments to the chord parser.

transposeArgs

An optional list of arguments passed to a transpose() call.

Defaults to an empty list().

Must be a list of named arguments to transpose().

inPlace

Should non-chord information be retained in the output string.

Defaults to FALSE.

Must be a singleton logical value: an on/off switch.

Details

Jazz/Pop

chord()

harte()

Classical

figuredBass()

tertian()

Roman Numerals

harm()

roman()

See also