Figured bass representation of harmony

This function outputs a figured bass representation of a tertian harmony.

Usage

# Default S3 method
figuredBass(
  x,
  ...,
  Key = NULL,
  transposeArgs = list(),
  parseArgs = list(),
  inPlace = FALSE
)

Arguments

x

An atomic vector.

The x argument can be any (atomic) vectors

Key

The diatonic key used by the parser, deparser, and transposer.

Defaults to NULL, which is interpreted as C major.

Must be a diatonicSet or something coercable to diatonicSet; must be either length 1 or length(x).

transposeArgs

An optional list of arguments passed to a transpose() call.

Defaults to an empty list().

Must be a list of named arguments to transpose().

parseArgs

An optional list of arguments to the chord parser.

Defaults to an empty list().

Must be a list of named arguments to the chord parser.

inPlace

Should non-chord information be retained in the output string.

Defaults to FALSE.

Must be a singleton logical value: an on/off switch.

Examples

romanNumerals <- c('2I', '2IV7', '1V', '2vi', '2-VI', '2iio7', '2Vb9')

figuredBass(romanNumerals)
#> [1] "C|"        "F|7"       "G|"        "A|"        "A-|b5"     "D|7b5"    
#> [7] "F-|b6b4b2"

tertian <- c('CM', 'CMm/3', 'FM', 'Fm', 'D-MM', 'GMmm')

figuredBass(tertian)
#> [1] "C|"     "E|6b53" "F|"     "F|b3"   "D-|7b5" "G|b9"  

B075 <- readHumdrum(humdrumRroot, "HumdrumData/BeethovenVariations/B075_00_05_a.krn")
#> Finding and reading files...
#> 	REpath-pattern '/home/nat/.tmp/RtmpjOz08Z/temp_libpath13452e707787ad/humdrumR/HumdrumData/BeethovenVariations/B075_00_05_a.krn' matches 1 text files in 1 directory.
#> One file read from disk.
#> Validating one file...
#> all valid.
#> Parsing one file...
#> Assembling corpus...
#> Done!
with(B075[[ , 2]], figuredBass(Token))
#> [1] "F|"    "F|"    "e|653" "F|"    "B-|"   "c|7"   "F|"