"Fold" humdrumR data into new fields

Many humdrum datasets encode data across multiple spines, spine-paths, or stops. By default, humdrumR parses each separate spine, spine-path, and stop as their own individual data points, taking up one row in the humdrum table. If we want to treat data in multiple spines/paths/stops as different aspects of the same data it is easiest to reshape the data so that the information is in different humdrumR fields rather than separate spines/paths/stops. We "fold" the data from one structural location over "on top" of other data using foldHumdrum.

Usage

foldHumdrum(
  humdrumR,
  fold,
  onto,
  what = "Spine",
  Piece = NULL,
  fromField = getActiveFields(humdrumR)[1],
  fillFromField = FALSE,
  newFieldNames = NULL
)

foldPaths(
  humdrumR,
  fromField = getActiveFields(humdrumR)[1],
  fillFromField = TRUE
)

foldStops(
  humdrumR,
  fromField = getActiveFields(humdrumR)[1],
  fillFromField = FALSE
)

Arguments

humdrumR

HumdrumR data.

Must be a humdrumR data object.

fold

Which target structure (spine, path, etc.) to "fold" onto another structural position.

Must be natural numbers.

onto

Which target structure (spine, path, etc.) is the data "folded" onto.

Must be natural numbers.

what

The structural (spine, path, etc.) which is folded across.

Defaults to "Spine".

Must be a single character string. Valid options are "Spine", "Path", "Stop", "Record",and "NData".

Piece

Which pieces in the corpus should be folded (see "Piece-Specific Folding" section, below).

Defaults to NULL.

Must be natural numbers; must be length length(onto).

fromField

Which field to "fold."

Defaults to getActiveFields(humdrumR)[1].

Must be a character string partially matching the name of a data field in the humdrumR input. For example, "Tok" would match the Token field. This is the field which is "folded" into a new field.

fillFromField

Should the content of the fromField be to copied unfolded sections?

Defaults to FALSE for foldHumdrum() and foldStops(); TRUE for foldPaths().

Must be a singleton logical value: an on/off switch.

This only comes into play if the folding field is smaller than the to field.

newFieldNames

Names to use for new fields created by the folding.

Defaults to NULL.

Must be character.

From where to where

The numeric fold and onto arguments specify where to fold from/to. fold indicates the Spine/Path/Stop to fold from, "on to" the Spine/Path/Stop indicated by onto. For example, if you specify foldHumdrum(mydata, fold = 2, onto = 1, what = 'Spine') spine 2 will be folded "on top of" spine 1. The fold and onto targets may not overlap.

The fold and onto arguments can be vectors of any length, which are interpreted in parallel: for example, the combination fold = 1:2 and onto = 3:4 would map the first spine to the third spine (1 -> 3) and the second spine to the 4th spine (2 -> 4). If the onto targets are duplicated, the fold spines will be folded onto multiple new fields: for example, the combination fold = 1:2 and onto = c(3, 3) will map first spine and the second spine on to two new fields of the third spine. If the fold target is duplicated, the same fold spines can be copied onto multiple onto spines: for example, the combination fold = 1 and onto = 2:3 will map the contents of the first spine onto the second and third spine, duplicating the spine-1 data.

The lengths of fold and onto are automatically matched, so arguments like fold = 1:2 and onto = 3 are equivalent to fold = 1:2, onto = c(3, 3). This makes it easy to do things like "copy all four spines onto spine 1": just write fold = 2:4, onto = 1.

To specify what structural field you want to fold across, use the what argument (character, length == 1). The default what value is "Spine"; other common fold options are "Path", and "Stop", though you might want to use the convenient foldPaths() and foldStops() functions directly (details below). (You may also fold across "Record" or "NData"), but these are advanced/tricky!)

Which fields

The fromField (character, length == 1) controls which field in the fold spine/path/stop is folded into a new field. The fromField argument defaults to the (first) active field, and must match (or partially match) a field in the humdrumR argument data set. In some cases, the fold data is smaller than the onto data---for instance, spine paths often only exist for part of a spine, so there is less data in the path than in the full spine. In these cases, it can be helpful to set fillFromField == TRUE, which causes the missing parts of fold to be filled with data from the from field. foldPaths does this by default.

The resulting new fields will automatically be named as appropriate Results fields. The newFieldNames argument (character) can be used to control the output names: one for each new field created by the fold. If you specify too many newFieldNames, the later names are ignored. If you specify too few newFieldNames, the later names will be given result names, consistent with the default behavior.

Piece-Specific Folding

By default, the same "fold" is carried out in each piece in the input corpus (humdrumR argument). If you need to specify different folds in different pieces, you have to specify the Piece argument (numeric, whole number). For every piece in the corpus you want to apply folds to, you must specify all the fold and onto arguments in parallel vectors with the Piece argument (even if this is reduendant for some files). For example, if we specify the combinations,

fold	onto	Piece

1	2	1

3	4	1

1	2	2

4	3	2

then

In Piece one:

the first spine is mapped to the second spine

the third spine is mapped to the fourth spine

In Piece two:

the first spine is mapped to the second spine

the fourth spine is mapped to the third spine

If any files in the corpus are not included, they will not be affected at all!

Predefined folds

The convenient foldStops() and foldPaths() functions automatically fold all stops/paths in a dataset onto the first stop/path, creating new fields named, e.g., Path1, Path2, etc. Another extremely useful function is foldExclusive(), which automatically folds spines based on their exclusive interpretation.

See also