The "Active expression" of a humdrumR object.

This "Active" expression is used as the default value in a lot of humdrumR code. For one, it is the data which is printed by show calls, i.e., whenever you return a humdrumR object in the terminal. In any expression within a call to with(in)Humdrum . is automatically replaced with the Active expression.

evalActive(
  humdrumR,
  dataTypes = "D",
  forceVector = FALSE,
  sep = ", ",
  nullAs = NA
)

getActive(humdrumR)

setActive(humdrumR, form)

setActiveFields(humdrumR, fieldnames)

Arguments

humdrumR
A humdrumR data object.

dataTypes
Which dataTypes of humdrum records to include. Legal values are 'G', 'L', 'I', 'M', 'D', 'd', 'P' or any combination of these in a single string (e.g., "LIM"). (see the humdrum table documentation Fields section for an explanation.).

forceVector
logical. If TRUE, the result is forced to be an atomic vector.

sep
A length-one character string. If forceVector == TRUE this value is used as a separator between tokens that are collapsed.

nullAsDot
A single atomic value. Any null tokens are coerced to this value (default is .).

Details

The active expression can be changed with the commands setActive or the $ operator. This is a handy way to quickly look at different fields in your data.

The Active expression is often just the name of a field: for instance, the default value is Token. However, it can actually be any complex expression which evaluates within the humdrum table. For instance, the Active expression could be: paste0(Token, " ", Record), which would automatically print each Token with its record number pasted to it.

Null data

HumdrumR identifies "null data" based on the active field. Anywhere the current active field evaluates to "." or NA is considered Null data, and assigned the type "d" in the internal humdrum table. As you work, there will often be data tokens which are null in one field, but not in another field. For example, if you load **kern data, a token like "4r" (quarter-note rest) token will be NA if you call pitch, but not NA if you call recip (rhythm).

kerndata <- readHumdrum(...)

kerndata$Token %hum>% pitch -> kerndata$Pitch
kerndata$Token %hum>% recip -> kerndata$Rhythm

Now, if you change the active field between Pitch and Rhythm you'll see that there are different numbers of (non-null) data tokens: ntokens(kerndata$Pitch) vs ntokens(kerndata$Rhythm) will return different numbers! (The different would be the number of rest tokens.) Similarly, if you apply functions/expressions to this data (using withinHumdrum for example), the result will depend on what the active field is:

kerndata$Pitch %hum<% ~length(Token)
kerndata$Rhythm %hum<% ~length(Token)

Once again, we'll get different numbers here! (Assuming there are rests in the data.) This is the case even though the do-expression isn't actually using the Pitch or Rhythm fields! If Pitch is the active field the rest tokens are null-data and will be ignored!