humdrumR Coercion.

Many users may wish to work with humdrum data without using the humdrumR API, instead using basic R data types. For this purpose, humdrumR data objects can be coerced to basic R data types.

# S4 method for humdrumR
as.vector(x, mode = "any")

as.lines(
  humdrumR,
  dataTypes = "GLIMDd",
  fieldname = NULL,
  alignColumns = FALSE,
  padPaths = FALSE,
  padder = ""
)

as.matrix(humdata)

as.data.frame(humdata)

as.matrices(
  humdrumR,
  dataTypes = "D",
  fieldnames = NULL,
  padder = NA,
  path.fold = TRUE
)

as.data.frames(
  humdrumR,
  dataTypes = "D",
  fieldnames = NULL,
  padder = NA,
  path.fold = TRUE
)

Arguments

mode
If the mode argument is not 'any', it can be a single character string naming an atomic mode---the output will be coerced to this mode (if possible).

dataTypes
Which types of humdrum records to include. Legal values are 'G', 'L', 'I', 'M', 'D', 'd' or any combination of these (e.g., "LIM"). (see the humdrum table documentation Fields section for explanation.).

padder
An atomic value of length one. If par.files or pad.paths are true, the padder argument is used to fill in the desired gaps.

pad.files
logical (default TRUE). If any pieces in the humdrumR corpus have fewer spines/columns than the maximum, should they be padded with the padder argument (par.files == TRUE) or should an an error occur (pad.files == FALSE)? Note that these "padded" points are not represented in the original humdrum data.

pad.paths
logical If any spine path splits ('*^') occur in the humdrumR data, should they be padded with the padder argument (par.files == TRUE) or should an an error occur (pad.paths == FALSE)? Note that these "padded" points are not represented in the original humdrum data.

field(s)
If the field argument is not NULL, it can instead be a character string matching the humdrumR object's fields. If so, these fields are extracted instead of the Active expression. For calls to as.vector and as.data.frame, only one field can be extracted. However, for calls to as.matrix, multiple fields can be extraced---these fields will be returned in a third matrix dimension, each field forming one rectangular slice.

Details

as.vector(humdata, types, mode, fields) evaluates the humdrumR object's Active expression, and (attempts) to force the result to a vector of mode. This method is essentially a wrapper for evalActive.

as.matrix(humdata, types, pad.files, pad.paths) also evaluates the humdrumR object's Active expression, but wraps it into a matrix of dimensions c(nrow(humdata), ncol(humdata). Note that "Columns" in humdrum data are not necesarrily the same as spines.

as.data.frame(humdata) first calls as.matrix then converts the matrix to a data.frame. as.data.table(humdata) first calls as.matrix then converts the matrix to a data.table.

as.matrices, as.data.frames, and as.data.tables call as.matrix/as.data.frame/as.data.table on each individual file in a humdrumR corpus, returning a list of matices/data.frames/data.tables.