Piping humdrumR data

These infix operators make it possible to use the with(in)Humdrum functions in a piping style, similar to the | (pipe) in unix-style terminals, or the pipe operator from the R package magrittr. (In fact, the magrittr pipe is imported by humdrumR, and we encourage you to incorporate it into your humdrumR work flows.)

humdrumR %hum>% formula

humdrumR %hum<% formula

humdrumR %humT% formula

lhs %s>% rhs

list %hums>% formula

list %hums<% formula

list %humsT% formula

list %hums[]% formula

Details

The key is that the function withinHumdrum always returns a new humdrumR data object. Thus, you can always send the output of withinHumdrum back in to a new call of withinHumdrum. This is exactly what the %hum>% is for: on the left-hand side, input a humdrumR object, on the right-hand side suitable arguments to withinHumdrum (i.e., a formula, a function, or list of formulae, functions, and named arguments). You can then chain this call with another call to %hum>% and more withinHumdrum arguments.

%hum<% acts the same way as %hum>% except it calls withHumdrum. Since withHumdrum does not return a humdrumR, the output can't be piped any further (using %hum>% or %hum<%). Thus, \%hum<\% should only be used as the last step in a pipe---you would do this if you want to extract the last step in your pipe from the data's Humdrum Table into a normal vector or list of R data.

' \%humT\% creates a "T" in the pipe, applying the desired expression but not keeping the result---the unaltered humdrumR input object is returned. This works simply by replacing all do~ with doplot~ in a call to withinHumdrum. The purpose of this option, is if you want to apply expressions for their side effects, for instance, for plotting.

%hum[]% is similar to %hum>% except it apply the formulae on its right-hand side using filterHumdrum. Thus, it can be used to filter/index a humdrumR data object on the fly.

Plural pipes

In R we often apply the same function to a list of data. "Plural pipes" expand this idea to piping: take a list of data and pipe each element in the list to an expression/function. There are plural pipe versions of each singular pipe operator. Just add an "s" to make them plural:

%>% (singular) : %s>% (plural);

%hum>% (singular) : %hums>% (plural);

%hum<% (singular) : %hums<% (plural);

%humT% (singular) : %humsT% (plural);

%hum[]% (singular) : %hums[]% (plural);

Note: %s>% is an expansion of the magrittr::%>% pipe operator, which makes use of some clever meta-programming---we can't guarantee it will always behave the way we might like!

Examples

humdata <- readHumdrum('path*.krn')

#> Finding and reading files...
#> 	REpath-pattern 'path*.krn' matches 0 text files in 1 directory.
#> Zero files read from disk. 
#> No files to validate.

humdata %hum>% ~table(.)

#> Error: %hum>% pipe operator can only be called with humdrumR data on left side.

humdata %hum>%
     c(by~Spine, do ~ table(.)) %hum>%
     c(mfcol ~ c(2,2), doplot~barplot(.))

#> Error: %hum>% pipe operator can only be called with humdrumR data on left side.

     
humdata %hum>%
     c(by ~ Spine, do ~ table(.)) %hum<%
     (do ~ sort(.))    

#> Error: %hum>% pipe operator can only be called with humdrumR data on left side.