humdrumR size and shape

These functions can be used to quickly get basic information about the size and "shape" of a humdrumR corpus. For more details, use the census function.

HumdrumR objects can be divided into "subcorpora." These functions tell us if there are any subcorpora and, if so, what they are called.

nrecords(humdrumR, dataTypes = "D")

ntokens(humdrumR, dataTypes = "D")

npieces(humdrumR)

anySubcorpora(humdrumR)

namesSubcorpora(humdrumR)

nfiles(humdrumR)

length(humdata)

# S4 method for humdrumR
nrow(x)

is.empty(humdrumR)

anyPaths(humdrumR)

anyStops(humdrumR)

Details

A few common base R methods are defined as synonyms for the humdrumR-specific sizing functions: length(humdata) is equivalent to npieces(humdata); nrow(humdata) is shortand for nrecords(., dataTypes = 'LIMDd') (i.e., local records only). ncol(humdata) returns the maximum value of the Column fields---the maximum number of tab-delineated columns in the humdrum files (irrespective of Spines/Paths). The results of nrow and ncol will match up with the dimensions of matrices/data.frames produced by calls to as.matrix/as.data.frame. dim(humdata) returns c(nrow(humdata), ncol(humdata)), as usual.

is.empty(humdata) asks if ntokens(humdata, dataTypes = 'D') == 0L.