Indexing humdrumR objects

R's built-in indexing operators, [] (single brakcets) and [[]] (double brackets) can be used as shortcuts for common calls to subset.humdrumR(), allowing you to filter out specific files, spines, or records.

Usage

humdata[] # returns unchanged

humdata[x:y]

humdata['regex']

humdata[[x:y]]

humdata[[ , x:y]]

humdata[['regex']]

humdata[[ , 'regex']]

humdata[[x:y, l:m]]

Arguments

x

HumdrumR data to index.

Must be a humdrumR data object.

i

Index for vectors or matrix/data.frame rows.

A numeric vector or a character string treated as a regular expression.

j

Index for matrix/data.frame columns.

A numeric vector or a character string treated as a regular expression.

Details

In R, the basic indexing operators, [] and [[]], are used to select subsets of data. For many data types (for instance, base R lists) the [single brackets] are used for "shallower" extraction while the [[double brackets]] are used for "deeper" extraction. HumdrumR corpus indexing follows this same basic pattern: [single brackets] are used to index humdrumR objects by piece while [[double brackets]] are used to index within pieces. (Accidentally writing [] when you need [[]] is a very common error, so watch out!)

Whether, indexing by piece or within, humdrumR objects can use two types of indexing arguments:

By numeric (ordinal integers)

By character string (regular expressions)

For more powerful/flexible indexing options, use subset directly.

Numeric indexing:

Indexing humdrumR corpora with [single brackets] will accept one numeric argument. This argument will be used to pick pieces within the humdrumR object ordinally. Thus, humdata[1:10] will select the first ten pieces in the data while humdata[42] will select only the 42nd piece.

Indexing humdrumR objects with [[double brackets]] will accept one or two numeric arguments, i and j, either of which can be used in isolation or in combination. (If j is used in isolation, it must be named or placed after a comma, as in humdata[[ , j ]].)

i is used to index data records (i.e., based on the humtable Record field). Thus, humdata[[1:20]] indexes the first twenty records from each file in the corpus, and humdata[[42]] extracts the 42nd record from each file.

j is used to index spines (i.e., based on the Spine field). Thus, humdata[[ , 3:4]] returns the third and fourth spines from each file in the corpus.

When indexing humdrumR corpora with numbers, all numeric (double) inputs are converted to integers.

Since subset always renumbers files/spines that remain after filtering/indexing, humdrumR indexing is entirely ordinal. For example,

humsubset <- humdata[11:20]
humsubset[2]

will return the 12th piece from the original humdata object. This is because the first call to [] returns the 11th through 20th pieces, which are renumbered 1:10 and the second index call returns the new 2nd index, which was the 12th originally. Similarly,

humsubset2 <- humdata[[ , 2:4]]
humsubset2[[ , 2]]

will return the third spine from the original data.

As in normal R indexing, negative numbers can be used, causing corresponding elements to be removed instead of retained. Thus, humdata[-3:-5] will remove the third, fourth, and fifth pieces from the data while humdata[[ , -3:-5]] will remove the third, fourth, and fifth spines from each piece. Positive and negative indices cannot be mixed in a single argument.

In all cases, indices outside of range (or of value 0) are ignored. E.g., if you have a corpus of twenty files and you call corpus[21], there is no 21st piece, so 21 is "out of range". If all your input indices are 0 and error will result. If all your input indices are out of range then an empty humdrumR object is returned. For instance, humdata[[401:500, ]] will return an empty humdrumR object if there are no pieces with more than 400 data records.

Character indexing:

Indexing humdrumR objects with [single brackets] will accept one vector of character strings. These strings are treated as regular expressions (regexes).

The tokens from the humdrumR object's active field(s) are searched for matches to any of the regular expressions you input. Any piece that contains any match to any of the regular expressions is retained---all other pieces are filtered out. Note that (because this is [single-bracket] indexing) the entire piece is retained, even if there is only one match. If no matches occur in any pieces, an empty humdrumR object is returned.

Indexing humdrumR objects with [[double brackets]] will accept one or two vectors of character strings, i and j, either of which can be used in isolation or in combination. (If j is used in isolation, it must be placed after a comma, as in humdata[[ , j]].) These strings are treated as regular expressions (regexes). The tokens from the humdrumR object's active field(s) are searched for matches to any of the regular expressions you input. Any record which contains at least one token matching any regex in i will be retained. Similarly, any spine which contains at least one token matching any regex in j is retained. If i and j are used together, matching spines (j) are indexed first, so that tokens matching the regular expression(s) in i must be found in the matching spines.

removeEmpty

By default, calls to indexing operators will completely remove data which you are filtering out. However, if you set the removeEmpty argument to FALSE, the filtered data is set to NULL, but not actually removed from the data object. (See subset.humdrumR() for more details.)

See also