Change or insert values in an expression

This function can be used to modify arguments to a functions within an existing expression (or quosure/formula).

interpolateArguments(quo, namedArgs)

Arguments

namedArgs
A list of named arguments. Unnamed arguments are simply ignored.

expr
A unevaluated expression object.

Details

interpolateArguments inteprets named value in its namedArgs argument in one of two ways: If the named value is a list, it interprets the name of the list as a function call, and inserts/swaps any arguments in that list into any instances of that function call within the expr. Named arguments are inserted or substituted if already present in expression. Unnamed argmuments are simply added to the call. Examples:

myexpr <- quote(dnorm(x, mean = 5))
interpolateArguments(myexpr, list(dnorm = list(mean = 2, sd = 5, TRUE)))

# result is new expresson: dnorm(x, mean = 2, sd = 5, TRUE)

If a named valued in the namedArgs argument is not a list, that name/value pair is substituted anywhere it is present in the expression. This approach is often more conscise, but arguments cannot be added to an expression this way, only substituted if already present. Examples:

myexpr <- quote(dnorm(x, mean = 5))
interpolateArguments(myexpr, mean = 2)

# result is new expression: dnorm(x, mean = 2)

Examples

myexpr2 <- quote(A + b*x + rnorm(length(a), mean(Z), sd = 2))

interpolateArguments(myexpr2,
                     list(sd = 10, mean = list(na.rm = TRUE)))

#> Error: `quo` must be a quosure

                       
# result is new expression: 
# a + b*x + rnorm(length(a), mean(Z, na.rm = TRUE), sd = 10)