Test the properties of tonal information

These functions test basic properties of pitch information. is.simple returns TRUE if pitch information is constrained in one octave. is.generic returns TRUE if pitch information is "natural" to the key (Key) argument.

Usage

is.simple(x, ...)

# S3 method for class 'tonalInterval'
is.simple(x, octave.round = floor, ...)

# Default S3 method
is.simple(x, ...)

is.generic(x, Key, ...)

# S3 method for class 'tonalInterval'
is.generic(x, Key = NULL)

Arguments

x

Pitch information.

Must be something that can be parsed as pitch information.

...

Parameters passed to tonalInterval().

octave.round

The rounding function.

Must be a rouding function.

Controls how simple intervals are interpreted relative to C.

Key

The onic key used to defined generic pitches.

Defaults to NULL.

Must be something that can be parsed as a onic key; must be either length 1 or length(x).

Details

These functions can be called directly on tonalIntervals; If called on anything else, the functions first calls the tonalInterval() parser. If any values fail to parse NA is returned.

"Simple" intervals fall in a particular octave relative to middle-C/unison. The octave.floor argument can be used to change how this works: The default option, floor, interprets the (**kern) pitches c, d, e, f, g, a, and b to be "simple." The most common alternative, round, identifies G, A, B, c, d, e, and f as "simple." See the pitch deparsing docs for a more detailed explanation.

"Generic" intervals belong to a key.

See also