Calculate melodic intervals

mint calculates melodic intervals in a vector, or across spine/paths of a humdrumR data object. A vector is interpreted as an ordered sequence of notes, forming a "melody," and the intervals between successive pitches are calculated.

Usage

mint(x, ...)

# S3 method for default
mint(
  x,
  lag = 1,
  deparser = interval,
  initial = kern,
  bracket = TRUE,
  classify = FALSE,
  ...,
  parseArgs = list(),
  Exclusive = NULL,
  Key = NULL,
  boundaries = list()
)

# S3 method for data.frame
mint(x, ...)

# S3 method for matrix
mint(x, margin = 2, ...)

# S3 method for humdrumR
mint(
  x,
  field = getActiveFields(x)[1],
  dataTypes = "D",
  ...,
  newField = paste0(field, "_mint")
)

Details

Input vector x is parsed as pitch information. (Parsing arguments can be passed via the parseArgs list, or parse(...) sugar. Key and Exclusive are also passed to the parser.)

The parsed pitch vector is copied and lagged using lag(), and pairs which cross boundaries are ignored. The melodic intervals are then "deparsed" into a standard representation; by default, the intervals() representation is used, but you can set the deparser argument to any pitch function. However, the only alternative deparser that would be commonly used (other than intervals()) would be semits().

The lag argument controls how far apart in the melody intervals are calculated. For instance, a lag of 2 will calculate intervals between every other note in the vector. Positive lags will calculate approaching intervals: each token represents the interval between the current note and the previous note. Negative lags will calculate departing intervals: each token reprseents the interval between the current note and the next note.

Note that you by passing directed = FALSE through the the deparser, the undirected (absolute value) of the melodic intervals can be returned.

Initial value padding

Any output of mint is necessarily padded by abs(lag) undefined intervals at the beginning (positive lag) or end (negative lag). The initial argument controls how these initial values are presented.

If initial is a function, the "initial" pitches are parsed "absolutely" relative to middle-C; in this case, initial should be another pitch function to deparse these pitches. The default is the kern() function. If bracket == TRUE, these initial values are are surrounded with [], so they are easier to distinguish from the actual melodic intervals.

If initial is an atomic value, these value are used as the padder; An atomic (vector) initial must be the same length as abs(lag). For example, you could set initial = 'start' to label these locations as the character string 'start'. If initial is NULL, the initial values are simply padded with NA.

Interval classification

If the classify argument is set to TRUE, intervals are classified as either "Unison", "Step", "Skip", or "Leap". Alternatively, skips can be interpreted as leaps by setting skips = FALSE. (Note that classification will only work if deparser = interval, which is the default).

By default, intervals are categorized tonally, meaning that the interval in tonal steps is used as the basis of classification. For example, an augmented 2nd is a step, and a diminished 3rd is a skip/leap. This means that augmented and diminished unisons are marked "Unison" as well! However, if directed = TRUE, augmented/diminished unisons will be marked with + or - to indicate direction, whereas perfect unisons are never marked with +/-.

Alternatively, you may choose to categorize intervals atonally by setting atonal = TRUE. If so, intervals are categorized based only on semitone (enharmonic) intervals: D# and Eb are classified the same.

Boundaries

In many cases we want to perform lagged calculations in a vector, but not across certain boundaries. For example, if your vector includes data from multiple pieces, we wouldn't want to calculate melodic intervals between pieces, only within pieces. The boundaries argument indicates one, or more, grouping vectors, which break the x (input) argument into groups. If more than boundaries vectors are given, a change in any vector indicates a boundary.

Value pairs which cross between groups are treated as if they were at the beginning. Basically, using boundaries should be essentially identical to using tapply(x, boundaries, laggedFunction, ...), except generally faster when the number of groups is large.

The most common use case in humdrum data, is looking at "melodies" within spines. For this, we want boundaries = list(File, Spine, Path ). In fact, humdrumR with(in) calls will automatically feed these three fields as boundaries arguments to certain functions: mint, hint, delta, sigma, lag, or ditto. Do any use of delta in a call to with(in), will automatically calculate the delta in a "melodic" way, within each spine path of each file. However, if you wanted, for instance, to calculate differences across spines (like harmonic intervals) you could manually set boundaries = list(File, Record).

See also