Find and read humdrum files into R

These functions find valid humdrum files on your local machine and read them into humdrumR.

findHumdrum(
  ...,
  contains = NULL,
  recursive = FALSE,
  allowDuplicates = FALSE,
  verbose = FALSE
)

readHumdrum(
  ...,
  recursive = FALSE,
  contains = NULL,
  allowDuplicates = FALSE,
  verbose = FALSE,
  tandems = "known",
  reference = "all"
)

Arguments

...
character: One or more patterns used to identify files to read. For details: see the "REpath-patterns" section below.

contains
character. If !is.null(contains), the contains argument is is treated as regular expressions: only files which contain matches to all of these regular expressions are read. Thus, readHumdrum('.*krn$', contains = "EEE") will only read kern files which contain matches to "EE"---which is kern for the E two octaves below middle C (or lower).

recursive
logical: If TRUE, the final part of the search pattern (i.e., the file search) is searched for recursively through all sub directories.

allowDuplicates
logical of length one, indicating what should happen if multiple search patterns match the same files. If allowDuplicates = TRUE, any such files are read multiple times, grouped into their respective corpora by the Label field. If allowDuplicates = FALSE, any redundant files are only read into the corpus of the first pattern they match.

verbose
logical: If TRUE, the names of matching files are printed before parsing begins. This is very useful as a check to make sure you aren't reading the wrong files!

tandems
character. This argument controls which, if any, tandem interpretations are parsed into their own fields. The default value is "known".

reference
character. This argument controls which, if any, reference records are parsed into their own fields. The default value is "all".

Details

findHumdrum does the work of finding and reading the text files into R. readHumdrum utilizes findHumdrum to read files, then parses them to create a humTable and build a humdrumR data object around the table.

REpath-patterns

"REpath-patterns" are specified using ... arguments. In combination, all the ... arguments are used to search for file paths. Each part of the search path you specify ("dirpart/dirpart/filepart", etc) are matched as regular expressions against directories/files on your disc. Thus, we can say things like findHumdrum("../^A.*/.*krn$"), which would match any kern files in any directory beginning with a capital "A" in the directory above the current working directory. For conveniance, you can break the path across multiple arguments instead of using delimited strings: For example, the code findHumdrum("..", "^A.*", ".*krn$") will give an identical result as the previous example (findHumdrum("../^A.*/,*krn$")). This is useful when searching for more than one pattern (see next paragraph) in the same directory.

If you want to search for more than one pattern, you can input them as a character vector: For instance, readHumdrum(c("mozart", "beethoven")---this command will search for filenames containing "mozart" OR "beethoven." This works for directories too: readHumdrum(c("Mozart", "Beethoven"), ".*krn$") will look for any kern files in directories containing "Mozart" OR "Beethoven." If patterns are named, these names will show up as identifying patterns in the humdrumR object's Label field. Unnamed patterns are simply labeled with numbers.

Normal (system appropriate) conventions (i.e., directories separated by "/", '~' at beginning to indicate home, ".." to indicate directory above working directory, etc.) are followed. If a pattern contains a solo dot followed by a file sep---e.g., "./", "x/./y"---this is treated as the current directory, not a regular expression. If a pattern contains two dots---e.g., "../"---this is treated as the directory above, not a regular expression. If you want to create a regular expression to match any directory, use ".*/".

The regex pattern "" matches any file (it is changed to ".*"). If you don't specifiy any ... argument, findHumdrum (or readHumdrum) will default to ".*" as well. Thus, readHumdrum() will read any humdrum files in the working directory.

(If two or more files in different directories share the same name, a unique name is created for each file by appending the names of the directories they occupy, recursively until the names are unique.)

If a single humdrum file has multiple pieces in it---meaning that all spine paths close with *-, then open again with **---then they are parsed separetely. They are distinguished in the Piece field. If there are no multi-piece files, Piece and File will be identical.

Validity

findHumdrum and readHumdrum automatically ignore non-text files. Whatsmore, any files which contain humdrum syntax errors (checked by validateHumdrum) are automatically skipped. If you want to see specifically what errors occured, call validateHumdrum directly and its errorReport.path argument.

Tandem Interpretations

The tandems argument controls which tandem interpretations parsed into their own fields. This can be helpful to either save processing time and memory by not parsing interpretations you won't need, or to parse interpretations that humdrumR doesn't recognize. The "known" tandem interpretations that humdrumR recognizes are encoded in a build humdrumR table called knownInterpretations. Each interpretation has a humdrumR name ("Clef", "TimeSignature", etc.) as well as a regular expression associated with it. The default value for the tandems argument is "known". If the tandems argument contains "known" all tandem interpretations in the built-in knownInterpretations table are parsed. Users may specify different interpretations to parse in two ways:

character strings matching one of the name values from the Name column of knownInterpretations. For instance, if you specify tandems = c('Clef', 'TimeSignature'), only clef (e.g., "*clefG2"), and time signature (e.g., "*M3/4") intepretations will be parsed.

if the chracter string(s) in tandem do not exactly match one of the names in knownInterpretations$Name, they are treated as regular expressions and used to match tandem interpretations in the data. This allows users to parse non-standard tandem interpretations that humdrumR doesn't already know about.

If any values in tandems are named, these names will be used for resulting fields. If no matches to an given interpretation are found, no field is created for that interpretation. If tandems = NULL, then no tandem interpretations are parsed.

Reference Records

By default (reference = "all"), humdrumR reads all reference records in the data. The reference code for each record (e.g, the "OTL", in "!!!OTL: xxx") is used as the name of an associated field. (If a reference record has no reference code (i.e., it lacks a colon), the field is called "Unkeyed.") In large datasets with many reference records, the reference data can actually make up a large portion of the humdrum table, and eat up a lot of memory. In these cases, we might not want to read all (or any) reference records---we can instead read only the reference records that we are planning to use in our analyses (if any). If reference = NULL, no reference records are parsed. Otherwise, the character values of reference are treated as reference codes and only matching reference records are parsed. For instance, readHumdrum(_, reference = "OTL") will only parse OTL reference records. If the values of reference are named, these names are used to name associated fields. Thus, by specifing reference = c(Title = 'OTL'), you can use "OTL" reference records to populate a field called "Title".

If there are more than one reference records with the same reference code, either explicitely numbered (e.g., "!!!COM1:", "!!!COM2:") all are read and rather than making two or more fields, a single field is created ("COM" in this) with the multiple values separated by ";".

Result

findHumdrum returns a "fileFrame" (data.table), listing all file names, the patterns they match, the directories they were found in, and the raw text content of these files.

readHumdrum returns a fully parsed humdrumR object.

Examples