Reciprocal representation of duration

The standard approach to representing conventional note values in humdrum is the "reciprocal" **recip. Representation. The **recip rhythmic values are often used as a part of **kern representation, which also includes pitch information and notation details.

Usage

recip(
  x,
  ...,
  sep = "%",
  parseArgs = list(),
  scale = 1,
  unit = 1,
  inPlace = FALSE
)

Arguments

x

(atomic vector) The x argument can be any (atomic) vector, or a rational (rhythmInterval), or NULL.

...

These arguments are passed to the rhythm deparser. There are also two hidden (advanced) arguments you can specify: memoize and deparse (see the details below).

sep

(character, length == 1) A character string to use as the separator between denominator and numerator.

parseArgs

(list) parseArgs can be a list of arguments that are passed to the rhythm parser.

scale

A numeric or rational value which is used as the output unit of measurement: the default value for most functions is rational(1, 1), a whole-note or "duration."

inPlace

(logical, length == 1) This argument only has an effect if the input (the x argument) is character strings, and there is extra, non-duration information in the input strings "besides" the rhythm information. If so, and inPlace = TRUE, the output will be placed into an output string beside the original non-rhythm information. If inPlace = FALSE, only the rhythm output information will be returned (details below).

Details

**recip values are literally the reciprocal of a duration value. Since most note values in conventional music notation are simple fractions the reciprocal approach is highly concise and very similar to conventional western notation and terminology. A "quarter note" is represented as the reciprocal of 1/4: simply "4". Full reciprocal fractions can be specified: "2%3" to indicate 3/2. The % separator can be changed using the sep argument.

As in conventional note values, "dots" can be added after a value to increase the duration by the ratio of (2 - (2^{-n})), where n is the number of dots. (One dot is 3/2; two dots is 7/4; etc.).

See also