Representation of rhythmic information

This S4 class is the core rhythm representation in the humdrumR package. The object is used to represent rhythmic durations and metric positions. Each duration is represented in whole note units. Numerically, these whole note units are represented as a ratio between integers, held in the slots @Numerator and @Denominator. This allows use to represent any rational number with no loss of precision due to rounding errors and weak decimal expansions (like 0.333333). Rhythm intervals are similar to standard musical termoniology (i.e, ``three eighth-notes'' is the ratio \(\frac{3}{8}\)).

#' @name RhythmScaling #' @export augment.rhythmInterval <- function(rint, scalar) rint * scalar

rhythmInterval(x, ...)

rint(denominator, numerator = 1L)

is.rhythmInterval(x)

# S4 method for rhythmInterval
is.numeric(x)

# S4 method for rhythmInterval
as.character(x)

# S3 method for rhythmInterval
as.double(x)

order.rhythmInterval(
  x,
  ...,
  na.last = TRUE,
  decreasing = FALSE,
  method = c("auto", "shell", "radix")
)

# S4 method for rhythmInterval,rhythmInterval
Compare(e1, e2)

# S4 method for rhythmInterval
Summary(x)

rhythmInterval(x, ...)

Details

#' @name RhythmScaling #' @export augment.character <- regexDispatch('Recip' = recip.rhythmInterval %.% augment.rhythmInterval %.% read.recip2rhythmInterval, '0-9+%/+' = as.ratio.rhythmInterval %.% augment.rhythmInterval %.% read.fraction2rhythmInterval, 'Decimal' = as.decimal.rhythmInterval %.% augment.rhythmInterval %.% read.numeric2rhythmInterval)

#' @name RhythmScaling #' @export diminish <- function(x, scalar = 2, ...) UseMethod('diminish')

#' @name RhythmScaling #' @export diminish.rhythmInterval <- function(rint, scalar) rint / scalar

#' @name RhythmScaling #' @export diminish.character <- regexDispatch('Recip' = recip.rhythmInterval %.% diminish.rhythmInterval %.% read.recip2rhythmInterval, '0-9+%/+' = as.ratio.rhythmInterval %.% diminish.rhythmInterval %.% read.fraction2rhythmInterval, 'Decimal' = as.decimal.rhythmInterval %.% diminish.rhythmInterval %.% read.numeric2rhythmInterval)

Slots

Numerator

Integers

Octave

Integers

Vectorization

rhythmInterval inherits from the virtual class struct. This means you can apply normal vectorized commands to rhythmIntervals, and even put them in data.frames.

Arithmetic

rhythmInterval objects have arithmetic operations defined. Addition and subtraction are straightword and intuitive (i.e., \(\frac{1/8} + \frac{3/8} = \frac{1/2}\)).

Multiplication and division are slightly more complicated: scalar multiplication is defined for rational numbers: \(\frac{3}{8} * 2 = \frac{3}{4}\) (the result is always a new rhythmInterval). However, note that a rhythmInterval cannot be multiplied by another rhythmInterval---afterall, what would a quarter-note times a quarter-note be? A rhythmInterval can be divided by another rhythmInterval to produce a real number: \(\frac{1}{2} / \frac{1}{4} = 2 \). Like other rational values in R we can also do either ``true'' (rational) division (using the / operator) or Euclidean division (using the %/% operator). Rational division (/) of a rhythmInterval by another rhythmInterval results in a rational number. For instance, \(\frac{1}{2} / \frac{1}{4} = 2\). Rational division of a rhythmInterval by a rational number results in a new rhythmInterval: \(\frac{1}{2} / 2 = \frac{1}{4}\). Eucliean (a.k.a., integer) division can only be applied between rhythmIntervals resulting in an integer quotient---the remainder, which is a rhythmInterval, can be calculated with the %% operator. The remainder (a.k.a., modulo) operator (%%) is especially useful, for instance in calculating metric positions.

Relational Operators

rhythmIntervals can be compared using the standard relational operators---==, !=, >, >=, etc.