Roman numeral key areas

Roman numeral key areas

Usage

romanKey(x, ..., Key = NULL, parseArgs = list())

Arguments

x

Input data, interpreted as diatonic keys.

Must be an atomic vector.

Key

The key used by the parser, deparser, and transposer.

Defaults to NULL.

Must be a diatonicSet or something coercable to diatonicSet; must be either length 1 or length(x)

parseArgs

An optional list of arguments passed to the key parser.

Defaults to an empty list().

Must be a list of named arguments to the key parser.

Examples

romanKey(c('C:', 'd:dor', 'G:', 'g:'))
#> [1] "I"     "iidor" "V"     "v"