Representation of tonal pitch information

The tonalInterval is the core tonal pitch representation in humdrumR. A tonalInterval is an abstract representation of tonal pitch, which can be translated to/from all standard "concrete" pitch representations: solfege, scientific pitch, semitones, frequencies, scale degrees, intervals, etc. For the most part, users should not need to interact with tonalIntervals directly---rather, tonalIntervals work behind the scene in numerous humdrumR pitch functions. See the pitchRepresentations and tonalTransformations documentation for details of usage and functionality or the Tonality in humdrumR vignette for a detailed explanation of the theory and specifics of tonalIntervals.

tonalInterval(
  x,
  direction,
  delta,
  sigma,
  generic,
  simple,
  roundContour,
  enharmonic.part,
  Key,
  Exclusive,
  inPlace,
  ...
)

tint(
  octave,
  LO5th = 0L,
  cent = numeric(length(octave)),
  partition = FALSE,
  Key = NULL,
  roundContour = floor
)

is.tonalInterval(x)

Details

The tonalInterval is a S4 subclass of humdrumR's virtual class struct, from which it inherits a lot of useful "vector-like" behaviors/functionality.

The constructor function tint can be used to create tonalIntervals directly. The three arguments corespond to the three slots: octave, LO5th (Fifth), and cent. All inputs will be coerced to match in length. What's more, the octave argument can be left blank, in which case the appropriate octave will automatically be computed to place the interval in the octave above middle-C.

By default, the as.character method, and thus (via struct) the show method, for tonalIntervals call kern(). Thus, if you return a tonalInterval on the command line (or call print one one) you'll see the kern pitch representation printed.

Slots

Octave

integers representing the octave offset.

Fifth

integers representing the "line-of-fifths" value.

Cent

numeric values representing cents (1200th of an octave).
Arithmetic

Technically, tonalIntervals are examples of algebraic modules over integers. This means that certain arithmetic operations are defined for tonalIntervals and can be called using standard arithmetic operators (+, -, etc.):

Addition: tonalIntervals can be added together, acting exactly as you'd expect (i.e., \(M3 + m3 = P5\)).

Subtraction: tonalIntervals can be subtracted just as they are added. Also, they can be negated with a single - operator (like -M3).

Multiplication: tonalIntervals can not be multiplied together. However, scalar (integer) multiplication is defined: thus, tonalIntervals can be multiplied by integers to create new tonalIntervals: e.g., \(M2 * 3L = A4\).

Division: as the natural inverse of scale multiplication, Euclidean division is defined for tonalIntervals---i.e., division by/into whole (integer) pieces, often with leftover "remainders" (modulo). In R, Euclidean division is achieved with the %/% operator---not /---, with the associated %% used for the remainder/modulo. Two tonalIntervals can be divided to produced an integer; Conversely, a tonalInterval can be divided by an integer to produce a tonalInterval.

Take note that the way humdrumR defines Euclidean division is based in tonal space---i.e., the line-of-fifths---not frequency or atonal-semitone space. For example, an augmented-fourth divided by a major-second is 3L, but a diminished-fifth divided by a major-second is not 3L---d5 %/% M2 equals -3L with a remainder of P8 (plus an octave)! The division algorithm works by applying standard Euclidean division to the @Fifth slot (line-of-fifths tonal space), and shifting the @Octave value in the remainder to the match the appropriate octave. This definition has the useful properties that specificinterval %% A1 = genericinterval and interval %% d2 = enharmonicinterval.

Relational Operators

tonalIntervals can be compared using the standard relational operations---==, !=, >, >=, etc. Two tonalIntervals are equal (according to ==) only if all their slots (Octave, Fifth, and Cent) are exactly identical. Thus, enharmonic notes (like C# and Db) are not equal. In contrast, ordinal comparisons (e.g., >, <=) between tonalIntervals are based on their semitone (equal temperament) size, so enharmonicity is irrelevant. Thus, m3 >= A2 and A2 >= m3 are both TRUE, even though m3 == A2 is not.

Coercion

humdrumR knows how to coerce several base-R atomic types into tonalIntervals. This can be done using the as function---e.g., as(3, "tonalInterval")---or more intuitively using the function tonalInterval(). Coercision methods are defined for

integer: interpreted as semitones

numeric: interpreted as frequency ratios, assuming a Pythagorean tuning.

character: interpreted using humdrumRs regular expression dispatch system, as explained fully here.

Since, coersion is defined, tonalInterval arithmatic can be applied when one of the two arguments is a tonalInterval and the other is a coercable atomic. For instance, M3 + 2L will interpret 2L as two semitones and add a major-second to the major-third! The clever dispatch system will even ignore character strings that are not recognized (see examples)!

Predefined Intervals:

humdrumR automatically exports a bunch of tonalIntervals, named by their musical interval representation. Every generic interval from 1 to 15 is combined with every interval quality dd (doubly diminished), d (diminished), m (minor), M (major), A (augumented) AA (doubly augmented). Thus, after loading humdrumR, you can type things like M3 + M3 and get A5. In addition, the variables unison (= P1 = tint(0, 0)), pythagorean.comma (= d2 = tint(-19,12)), and octave (tint(1, 0)) are exported as well.

Examples

M3 <- tint(   , 4L)

M2 <- tint(   , 2L)
M9 <- tint(-1L, 2L)

M9 - M2 

#> tonalInterval[ , ]
#> [1] ccc

# = octave
M9 - 2L

#> tonalInterval[ , ]
#> [1] ccc

# = octave

M3 %/% M2 

#> [1] 2

# = 2

###

cMajor <- sort(tint( , -1:5))
eMajor <- cMajor + M3
eMajor + 2L 

#> tonalInterval[ , ]
#> [1] f#  g#  a#  b   cc# dd# ee#

# f# g# a# b cc# dd# ee#

eMajor[4:5] - octave 

#> tonalInterval[ , ]
#> [1] A B

# = A B

"4.ee" + P5 

#> [1] "4.bb"

# = "4.bb"