humApply is just a wrapper for with(in)Humdrum, included to parallel the R family of _apply functions. humApply uses non-standard evaluation to capture arguments fed to it without the user needing to make explicit formula using ~. This is only guaranteed to work in the global environment, so be careful. If you run into problems, switch over to with(in)Humdrum and use explicit X~formulas.

humApply is just a wrapper for with(in)Humdrum, included to parallel the R family of _apply functions. humApply uses non-standard evaluation to capture arguments fed to it without the user needing to make explicit formula using ~. This is only guaranteed to work in the global environment, so be careful. If you run into problems, switch over to with(in)Humdrum and use explicit X~formulas.

humApply(humdrumR, FUN, ..., within = TRUE, doplot = FALSE)

Arguments

humdrumR

A humdrumR data object.

FUN

A function to apply to the Active field(s) in the humdrumR object.

...

Any arguments which can be fed to with(in)Humdrum as formulae (except for do expressions, which are replaced by the FUN argument!). However, rather that writinging formula in the format Keyword ~ Expression, humApply arguments should be written as normal R arguments: Keyword = Expression. Unnamed arguments are ignored. #' @param within A logical. If TRUE (the default), withinHumdrum is used to apply the function---meaning that the output is reconstituted into a new field in the humdrumR object. If within == FALSE, withHumdrum is used instead, which results in the function's output being returned inprocessed.

doplot

Boolean. If TRUE the FUN argument is treated as a doplot expression by with(in)Humdrum, so the result is ignored (for plotting or side-effects purposes).