with(in)Humdrum

Apply arbitrary expressions to fields within [S4class][humdrumRclass] data.

Usage

withinHumdrum(humdrumR, ...)

withHumdrum(humdrumR, ..., drop = TRUE)

inHumdrum(humdrumR, ...)

humApply(humdrumR, FUN, ..., within = TRUE, doplot = FALSE)

Arguments

humdrumR

A humdrumRclass data object.

...

Any arguments which can be fed to [humdrumR:with-in-Humdrum][with(in)Humdrum] as formulae (except for do expressions, which are replaced by the FUN argument!). However, rather that writinging formula in the format Keyword ~ Expression, humApply arguments should be written as normal R arguments: Keyword = Expression. Unnamed arguments are ignored. #' @param within A logical. If TRUE (the default), [humdrumR:with-in-Humdrum][withinHumdrum] is used to apply the function---meaning that the output is reconstituted into a new field in the humdrumR object. If within == FALSE, [humdrumR:with-in-Humdrum][withHumdrum] is used instead, which results in the function's output being returned inprocessed.

drop

This argument is concetually similar to the drop argument in R matrices and data.frames. If drop = TRUE, the output of withHumdrum is simplified as much as possible (trying to return the "raw" vector, list, table, etc. within it). If drop = FALSE, the result is always a data.table. The default value (drop = TRUE) is usually what we want because it is more intuitive, but in more complex code, it can be helpful to set drop = FALSE so that the output is consistent.

FUN

A function to apply to the humdrumR:humdrumRfield(s) in thehumdrumR` object.

doplot

Boolean. If TRUE the FUN argument is treated as a doplot expression by [humdrumR:with-in-Humdrum][with(in)Humdrum], so the result is ignored (for plotting or side-effects purposes).

Value

From withinHumdrum and inHumdrum, a new humdrumR data object. From withHumdrum, whatever value is returned by expression.

Overview

These functions are the primary means of working with humdrumR data. (They are analogous to the base functions [base][with] and [base][within] as applied to [base:data.frame][data.frames].) Specifically they allow you to evaluate arbitrary expressions involving fields in a humdrumR data object. They also includes a number of special evaluation options:

Evaluate an expression in only matching parts of the data.

Evaluate an expression separately in subsets of the data.

Evaluate an expression across windows in the data (e.g., ngrams, rolling windows).

Evaluate an expression which produces a plot, with particular plotting parameters set using [graphics][par].

The difference between withHumdrum and withinHumdrum is analogous to the difference between [base][with] and [base][within]. withHumdrum evaluates your expression(s) and then simply returns the result of the evaluation. withinHumdrum evaluates your expression(s) and then (attempts) to insert the results back into the humdrumR object, generating new fields called PipeX (see details).

inHumdrum is simply a short hand for withinHumdrum.

Formulae

Every formula in the formulae argument is treated as a Keyword ~ Expression(s) pairing. Multiple expressions can be input using multiple ~ operators: Keyword ~ Expression1 [~ Expression2 ~ ... ~ ExpressionN] (the leftmost expression is treated as the keyword.) If there is no leftmost expression (i.e., ~ Expression), the Keyword defaults to "do." The keyword expression must be a single, simple name/symbol, following standard R rules (i.e., ".foobar" is acceptable but "3 + foobar" is not).

Legal keywords, and their meanings are:

do An expression to be evaluated within the humdrumR data object (see "Expression evaluation").

doplot An expression to be evaluated within the humdrumR data object while ignoring the result of the expression (see "Expression evaluation" and "Plotting".

by An expression used to break the data into groups, with the do expression(s) evaluated separately in each group (see "Partitioning").

where An expression indicating a subset of the data in which to evaluate the do expression (see "Partitioning").

ngrams A positive number n. The expression is evaluated across overlapping length-n windows.

recordtypes A string or vector of characters drawn from c("D", "d", "I", "L", "M","G"). These characters correspond to types of humdrum records: Data, null data, Interpretations, Measures, Local comments, and Global comments respectively. The expression is only evaluated on data drawn from the specified record types (defaults to "D").

pre An expression to evaluate once before evaluating the do expression(s). Useful, for instance, for taking logs or opening a graphing window. The pre expression is evaluated in the global environment.

post An expression evaluate once after evaluating the do expression(s). Always evaluated in the global environment.

Expression evaluation

The right-hand side of any formula in the formulae argument with the keyword do or doplot (or with no keyword specified) is evaluated within the humdrumR data object. The expression can, thus, refer to any field in the humdrumR object (Record, Token, File, etc.). You can also include a . anywhere in the expression, which will be interpreted as the humdrumR object's current [dest=humdrumR][Active] expression.

humdata <- readHumdrum('directorywithdata/*.krn') # read some data

withinHumdrum(humdata, ~getPitch(Token)) # The most basic pattern
withinHumdrum(humdata, ~getPitch(.)) # Same as previous (unless `Active` field has been changed))

withinHumdrum(humdata, ~solfa(getPitch(Token), key = Key)) 
# Assumes that the Key field was parsed during the call to `[readHumdrum][readHumdrum]`

withinHumdrum(humdata, ~getSemits(Token) - mean(getSemits(Token)))

If multiple do expressions are provided, they are each evaluated one at a time, with the result of each piped into the next. Other, non-do, formulae (like by~ or ngrams~) are reused for each expression evaluated.

Partitioning

A by expression is used to break the data into subsets, with the do expression(s) evaluated separately within each subset. This works the similarly to the by argument in [data.table][data.table]s, the INDEX argument of [base][tapply], or the INDICES argument of [base][by]. Each by expression must evaluate, within the humdrumR data object, to a vector (or a list of vectors of equal length) of categories to group the data by. Most commonly, the by expression(s) are simply field(s) in the data: for instance,

withinHumdrum(humdata,
         do ~ table(Token),
         by ~ File)

will apply the function [base][table] to the Token field separately for each file in the humdrumR data. However, we can also use more complex expressions like

withinHumdrum(humdata,
         do ~ table(Token), 
         by ~ Spine > 3 | Record \%\% 2 == 0)

which will evaluate the do expression in two groups, one where either the spine number is three or less or the record number is even, and another group where the opposite is true.

If the by expression evaluates to a list of grouping vectors, the do expressions are evaulated across every combination of categories in all the vectors. Thus, withinHumdrum(humdata, do ~ table(Token), by ~ list(File, Spine)) will apply table to Token across each spine in each file. As some syntactic sugar, if the by expression has more than two parts, all parts except the (leftmost) keyword part are combined in a list (i.e., by ~ File ~ Spine becomes by ~ list(File, Spine)}. Thus the previous example can also be written: withinHumdrum(humdata, do ~ table(Token), by ~ File ~ Spine)

A where expression is used to identify a subset of the data and evaluate the do expression(s) only in that subset. where expressions must evaluated, within the humdrumR data object, to a single logical vector. The do expression(s) are only evaluated where this logical vector is TRUE. Wherever the where expression evaluates to FALSE, the original humdrumR data is kept unchanged.

If multiple partitioning formulae (i.e, by and where) expressions are evaluated recursively, in order from left to right. Thus if you specify withinHumdrum(humdata, do ~ sd(semits), by ~ File, where ~ semits > mean(semits)) a the standard deviation of the semits field will be calculated only in each file, but only where the semits field is greater than the mean semits value within that file. Contrast this with this call: withinHumdrum(humdata, do ~ sd(semits) where ~ semits > mean(semits), by ~ File) wherein the standard deviation of semits is, again, calculated for each file, but this time wherever the semits field is greater than the mean value across all the data.

Plotting

The doplot keyword behaves exactly like the do keyword, except that the result of the evaluation is ignored. This is useful for plotting as well as other side-effects (like writing to a file). If doplot is used with withHumdrum, the function simply returns NULL (after executing the doplot expression If doplot is used with withinHumdrum (or inHumdrum), the function simply returns the unaltered humdrumR argument.

withinHumdrum also allows you to specify plotting options in line, without having to make a separate call to [graphics][par]. Any [graphics][par] argument can be specified as a Keyword ~ Expression pair in the formulae argument. For instance, if you call a doplot expression with a by expression that creates four groups, R will create four plots---but you will only see the last one! Normally, you would need to call par(mfcol = c(2,2)) before calling your plotting function. However, with withinHumdrum you can soecific mfcol = c(2,2) right in a formulae formula: withinHumdrum(humdata, doplot ~ fooplot(.), by ~ list(Two, byTwo), mfcol ~ c(2, 2)) The best part is withinHumdrum will reset par to it's previous state after withinHumdrum is done.

Tandem interpretations

The function [readHumdrum][readHumdrum] automatically parses tandem interpretations (that it recognizes) into their own fields in the resulting [humdrumR][humdrumRclass] data. For instance, data with a '*clefF4' will show up as a Clef field. However, users might read humdrum data with their own custom tandem interpretations that are not built into humdrumR. humdrumR includes the function [getTandem][getTandem] to help us extract arbitrary tandem intrpretation data. Luckily, withinHumdrum knows some syntactic sugar which makes it easy to do this anywhere in our expressions, simply by putting a named object beginning with the symbol *. Of course, R doesn't normally allow names to begin with symbols like *, but you can force it by placing grave symbols around the name *name. If you do this in a withinHumdrum expression, withinHumdrum will treat this name as a regular expression and substitute a call getTandem(Tandem, 'regular expression') in the expression. This means you can could do something like withinHumdrum(humdata, do ~ myFunction(Token, *mytandempattern)) and myFunction will be called with the first argument being the Token field, and the second argument being tandem interpretations which match 'mytandempattern' (extracted from the Tandem field).

Splatting

("Splatting" refers to feeding a function a list/vector of arguments.) Sometimes we want to divide our data into pieces (a l\'a partition option), but rather than applying the same expression to each piece, we want to feed the separate pieces as separate arguments to the same function. In withinHumdrum you can use some syntactic sugar to do just this, using the @ symbol in the format myFunction(TargetExpr@GroupingExpr). If we make this call

withinHumdrum(humdata, do ~ myFunction(Token@Spine))

and there are four spines this is how withinHumdrum will intepret the expression:

withinHumdrum(humData, do ~ myFunction(TokenSpine == 1, # first argument when Spine == 1 TokenSpine == 2, # second argument when Spine == 2 TokenSpine == 3, # etc. TokenSpine == 4))

Argument interpolation

Any named arguments to withinHumdrum are [humdrumR:interpolateArguments][interpolated] into the do expressions. This is useful if you've already created a list of formulas that you like, but would like to make small changes to a function call within the do expressions, without starting from scratch. Examples:

mycommand <- c(do ~ mean(., na.rm = TRUE), by ~ Spine ~ File)
withinHumdrum(humdata,
              mycommand,
              na.rm = FALSE)
# mycommand is executed with na.rm changed to FALSE              

Piping

For calls to withinHumdrum, the result of each do expression is insterted back into the [humtable][humdrum table]. The results are put into new field(s) labeled Pipe1, PipeX, ..., PipeN. If the results of the expression are shorter than the rows in the humtable, or an object, the humdrum table is shrunk to fit them.

Examples

humdata <- readHumdrum('directorywithdata/*.krn')
#> Finding and reading files...
#> 	REpath-pattern 'directorywithdata/*.krn' matches 0 text files in 0 directories.
#> Zero files read from disk. 
#> No files to validate.

withinHumdrum(humdata, ~nchar(.)) # counts characters in each data token.
#> humdrumR error in:
#> 	tryCatch(withCallingHandlers({
#>   NULL
#>   saveRDS(do.call(do.call, c(readRDS(
#>     "/home/nat/.tmp/RtmpqwGfV0/callr-fun-15ea7056d2377"), list(envir = .GlobalEnv,
#>     quote = TRUE)), envir = .GlobalEnv, quote = TRUE), file = "/home/nat/.tmp/RtmpqwGfV0/callr-res-15ea701f9a8237")
#>   flush(stdout())
#>   flush(stderr())
#>   NULL
#>   invisible()
#> }, error = function(e) {
#>   {
#>     callr_data <- as.environment("tools:callr")$`__callr_data__`
#>     err <- callr_data$err
#>     assign(".Traceback", .traceback(4), envir = callr_data)
#>     dump.frames("__callr_dump__")
#>     assign(".Last.dump", .GlobalEnv$`__callr_dump__`, envir = callr_data)
#>     rm("__callr_dump__", envir = .GlobalEnv)
#>     e$call <- deparse(conditionCall(e), nlines = 6)
#>     e2 <- err$new_error(conditionMessage(e), call. = conditionCall(e))
#>     class(e2) <- c("callr_remote_error", class(e2))
#>     e2$error <- e
#>     calls <- sys.calls()
#>     dcframe <- which(vapply(calls, function(x) length(x) >= 1 && identical(x[[
#>       1]], quote(do.call)), logical(1)))[1]
#>     if (!is.na(dcframe)) e2$`_ignore` <- list(c(1, dcframe + 1L))
#>     e2$`_pid` <- Sys.getpid()
#>     e2$`_timestamp` <- Sys.time()
#>     if (inherits(e, "rlib_error_2_0")) e2$parent <- e$parent
#>     e2 <- err$add_trace_back(e2, embed = FALSE)
#>     saveRDS(list("error", e2), file = paste0(
#>       "/home/nat/.tmp/RtmpqwGfV0/callr-res-15ea701f9a8237", ".error"))
#>   }
#> }, interrupt = function(e) {
#>   {
#>     callr_data <- as.environment("tools:callr")$`__callr_data__`
#>     err <- callr_data$err
#>     assign(".Traceback", .traceback(4), envir = callr_data)
#>     dump.frames("__callr_dump__")
#>     assign(".Last.dump", .GlobalEnv$`__callr_dump__`, envir = callr_data)
#>     rm("__callr_dump__", envir = .GlobalEnv)
#>     e$call <- deparse(conditionCall(e), nlines = 6)
#>     e2 <- err$new_error(conditionMessage(e), call. = conditionCall(e))
#>     class(e2) <- c("callr_remote_error", class(e2))
#>     e2$error <- e
#>     calls <- sys.calls()
#>     dcframe <- which(vapply(calls, function(x) length(x) >= 1 && identical(x[[
#>       1]], quote(do.call)), logical(1)))[1]
#>     if (!is.na(dcframe)) e2$`_ignore` <- list(c(1, dcframe + 1L))
#>     e2$`_pid` <- Sys.getpid()
#>     e2$`_timestamp` <- Sys.time()
#>     if (inherits(e, "rlib_error_2_0")) e2$parent <- e$parent
#>     e2 <- err$add_trace_back(e2, embed = FALSE)
#>     saveRDS(list("error", e2), file = paste0(
#>       "/home/nat/.tmp/RtmpqwGfV0/callr-res-15ea701f9a8237", ".error"))
#>   }
#> }, callr_message = function(e) {
#>   try(signalCondition(e))
#> }), error = function(e) {
#>   NULL
#>   try(stop(e))
#> }, interrupt = function(e) {
#>   NULL
#>   e
#> })
#> 	 -> tryCatchList(expr, classes, parentenv, handlers)
#> 	 -> tryCatchOne(tryCatchList(expr, names[-nh], parentenv, handlers[-nh]), names[nh],
#> parentenv, handlers[[nh]])
#> 	 -> doTryCatch(return(expr), name, parentenv, handler)
#> 	 -> tryCatchList(expr, names[-nh], parentenv, handlers[-nh])
#> 	 -> tryCatchOne(expr, names, parentenv, handlers[[1L]])
#> 	 -> doTryCatch(return(expr), name, parentenv, handler)
#> 	 -> withCallingHandlers({
#>   NULL
#>   saveRDS(do.call(do.call, c(readRDS(
#>     "/home/nat/.tmp/RtmpqwGfV0/callr-fun-15ea7056d2377"), list(envir = .GlobalEnv,
#>     quote = TRUE)), envir = .GlobalEnv, quote = TRUE), file = "/home/nat/.tmp/RtmpqwGfV0/callr-res-15ea701f9a8237")
#>   flush(stdout())
#>   flush(stderr())
#>   NULL
#>   invisible()
#> }, error = function(e) {
#>   {
#>     callr_data <- as.environment("tools:callr")$`__callr_data__`
#>     err <- callr_data$err
#>     assign(".Traceback", .traceback(4), envir = callr_data)
#>     dump.frames("__callr_dump__")
#>     assign(".Last.dump", .GlobalEnv$`__callr_dump__`, envir = callr_data)
#>     rm("__callr_dump__", envir = .GlobalEnv)
#>     e$call <- deparse(conditionCall(e), nlines = 6)
#>     e2 <- err$new_error(conditionMessage(e), call. = conditionCall(e))
#>     class(e2) <- c("callr_remote_error", class(e2))
#>     e2$error <- e
#>     calls <- sys.calls()
#>     dcframe <- which(vapply(calls, function(x) length(x) >= 1 && identical(x[[
#>       1]], quote(do.call)), logical(1)))[1]
#>     if (!is.na(dcframe)) e2$`_ignore` <- list(c(1, dcframe + 1L))
#>     e2$`_pid` <- Sys.getpid()
#>     e2$`_timestamp` <- Sys.time()
#>     if (inherits(e, "rlib_error_2_0")) e2$parent <- e$parent
#>     e2 <- err$add_trace_back(e2, embed = FALSE)
#>     saveRDS(list("error", e2), file = paste0(
#>       "/home/nat/.tmp/RtmpqwGfV0/callr-res-15ea701f9a8237", ".error"))
#>   }
#> }, interrupt = function(e) {
#>   {
#>     callr_data <- as.environment("tools:callr")$`__callr_data__`
#>     err <- callr_data$err
#>     assign(".Traceback", .traceback(4), envir = callr_data)
#>     dump.frames("__callr_dump__")
#>     assign(".Last.dump", .GlobalEnv$`__callr_dump__`, envir = callr_data)
#>     rm("__callr_dump__", envir = .GlobalEnv)
#>     e$call <- deparse(conditionCall(e), nlines = 6)
#>     e2 <- err$new_error(conditionMessage(e), call. = conditionCall(e))
#>     class(e2) <- c("callr_remote_error", class(e2))
#>     e2$error <- e
#>     calls <- sys.calls()
#>     dcframe <- which(vapply(calls, function(x) length(x) >= 1 && identical(x[[
#>       1]], quote(do.call)), logical(1)))[1]
#>     if (!is.na(dcframe)) e2$`_ignore` <- list(c(1, dcframe + 1L))
#>     e2$`_pid` <- Sys.getpid()
#>     e2$`_timestamp` <- Sys.time()
#>     if (inherits(e, "rlib_error_2_0")) e2$parent <- e$parent
#>     e2 <- err$add_trace_back(e2, embed = FALSE)
#>     saveRDS(list("error", e2), file = paste0(
#>       "/home/nat/.tmp/RtmpqwGfV0/callr-res-15ea701f9a8237", ".error"))
#>   }
#> }, callr_message = function(e) {
#>   try(signalCondition(e))
#> })
#> 	 -> saveRDS(do.call(do.call, c(readRDS(
#>   "/home/nat/.tmp/RtmpqwGfV0/callr-fun-15ea7056d2377"), list(envir = .GlobalEnv,
#>   quote = TRUE)), envir = .GlobalEnv, quote = TRUE), file = "/home/nat/.tmp/RtmpqwGfV0/callr-res-15ea701f9a8237")
#> 	 -> do.call(do.call, c(readRDS("/home/nat/.tmp/RtmpqwGfV0/callr-fun-15ea7056d2377"),
#> list(envir = .GlobalEnv, quote = TRUE)), envir = .GlobalEnv, quote = TRUE)
#> 	 -> <function(what, args, quote = FALSE, envir = parent.frame()) {
#>   if (!is.list(args)) stop("second argument must be a list")
#>   if (quote) args <- lapply(args, enquote)
#>   .Internal(do.call(what, args, envir))
#> }>(base::quote(<function(..., crayon_enabled, crayon_colors, pkgdown_internet) {
#>   options(crayon.enabled = crayon_enabled, crayon.colors = crayon_colors,
#>     pkgdown.internet = pkgdown_internet)
#>   pkgdown::build_site(...)
#> }>), base::quote(<list: pkg = <pkgdown>, examples = TRUE, run_dont_run = FALSE,
#>   seed = 1014, lazy = FALSE, ...>), envir = base::quote(<environment>), quote = base::quote(
#>   TRUE))
#> 	 -> <function(..., crayon_enabled, crayon_colors, pkgdown_internet) {
#>   options(crayon.enabled = crayon_enabled, crayon.colors = crayon_colors,
#>     pkgdown.internet = pkgdown_internet)
#>   pkgdown::build_site(...)
#> }>(pkg = base::quote(<pkgdown>), examples = base::quote(TRUE), run_dont_run = base::quote(
#>   FALSE), seed = base::quote(1014), lazy = base::quote(FALSE), override = base::quote(
#>   <list: >), install = base::quote(FALSE), preview = base::quote(FALSE),
#> new_process = base::quote(FALSE), devel = base::quote(FALSE), crayon_enabled = base::quote(
#>   TRUE), crayon_colors = base::quote(256L), pkgdown_internet = base::quote(TRUE))
#> 	 -> pkgdown::build_site(...)
#> 	 -> build_site_local(pkg = pkg, examples = examples, run_dont_run = run_dont_run,
#>   seed = seed, lazy = lazy, override = override, preview = preview, devel = devel)
#> 	 -> build_reference(pkg, lazy = lazy, examples = examples, run_dont_run = run_dont_run,
#>   seed = seed, override = override, preview = FALSE, devel = devel)
#> 	 -> purrr::map(topics, build_reference_topic, pkg = pkg, lazy = lazy, examples_env = examples_env,
#>   run_dont_run = run_dont_run)
#> 	 -> .f(.x[[i]], ...)
#> 	 -> withCallingHandlers(data_reference_topic(topic, pkg, examples_env = examples_env,
#>   run_dont_run = run_dont_run), error = function(err) {
#>   msg <- c(paste0("Failed to parse Rd in ", topic$file_in), i = err$message)
#>   abort(msg, parent = err)
#> })
#> 	 -> data_reference_topic(topic, pkg, examples_env = examples_env, run_dont_run = run_dont_run)
#> 	 -> run_examples(tags$tag_examples[[1]], env = if (is.null(examples_env)
#> ) NULL else new.env(parent = examples_env), topic = tools::file_path_sans_ext(
#>   topic$file_in), run_dont_run = run_dont_run)
#> 	 -> highlight_examples(code, topic, env = env)
#> 	 -> downlit::evaluate_and_highlight(code, fig_save = fig_save_topic, env = child_env(
#>   env), output_handler = evaluate::new_output_handler(value = pkgdown_print))
#> 	 -> evaluate::evaluate(code, child_env(env), new_device = TRUE, output_handler = output_handler)
#> 	 -> evaluate_call(expr, parsed$src[[i]], envir = envir, enclos = enclos, debug = debug,
#> last = i == length(out), use_try = stop_on_error != 2L, keep_warning = keep_warning,
#> keep_message = keep_message, output_handler = output_handler, include_timing = include_timing)
#> 	 -> timing_fn(handle(ev <- withCallingHandlers(withVisible(eval(expr, envir, enclos)),
#> warning = wHandler, error = eHandler, message = mHandler)))
#> 	 -> handle(ev <- withCallingHandlers(withVisible(eval(expr, envir, enclos)),
#> warning = wHandler, error = eHandler, message = mHandler))
#> 	 -> try(f, silent = TRUE)
#> 	 -> tryCatch(expr, error = function(e) {
#>   call <- conditionCall(e)
#>   if (!is.null(call)) {
#>     if (identical(call[[1L]], quote(doTryCatch))) call <- sys.call(-4L)
#>     dcall <- deparse(call, nlines = 1L)
#>     prefix <- paste("Error in", dcall, ": ")
#>     LONG <- 75L
#>     sm <- strsplit(conditionMessage(e), "\n")[[1L]]
#>     w <- 14L + nchar(dcall, type = "w") + nchar(sm[1L], type = "w")
#>     if (is.na(w)) w <- 14L + nchar(dcall, type = "b") + nchar(sm[1L], type = "b")
#>     if (w > LONG) prefix <- paste0(prefix, "\n  ")
#>   } else prefix <- "Error : "
#>   msg <- paste0(prefix, conditionMessage(e), "\n")
#>   .Internal(seterrmessage(msg[1L]))
#>   if (!silent && isTRUE(getOption("show.error.messages"))) {
#>     cat(msg, file = outFile)
#>     .Internal(printDeferredWarnings())
#>   }
#>   invisible(structure(msg, class = "try-error", condition = e))
#> })
#> 	 -> tryCatchList(expr, classes, parentenv, handlers)
#> 	 -> tryCatchOne(expr, names, parentenv, handlers[[1L]])
#> 	 -> doTryCatch(return(expr), name, parentenv, handler)
#> 	 -> withCallingHandlers(withVisible(eval(expr, envir, enclos)), warning = wHandler,
#> error = eHandler, message = mHandler)
#> 	 -> withVisible(eval(expr, envir, enclos))
#> 	 -> eval(expr, envir, enclos)
#> 	 -> eval(expr, envir, enclos)
#> 	 -> withinHumdrum(humdata, ~ nchar(.))
#> 	
#> Error: In the call withinHumdrum(humdrumR = _), the argument humdrumR must be a humdrumR object.
withinHumdrum(humdata, ~table(.), by ~ Spine) # Tabulates data tokens in each Spine.
#> humdrumR error in:
#> 	tryCatch(withCallingHandlers({
#>   NULL
#>   saveRDS(do.call(do.call, c(readRDS(
#>     "/home/nat/.tmp/RtmpqwGfV0/callr-fun-15ea7056d2377"), list(envir = .GlobalEnv,
#>     quote = TRUE)), envir = .GlobalEnv, quote = TRUE), file = "/home/nat/.tmp/RtmpqwGfV0/callr-res-15ea701f9a8237")
#>   flush(stdout())
#>   flush(stderr())
#>   NULL
#>   invisible()
#> }, error = function(e) {
#>   {
#>     callr_data <- as.environment("tools:callr")$`__callr_data__`
#>     err <- callr_data$err
#>     assign(".Traceback", .traceback(4), envir = callr_data)
#>     dump.frames("__callr_dump__")
#>     assign(".Last.dump", .GlobalEnv$`__callr_dump__`, envir = callr_data)
#>     rm("__callr_dump__", envir = .GlobalEnv)
#>     e$call <- deparse(conditionCall(e), nlines = 6)
#>     e2 <- err$new_error(conditionMessage(e), call. = conditionCall(e))
#>     class(e2) <- c("callr_remote_error", class(e2))
#>     e2$error <- e
#>     calls <- sys.calls()
#>     dcframe <- which(vapply(calls, function(x) length(x) >= 1 && identical(x[[
#>       1]], quote(do.call)), logical(1)))[1]
#>     if (!is.na(dcframe)) e2$`_ignore` <- list(c(1, dcframe + 1L))
#>     e2$`_pid` <- Sys.getpid()
#>     e2$`_timestamp` <- Sys.time()
#>     if (inherits(e, "rlib_error_2_0")) e2$parent <- e$parent
#>     e2 <- err$add_trace_back(e2, embed = FALSE)
#>     saveRDS(list("error", e2), file = paste0(
#>       "/home/nat/.tmp/RtmpqwGfV0/callr-res-15ea701f9a8237", ".error"))
#>   }
#> }, interrupt = function(e) {
#>   {
#>     callr_data <- as.environment("tools:callr")$`__callr_data__`
#>     err <- callr_data$err
#>     assign(".Traceback", .traceback(4), envir = callr_data)
#>     dump.frames("__callr_dump__")
#>     assign(".Last.dump", .GlobalEnv$`__callr_dump__`, envir = callr_data)
#>     rm("__callr_dump__", envir = .GlobalEnv)
#>     e$call <- deparse(conditionCall(e), nlines = 6)
#>     e2 <- err$new_error(conditionMessage(e), call. = conditionCall(e))
#>     class(e2) <- c("callr_remote_error", class(e2))
#>     e2$error <- e
#>     calls <- sys.calls()
#>     dcframe <- which(vapply(calls, function(x) length(x) >= 1 && identical(x[[
#>       1]], quote(do.call)), logical(1)))[1]
#>     if (!is.na(dcframe)) e2$`_ignore` <- list(c(1, dcframe + 1L))
#>     e2$`_pid` <- Sys.getpid()
#>     e2$`_timestamp` <- Sys.time()
#>     if (inherits(e, "rlib_error_2_0")) e2$parent <- e$parent
#>     e2 <- err$add_trace_back(e2, embed = FALSE)
#>     saveRDS(list("error", e2), file = paste0(
#>       "/home/nat/.tmp/RtmpqwGfV0/callr-res-15ea701f9a8237", ".error"))
#>   }
#> }, callr_message = function(e) {
#>   try(signalCondition(e))
#> }), error = function(e) {
#>   NULL
#>   try(stop(e))
#> }, interrupt = function(e) {
#>   NULL
#>   e
#> })
#> 	 -> tryCatchList(expr, classes, parentenv, handlers)
#> 	 -> tryCatchOne(tryCatchList(expr, names[-nh], parentenv, handlers[-nh]), names[nh],
#> parentenv, handlers[[nh]])
#> 	 -> doTryCatch(return(expr), name, parentenv, handler)
#> 	 -> tryCatchList(expr, names[-nh], parentenv, handlers[-nh])
#> 	 -> tryCatchOne(expr, names, parentenv, handlers[[1L]])
#> 	 -> doTryCatch(return(expr), name, parentenv, handler)
#> 	 -> withCallingHandlers({
#>   NULL
#>   saveRDS(do.call(do.call, c(readRDS(
#>     "/home/nat/.tmp/RtmpqwGfV0/callr-fun-15ea7056d2377"), list(envir = .GlobalEnv,
#>     quote = TRUE)), envir = .GlobalEnv, quote = TRUE), file = "/home/nat/.tmp/RtmpqwGfV0/callr-res-15ea701f9a8237")
#>   flush(stdout())
#>   flush(stderr())
#>   NULL
#>   invisible()
#> }, error = function(e) {
#>   {
#>     callr_data <- as.environment("tools:callr")$`__callr_data__`
#>     err <- callr_data$err
#>     assign(".Traceback", .traceback(4), envir = callr_data)
#>     dump.frames("__callr_dump__")
#>     assign(".Last.dump", .GlobalEnv$`__callr_dump__`, envir = callr_data)
#>     rm("__callr_dump__", envir = .GlobalEnv)
#>     e$call <- deparse(conditionCall(e), nlines = 6)
#>     e2 <- err$new_error(conditionMessage(e), call. = conditionCall(e))
#>     class(e2) <- c("callr_remote_error", class(e2))
#>     e2$error <- e
#>     calls <- sys.calls()
#>     dcframe <- which(vapply(calls, function(x) length(x) >= 1 && identical(x[[
#>       1]], quote(do.call)), logical(1)))[1]
#>     if (!is.na(dcframe)) e2$`_ignore` <- list(c(1, dcframe + 1L))
#>     e2$`_pid` <- Sys.getpid()
#>     e2$`_timestamp` <- Sys.time()
#>     if (inherits(e, "rlib_error_2_0")) e2$parent <- e$parent
#>     e2 <- err$add_trace_back(e2, embed = FALSE)
#>     saveRDS(list("error", e2), file = paste0(
#>       "/home/nat/.tmp/RtmpqwGfV0/callr-res-15ea701f9a8237", ".error"))
#>   }
#> }, interrupt = function(e) {
#>   {
#>     callr_data <- as.environment("tools:callr")$`__callr_data__`
#>     err <- callr_data$err
#>     assign(".Traceback", .traceback(4), envir = callr_data)
#>     dump.frames("__callr_dump__")
#>     assign(".Last.dump", .GlobalEnv$`__callr_dump__`, envir = callr_data)
#>     rm("__callr_dump__", envir = .GlobalEnv)
#>     e$call <- deparse(conditionCall(e), nlines = 6)
#>     e2 <- err$new_error(conditionMessage(e), call. = conditionCall(e))
#>     class(e2) <- c("callr_remote_error", class(e2))
#>     e2$error <- e
#>     calls <- sys.calls()
#>     dcframe <- which(vapply(calls, function(x) length(x) >= 1 && identical(x[[
#>       1]], quote(do.call)), logical(1)))[1]
#>     if (!is.na(dcframe)) e2$`_ignore` <- list(c(1, dcframe + 1L))
#>     e2$`_pid` <- Sys.getpid()
#>     e2$`_timestamp` <- Sys.time()
#>     if (inherits(e, "rlib_error_2_0")) e2$parent <- e$parent
#>     e2 <- err$add_trace_back(e2, embed = FALSE)
#>     saveRDS(list("error", e2), file = paste0(
#>       "/home/nat/.tmp/RtmpqwGfV0/callr-res-15ea701f9a8237", ".error"))
#>   }
#> }, callr_message = function(e) {
#>   try(signalCondition(e))
#> })
#> 	 -> saveRDS(do.call(do.call, c(readRDS(
#>   "/home/nat/.tmp/RtmpqwGfV0/callr-fun-15ea7056d2377"), list(envir = .GlobalEnv,
#>   quote = TRUE)), envir = .GlobalEnv, quote = TRUE), file = "/home/nat/.tmp/RtmpqwGfV0/callr-res-15ea701f9a8237")
#> 	 -> do.call(do.call, c(readRDS("/home/nat/.tmp/RtmpqwGfV0/callr-fun-15ea7056d2377"),
#> list(envir = .GlobalEnv, quote = TRUE)), envir = .GlobalEnv, quote = TRUE)
#> 	 -> <function(what, args, quote = FALSE, envir = parent.frame()) {
#>   if (!is.list(args)) stop("second argument must be a list")
#>   if (quote) args <- lapply(args, enquote)
#>   .Internal(do.call(what, args, envir))
#> }>(base::quote(<function(..., crayon_enabled, crayon_colors, pkgdown_internet) {
#>   options(crayon.enabled = crayon_enabled, crayon.colors = crayon_colors,
#>     pkgdown.internet = pkgdown_internet)
#>   pkgdown::build_site(...)
#> }>), base::quote(<list: pkg = <pkgdown>, examples = TRUE, run_dont_run = FALSE,
#>   seed = 1014, lazy = FALSE, ...>), envir = base::quote(<environment>), quote = base::quote(
#>   TRUE))
#> 	 -> <function(..., crayon_enabled, crayon_colors, pkgdown_internet) {
#>   options(crayon.enabled = crayon_enabled, crayon.colors = crayon_colors,
#>     pkgdown.internet = pkgdown_internet)
#>   pkgdown::build_site(...)
#> }>(pkg = base::quote(<pkgdown>), examples = base::quote(TRUE), run_dont_run = base::quote(
#>   FALSE), seed = base::quote(1014), lazy = base::quote(FALSE), override = base::quote(
#>   <list: >), install = base::quote(FALSE), preview = base::quote(FALSE),
#> new_process = base::quote(FALSE), devel = base::quote(FALSE), crayon_enabled = base::quote(
#>   TRUE), crayon_colors = base::quote(256L), pkgdown_internet = base::quote(TRUE))
#> 	 -> pkgdown::build_site(...)
#> 	 -> build_site_local(pkg = pkg, examples = examples, run_dont_run = run_dont_run,
#>   seed = seed, lazy = lazy, override = override, preview = preview, devel = devel)
#> 	 -> build_reference(pkg, lazy = lazy, examples = examples, run_dont_run = run_dont_run,
#>   seed = seed, override = override, preview = FALSE, devel = devel)
#> 	 -> purrr::map(topics, build_reference_topic, pkg = pkg, lazy = lazy, examples_env = examples_env,
#>   run_dont_run = run_dont_run)
#> 	 -> .f(.x[[i]], ...)
#> 	 -> withCallingHandlers(data_reference_topic(topic, pkg, examples_env = examples_env,
#>   run_dont_run = run_dont_run), error = function(err) {
#>   msg <- c(paste0("Failed to parse Rd in ", topic$file_in), i = err$message)
#>   abort(msg, parent = err)
#> })
#> 	 -> data_reference_topic(topic, pkg, examples_env = examples_env, run_dont_run = run_dont_run)
#> 	 -> run_examples(tags$tag_examples[[1]], env = if (is.null(examples_env)
#> ) NULL else new.env(parent = examples_env), topic = tools::file_path_sans_ext(
#>   topic$file_in), run_dont_run = run_dont_run)
#> 	 -> highlight_examples(code, topic, env = env)
#> 	 -> downlit::evaluate_and_highlight(code, fig_save = fig_save_topic, env = child_env(
#>   env), output_handler = evaluate::new_output_handler(value = pkgdown_print))
#> 	 -> evaluate::evaluate(code, child_env(env), new_device = TRUE, output_handler = output_handler)
#> 	 -> evaluate_call(expr, parsed$src[[i]], envir = envir, enclos = enclos, debug = debug,
#> last = i == length(out), use_try = stop_on_error != 2L, keep_warning = keep_warning,
#> keep_message = keep_message, output_handler = output_handler, include_timing = include_timing)
#> 	 -> timing_fn(handle(ev <- withCallingHandlers(withVisible(eval(expr, envir, enclos)),
#> warning = wHandler, error = eHandler, message = mHandler)))
#> 	 -> handle(ev <- withCallingHandlers(withVisible(eval(expr, envir, enclos)),
#> warning = wHandler, error = eHandler, message = mHandler))
#> 	 -> try(f, silent = TRUE)
#> 	 -> tryCatch(expr, error = function(e) {
#>   call <- conditionCall(e)
#>   if (!is.null(call)) {
#>     if (identical(call[[1L]], quote(doTryCatch))) call <- sys.call(-4L)
#>     dcall <- deparse(call, nlines = 1L)
#>     prefix <- paste("Error in", dcall, ": ")
#>     LONG <- 75L
#>     sm <- strsplit(conditionMessage(e), "\n")[[1L]]
#>     w <- 14L + nchar(dcall, type = "w") + nchar(sm[1L], type = "w")
#>     if (is.na(w)) w <- 14L + nchar(dcall, type = "b") + nchar(sm[1L], type = "b")
#>     if (w > LONG) prefix <- paste0(prefix, "\n  ")
#>   } else prefix <- "Error : "
#>   msg <- paste0(prefix, conditionMessage(e), "\n")
#>   .Internal(seterrmessage(msg[1L]))
#>   if (!silent && isTRUE(getOption("show.error.messages"))) {
#>     cat(msg, file = outFile)
#>     .Internal(printDeferredWarnings())
#>   }
#>   invisible(structure(msg, class = "try-error", condition = e))
#> })
#> 	 -> tryCatchList(expr, classes, parentenv, handlers)
#> 	 -> tryCatchOne(expr, names, parentenv, handlers[[1L]])
#> 	 -> doTryCatch(return(expr), name, parentenv, handler)
#> 	 -> withCallingHandlers(withVisible(eval(expr, envir, enclos)), warning = wHandler,
#> error = eHandler, message = mHandler)
#> 	 -> withVisible(eval(expr, envir, enclos))
#> 	 -> eval(expr, envir, enclos)
#> 	 -> eval(expr, envir, enclos)
#> 	 -> withinHumdrum(humdata, ~ table(.), by ~ Spine)
#> 	
#> Error: In the call withinHumdrum(humdrumR = _), the argument humdrumR must be a humdrumR object.