Propagate data points to "fill" null data.

ditto is a function that allow you to "fill" null values in a vector with non-null values from earlier/later in the same vector. The default, "forward," behavior fills each null value with the previous (lower index) non-null value, if there are any. The reverse argument can be used to cause "backward" filling, where the next (higher index) non-null value is used.

Usage

ditto(x, ...)

# S3 method for default
ditto(
  x,
  nonnull = function(x) !is.na(x) & x != ".",
  reverse = FALSE,
  boundaries = list()
)

# S3 method for data.frame
ditto(x, ...)

# S3 method for matrix
ditto(x, margin = 2, ...)

# S3 method for humdrumR
ditto(
  x,
  field = getActiveFields(x)[1],
  ...,
  newField = paste0(field, "_ditto")
)

Arguments

x

The input argument. Should be a vector (including lists), matrix, or data.frame.

nonnull

Either a logical vector where (length(x) == length(nonnull)), a numeric vector of positive indices, or a function which, when applied to x returns an appropriate logical/numeric vector.

reverse

(logical & length == 1) If reverse == TRUE, the "non-null" values are coped to overwrite null values earlier (lower indices) in the vector.

boundaries

A vector or list of vectors, all of the same length as x. Each segment of x delineated by the boundaries vector(s) is treated separately.

margin

a vector giving the subscripts which the function will be applied over. E.g., for a matrix 1 indicates rows, 2 indicates columns, c(1, 2) indicates rows and columns. Where X has named dimnames, it can be a character vector selecting dimension names.

field

Which field (partially matched) in the humdrumR dataset should be dittoed?

newField

(character of length == 1) What to name the new (dittoed) field.

Details

Which values are considered "non-null" can be controlled using the nonnull argument. The nonnull argument can either be a logical vector which is the same length as the input (x) argument, a numeric vector of positive indices, or a function which, when applied to x returns an appropriate logical/numeric vector. The values of x where nonnull == TRUE are copied forward/backwards to replace any adjacent vales where nonnull == FALSE. By default, nonnull is the function \(x) !is.na(x) & x != '.', which means that NA values and the string "." are "null", and are overwritten by adjacent values.

ditto methods are defined for data.frames and arrays (including matrices). The data.frame method simply applies ditto to each column of the data.frame separately. For arrays, ditto can be applied across columns (margin == 2), rows (margin == 1), or other dimensions.

The ditto method for a humdrumR object applies ditto to each spine-path within each file in the corpus. The field argument indicates which field to apply ditto to. The result of the dittoing is saved to a new field---the newField argument can be used to control what to name the new field.

Boundaries

In many cases we want to perform lagged calculations in a vector, but not across certain boundaries. For example, if your vector includes data from multiple pieces, we wouldn't want to calculate melodic intervals between pieces, only within pieces. The boundaries argument indicates one, or more, grouping vectors, which break the x (input) argument into groups. If more than boundaries vectors are given, a change in any vector indicates a boundary.

Value pairs which cross between groups are treated as if they were at the beginning. Basically, using boundaries should be essentially identical to using tapply(x, boundaries, laggedFunction, ...), except generally faster when the number of groups is large.

The most common use case in humdrum data, is looking at "melodies" within spines. For this, we want boundaries = list(File, Spine, Path ). In fact, humdrumR with(in) calls will automatically feed these three fields as boundaries arguments to certain functions: mint, delta, sigma, lag, or ditto. Do any use of delta in a call to with(in), will automatically calculate the delta in a "melodic" way, within each spine path of each file. However, if you wanted, for instance, to calculate differences across spines (like harmonic intervals) you could manually set boundaries = list(File, Record).

See also