Visualize data

The draw() function is humdrumR's go-to plotting function. draw() will make a variety of graphs, depending on the type of data you give it. For the most part, draw() is simply a stylish, easy to use wrapper around the base-R graphics functions plot(), barplot(), and hist().

Usage

draw(
  x,
  y,
  facets = list(),
  ...,
  xlab = NULL,
  ylab = NULL,
  axes = 1:4,
  legend = TRUE,
  aspect = NULL,
  margin = 0.2,
  main = "",
  sub = "",
  col = 1,
  cex = NULL
)

Details

draw() is a generic function, which does different plots depending on the data you pass to its x and y arguments.

x and y both numeric: scatter plot.

x numeric by itself: histogram.

y numeric by itself: quantile plot.

x is a table: barplot or heatmap.

y is numeric, x is character or factor: a violin plot.

draw() is simply a wrapper built on top of R's "base" plotting system. This means that all the standard arguments to base-R plots can be used to customize plots (See par() for a full list) or add to them (for example, using points() or mtext()). However, draw() has a number of special additional features, including easily plotting "facets" (dividing data into multiple plots). Anything that draw() does can be done using normal base-R plotting functions (plot(), barplot(), etc.), but draw() makes making good looking plots faster and easier.

The following arguments are the main arguments used with draw(), though not all arguments will have an effect in all types of plots.

Base R:

x: left/right position.

y: up/down position.

col: color.

humdrumR implements a custom color scheme, which can be accessed by using whole number values for col.

cex: point size.

log: draw on log scale?

draw() specific:

square: Should the plot be forced to be square?

alpha: control transparency.

quantiles: mark distribution quantiles.

facets: divide the data into multiple plots.

Specific methods also have additional arguments.

Scatter Plot

If both x and y are real numbers (numeric() in R), a scatter plot is drawn. These vectors should be the same length, but if one of them is length == 1, all the points will simply be drawn in a line at that value.

Coloring points (col arg)

If a single col value is provided, all points are draw this color.

If the col value is the same length as x and y, a scale of colors (either discrete of continuos) is generated to match the values this variable takes.

A legend is drawn.

Point size (cex arg)

If a single cex value is provided, all points are drawn the same size.

If the cex value is numeric and the same length as x and y, a scale of sizes is generated to match the values this variable takes. This auto-sizing should scale with the area of the points.

A legend is drawn.