Calculate Entropy or Information Content of variables

Information content and entropy are fundamental concepts in information theory, quantifying the amount of information in samples from a random variable; they are often characterized as measures of how "expected" (low information) or "surprising" (high information) data is. Both concepts are closely related the probability density/mass of events: improbable events have higher information content. The probability of each (point-wise) observation maps to the information content; The average information content of a variable is the entropy. Information content/entropy can be calculated for discrete probabilities or continuous probabilities, and humdrumR defines methods for calculating both.

Usage

entropy(..., model, base = 2)

H(..., model, base = 2)

# S3 method for probability
entropy(pdist, model, condition = NULL, base = 2)

# S3 method for numeric
entropy(x, model, base = 2, na.rm = TRUE, ...)

# S3 method for density
entropy(x, model, base = 2, na.rm = TRUE)

# S3 method for default
entropy(..., model, base = 2)

xentropy(..., model, base = 2)

kld(..., model, base = 2)

# S3 method for probability
kld(pdist, model, condition = NULL, base = 2)

# S3 method for default
kld(..., model, base = 2)

info(
  ...,
  model,
  base = 2,
  condition = NULL,
  na.rm = FALSE,
  .drop = FALSE,
  binArgs = list()
)

Arguments

...

Distribution (or atomic vectors) to compute entropy/information of.

Must either be a distribution object (created by table(), density(), count(), or pdist()), or one or more atomic vectors of equal length.

If atomic vectors are provided, and model is missing, the atomic vectors are passed to pdist() in order to calculate the model.

model

The expected probability model.

Must either be omitted (not allowed in calls to xentropy()) or must be a probability distribution created by pdist().

In calls to entropy() or info(), if model is missing, ... arguments are used to generate the model.

base

The logarithmic base.

Defaults to 2, so information is measured in units "bits."

Must be a single, non-zero positive number.

Use base = exp(1) for natural-log "nats," or base = 10 for Hartley/"dits".

condition

Compute conditional entropy/information, conditoned on this variable.

Defaults to NULL (no condition), so the joint entropy is calculated.

Must be a non-empty character string, which matches the name of one or of the named variables in the distribution.

This argument is simply passed to pdist().

Details

To calculate information content or entropy, we must assume (or more often, estimate) a probability distribution. HumdrumR's count() and pdist() methods (or R's standard table() function) can be used calculate empirical distributions of atomic data. For numeric, data we can also use R's standard stats::density() function to estimate the continuous probability density.

The entropy() function takes an object representing a probability distribution---ideally a humdrumR distribution object, base-R table, or a density() object (for continuous variables)---and returns the entropy, defaulting to base-2 entropy ("bits"). However, if you are lazy, you can pass entropy() our atomic data vectors directly, and it will automatically pass them to the pdist() function for you; for example, if you want to calculate the joint entropy of variablesx and y (which must be the same length), you can call entropy(pdist(x, y)) or just entropy(x, y). Other arguments can be provided to pdist() as well; notably, if you want to calculate the conditional entropy, you can, for example, say entropy(x, y, condition = 'y').

The info() function is used similarly to the calling entropy() directly on data vectors: anywhere where you can call entropy(x, y), you can call info(x, y) instead. The difference is that info() will return a vector of numbers, the same length as the representing the information content of each input observation. By definition, entropy of the data distribution is the average of all these point-wise information values: thus, mean(info(x, y)) == entropy(x, y).

Cross entropy

In many cases, we simply use entropy/information content to describe a set of data. In this case, the data we observe and the probability model (distribution) are the same---the probability model is the distribution of the data itself. However, we can also use a different model---in this case, a different probability distribution---to describe data. We thus get a measure of how well the model fits the data; this is called the cross entropy. The minimum cross entropy occurs when the data matches the model exactly, and that minimum is the normal "self" entropy of the model. If a data matches the model well, the cross entropy will be a bit higher than the self entropy; if the data matches the model poorly, the cross entropy can be much higher. The difference between the cross entropy and the self entropy is always positive (or zero), and is called the Kullback-Leibler Divergence.

To calculate cross entropy, use the xentropy() command. (The Kullback-Leibler Divergence can be calculated in the same way using the kld() function.) The xentropy() command works just like the entropy command, except you need to provide it a model argument, which must be another probability distribution. Note that that data and the model have to have the exact same variable names, or humdrumR will throw an error! Name your arguments to avoid this (this is illustrated in the example below, where we name everything X). To illustrate, lets create three sets of data, two of which are similar, and one which is very different:

dorian <- c('A', 'B', 'C', 'D', 'E', 'F#', 'G')
N <- 1000

sample1 <- sample(dorian, N, replace = TRUE, prob = 7:1)
sample2 <- sample(dorian, N, replace = TRUE, prob = 7:1)
sample3 <- sample(dorian, N, replace = TRUE, prob = 1:7)

## first the self entropy
entropy(X = sample1)
entropy(X = sample2)
entropy(X = sample3)

## now the cross entropy

xentropy(X = sample1, model = pdist(X = sample2))
xentropy(X = sample2, model = pdist(X = sample2))
xentropy(X = sample3, model = pdist(X = sample2))

sample1 and sample2 have very similar distributions, so when we use sample2 as a model for sample1, the cross entropy is only slightly higher than the self entropy of sample1. However, when we use sample2 as the model for sample3 (which is distributed very differently) the entropy is quite a lot higher than the entropy of sample 3.

The info() command can also be passed a model argument. As always, mean(info(x, model)) == xentropy(x, model). There is no standard name for this "cross information content." However, cross entropy/information-content are closely related to the more general concept of likelihood (see the next section).

Likelihood

The output of info() is identical to the log (base 2 by default) of the modeled likelihood of each data point, which can be computed using the like() function. Literally, info(x, base) == log(like(x), base). The like() function works just like info(), computing pointwise probabilities for each data point based on the probability distribution in model. However, we can use it to, for example, calculate the total log likelihood of data using sum(log(like(...))). This value divided by N is the cross entropy (make sure to use the right log base!): -sum(log(like(...), base = 2)) == xentropy(...).

See also