The "Active expression" of a humdrumR object

humdrumR objects contain many fields of data stored in their underlying humdrum table; You can explicitly access any of these fields using with(in)Humdrum. When you don't explicitly indicate a field, humdrumR will generally default to showing/using an the objects "Active expression".

Usage

evalActive(
  humdrumR,
  dataTypes = "D",
  forceAtomic = TRUE,
  sep = ", ",
  nullChar = FALSE
)

getActive(humdrumR)

getActiveFields(humdrumR)

setActive(humdrumR, expr)

setActiveFields(humdrumR, fieldnames)

# S4 method for humdrumR
$(x, name)

Arguments

humdrumR

HumdrumR data.

Must be a humdrumR data object.

dataTypes

Which types of humdrum records to include in the census.

Defaults to "d".

Must be character. Legal values are 'G', 'L', 'I', 'M', 'D', 'd' or any combination of these (e.g., "LIM"). (See the humdrum table documentation Fields section for explanation.)

forceAtomic

Whether he evaluated active field is forced/coerced into a atomic vector.

Defaults to TRUE.

Must be a singleton logical value: an on/off switch

sep

Separator to paste between collapsed strings.

Defaults to ", ".

Must be a single character string.

Only used if forceAtomic == TRUE, wherein lists of vectors are pasted together.

nullChar

Should null tokens be returned as characters or NA?

Defaults to FALSE.

Must be a singleton logical value: an on/off switch

If nullChar == TRUE and the output is character, NAs in the output vector are replaced with humdrum null character tokens: ".", "!", "=", or "*". If forceAtomic is FALSE, nullChar is ignored.

Details

Most of the time, the active expression just points to a single field: when first read in, the active expression/field is Token. However, the active expression can be any arbitrary R expression involving fields of the humdrum table. When called for, the expression is evaluated within the object's humdrum table (similar to a "within expression," without any extra evaluation options). For instance, the active expression could be: paste(Token, Record), which would print each Token with its record number pasted to it. Any fields referenced in the active expression are called "active fields."

Common commands which evaluate the active expression include:

When is printing a humdrumR object in the terminal, the active expression is shown. (evalActive is used to evaluate the expression as a character string, if needed.) When a humdrumR object prints, the active fields are marked with "*" by their name(s).

When writing to files, the active expression is written.

In a "within expression," the variable . is automatically replaced with the active expression.

Functions like collapseHumdrum], foldHumdrum(), and fields(), use the active field(s) for default arguments.

The current active field can be seen by calling getActive(humData). A character vector of all fields being used by the active expression can be extracted with getActiveFields(humData).

Setting the active expression

The active expression can be changed in several ways. The simplest and most common is using the $ operator, which takes a field name (partially matched) and sets the active expression to simply call that field. This is a handy way to quickly look at different fields in your data:

humData$Token

humData$Spine
humData$Sp # same as last one, because it partially matches Spine

More complex active expressions can be set using setActive, specified directly as the second argument: e.g., setActive(humData, paste(Token, Record)). Notice that the active field must

Refer to at least one field in the humdrum table.

Evaluate to an vector that is same length as the humdrum table (given the target dataTypes), or a list of vectors of that length.

For programmatic work, setActiveFields accepts a character vector of partially matched field names; If one field name is given, the active field just calls that field. If two or more field names are given, the active expression is set to an expression of the form list(Field1, Field2, Field3, ...). This is the easiest way to quickly see two or three fields side by side. As special syntactic sugar, if you call humData$All, a liit of all the data fields is set to the active field. This is a useful way to look at all your data fields.

Null data

humdrumR`` identifies "null data" based on the active field---it might not be obvious, but this is one of the most important jobs of the active field! Anywhere the current active field evaluates to "."orNA(orNULLfor [lists][base::list()]) is considered null data; in the internal [humdrum table][humTable] these data points are set toTRUEin theNullfield and assigned the type"d"in theType` field. Null data is updated whenever the active field is changed or reset, including by functions which create new fields, like foldHumdrum() and within.humdrumR().

As you work, there will often be data tokens which are null in one field, but not in another field. For example, if you load **kern data, a token like "4r" (quarter-note rest) token will be NA if you call pitch, but not NA if you call recip (rhythm).

kerndata <- readHumdrum(...)

within(kerndata$Token,
       Pitch  <- pitch(.),
       Rhythm <- recip(.)) -> kerndata

Now, if you change the active field between Pitch and Rhythm (using $) you'll see that there are different numbers of (non-null) data tokens: ntoken(kerndata$Pitch ,'D') vs ntoken(kerndata$Rhythm, 'D) will return different numbers! (The difference would be the number of rest tokens.) Similarly, if you apply functions/expressions to this data (using withinHumdrum() for example), the result will depend on what the active field is:

with(kerndata$Pitch, length(.))
with(kerndata$Rhythm, length(.))

Once again, we'll get different numbers here! (Assuming there are rests in the data.) This is the case even though the do-expression isn't actually using the Pitch or Rhythm fields! If Pitch is the active field, the rest tokens are null-data and will be ignored!

Evaluating the active expression

Evaluation of the active expression is usually something done automatically by humdrumR functions, especially for printing data at the console. However, you can also do it manually using the evalActive command. The "raw" result of evaluating the active expression can be returned by specifying forceAtomic == FALSE. However, by default forceAtomic == TRUE which causes evalActive to coerce the evaluated results into an atomic vector. Obviously, the evaluated active result is an atomic vector, no coercion is needed.

If the evaluated active result is a list, it must be either the full length of the humdrum table, or a list of vectors/lists of that length. In other words, the result must be one or more "full length" vector/lists. For each full length list, each element of the list is coerced to a single atomic value and then unlisted to create an atomic vector. If the elements of the list are not themselves atomic, they are converted to various character representations.

tables are coerced to the string "<table: k=x, n=y>", where x is the number of categories in the table and y is the total number of values in the table (sum(table(...))).

lists of length < 5 are coerced to "list(a, b, c, d, e)", where a-e are the elements of the list. Longer lists are coerced to "list[n]:, where n is the length of the list.

All other R objects are coerced to <class>, where class is the class of the object.

Finally, all the thus-generated full-length vectors (if there are more than one) are pasted together, separated by sep (default = ", "). A common practical illustration/application of this last is to specify active fields that are lists of fields---for example, list(Token, Spine, Record). Following the algorithm above, the evaluated result is would be character vector looking like "Token, Spine, Record". This is exactly what setActiveFields does when fed multiple fieldNames, as well as the special call humData$All.