Tabulate records and tokens in a humdrumR corpus

census is one of humdrumR's summary functions, used to tabulate the raw size of a humdrumR corpus. census takes a humdrumR object and and returns a humCensus table. The dataType argument controls what types of records to tabulate: legal values are 'G', 'L', 'I', 'M', 'D', 'd' or any combination of these (e.g., "LIM"). The default is "D".

Usage

census(humdrumR, dataTypes = "GLIMDd", by = "Filename", removeEmpty = FALSE)

census(humdata)[i]

# S3 method for humCensus
print(censusTable, showEach = TRUE, screenWidth = options("width")$width - 10L)

Arguments

humdrumR

A humdrumR object

dataTypes

A character string of length 1.

by

A character string of length 1. Must be a field in the humdrumR object.

Details

A humCensus table has five columns of information:

Records

The total number of records.

Tokens

The total number of tokens.

(unique)

The number of unique tokens

Characters

The total numder of characters. This includes humdrum control characters like * and !!.

(per token)

This is simply Characters / Tokens, indicating the mean length of each token.

A humCensus table has one row for each file in the corpus. Rows are labeled with each file's corresponding number (from the humTable's File field) and name (the Filename field). In addition, when a humCensus object is printed, the totals across all files are printed as well---(unique) and (per token) across all files are calculated across all files as well, not summed.

Indexing

Rows of a humCensus object can be selected with a single argument i: e.g., censusTable[i]. If i is numeric, the corresponding rows are selected ordinally (not by File number). If i is a character string, this string is mached as a regular expression against file names. If i is a formula, the right-hand side of the formula is evaluated within the table---if it evaluates to a logical vector, files are selected accordingly. For instance, censusTable[~Tokens > 100] will select all files with more than 100 tokens. (The '(unique)' and '(per token)' columns must be referred to with their names enclosed in \``---for example, censusTable~\verb(unique)\ > 100` will return all files with more than 100 unique tokens.

A drop argument is also available. If TRUE, a plain data.table::data.table() is returned.

See also