Piping humdrumR data

These infix operators make it possible to use the [humdrumR:with-in-Humdrum][with(in)Humdrum] functions in a piping style, similar to the | (pipe) in unix-style terminals, or the |> pipe operator. from the R package magrittr.

Usage

humdrumR %hum>% formula

humdrumR %hum<% formula

humdrumR %humT% formula

e1 %s>% e2

list %hums>% formula

list %hums<% formula

list %humsT% formula

list %hums[]% formula

Details

The key is that the function [withinHumdrum][withinHumdrum] always returns a new [humdrumR][humdrumR] data object. Thus, you can always send the output of [withinHumdrum][withinHumdrum] back in to a new call of [withinHumdrum][withinHumdrum]. This is exactly what the %hum>% is for: on the left-hand side, input a [humdrumR][humdrumR] object, on the right-hand side suitable arguments to [withinHumdrum][withinHumdrum] (i.e., a formula, a function, or list of formulae, functions, and named arguments). You can then chain this call with another call to %hum>% and more [withinHumdrum][withinHumdrum] arguments.

%hum<% acts the same way as %hum>% except it calls [withHumdrum][withinHumdrum]. Since [withHumdrum][withinHumdrum] does not return a [humdrumR][humdrumR], the output can't be piped any further (using %hum>% or %hum<%). Thus, \%hum<\% should only be used as the last step in a pipe---you would do this if you want to extract the last step in your pipe from the data's [humdrumR:humtable][Humdrum Table] into a normal vector or list of R data.

' \%humT\% creates a "T" in the pipe, applying the desired expression but not keeping the result---the unaltered humdrumR input object is returned. This works simply by replacing all do~ with doplot~ in a call to [withinHumdrum][withinHumdrum]. The purpose of this option, is if you want to apply expressions for their side effects, for instance, for plotting.

%hum[]% is similar to %hum>% except it apply the formulae on its right-hand side using [humdrumR][filterHumdrum]. Thus, it can be used to filter/index a [humdrumR][humdrumR] data object on the fly.

Plural pipes

In R we often apply the same function to a list of data. "Plural pipes" expand this idea to piping: take a list of data and pipe each element in the list to an expression/function. There are plural pipe versions of each singular pipe operator. Just add an "s" to make them plural:

|> (singular) : %s>% (plural);

%hum>% (singular) : %hums>% (plural);

%hum<% (singular) : %hums<% (plural);

%humT% (singular) : %humsT% (plural);

%hum[]% (singular) : %hums[]% (plural);

Note: %s>% is an expansion of the \> pipe operator, which makes use of some clever meta-programming---we can't guarantee it will always behave the way we might like!

Examples

humdata <- readHumdrum('path*.krn')
#> Finding and reading files...
#> 	REpath-pattern 'path*.krn' matches 0 text files in 1 directory.
#> Zero files read from disk.
#> No files to validate.

humdata %hum>% ~table(.)
#> humdrumR error in:
#> 	tryCatch(withCallingHandlers({
#>   NULL
#>   saveRDS(do.call(do.call, c(readRDS(
#>     "/home/nat/.tmp/RtmpAJb3S4/callr-fun-2f64c25cfb0d5"), list(envir = .GlobalEnv,
#>     quote = TRUE)), envir = .GlobalEnv, quote = TRUE), file = "/home/nat/.tmp/RtmpAJb3S4/callr-res-2f64c750f9f6e")
#>   flush(stdout())
#>   flush(stderr())
#>   NULL
#>   invisible()
#> }, error = function(e) {
#>   {
#>     callr_data <- as.environment("tools:callr")$`__callr_data__`
#>     err <- callr_data$err
#>     assign(".Traceback", .traceback(4), envir = callr_data)
#>     dump.frames("__callr_dump__")
#>     assign(".Last.dump", .GlobalEnv$`__callr_dump__`, envir = callr_data)
#>     rm("__callr_dump__", envir = .GlobalEnv)
#>     e$call <- deparse(conditionCall(e), nlines = 6)
#>     e2 <- err$new_error(conditionMessage(e), call. = conditionCall(e))
#>     class(e2) <- c("callr_remote_error", class(e2))
#>     e2$error <- e
#>     calls <- sys.calls()
#>     dcframe <- which(vapply(calls, function(x) length(x) >= 1 && identical(x[[
#>       1]], quote(do.call)), logical(1)))[1]
#>     if (!is.na(dcframe)) e2$`_ignore` <- list(c(1, dcframe + 1L))
#>     e2$`_pid` <- Sys.getpid()
#>     e2$`_timestamp` <- Sys.time()
#>     if (inherits(e, "rlib_error_2_0")) e2$parent <- e$parent
#>     e2 <- err$add_trace_back(e2, embed = FALSE)
#>     saveRDS(list("error", e2), file = paste0(
#>       "/home/nat/.tmp/RtmpAJb3S4/callr-res-2f64c750f9f6e", ".error"))
#>   }
#> }, interrupt = function(e) {
#>   {
#>     callr_data <- as.environment("tools:callr")$`__callr_data__`
#>     err <- callr_data$err
#>     assign(".Traceback", .traceback(4), envir = callr_data)
#>     dump.frames("__callr_dump__")
#>     assign(".Last.dump", .GlobalEnv$`__callr_dump__`, envir = callr_data)
#>     rm("__callr_dump__", envir = .GlobalEnv)
#>     e$call <- deparse(conditionCall(e), nlines = 6)
#>     e2 <- err$new_error(conditionMessage(e), call. = conditionCall(e))
#>     class(e2) <- c("callr_remote_error", class(e2))
#>     e2$error <- e
#>     calls <- sys.calls()
#>     dcframe <- which(vapply(calls, function(x) length(x) >= 1 && identical(x[[
#>       1]], quote(do.call)), logical(1)))[1]
#>     if (!is.na(dcframe)) e2$`_ignore` <- list(c(1, dcframe + 1L))
#>     e2$`_pid` <- Sys.getpid()
#>     e2$`_timestamp` <- Sys.time()
#>     if (inherits(e, "rlib_error_2_0")) e2$parent <- e$parent
#>     e2 <- err$add_trace_back(e2, embed = FALSE)
#>     saveRDS(list("error", e2), file = paste0(
#>       "/home/nat/.tmp/RtmpAJb3S4/callr-res-2f64c750f9f6e", ".error"))
#>   }
#> }, callr_message = function(e) {
#>   try(signalCondition(e))
#> }), error = function(e) {
#>   NULL
#>   try(stop(e))
#> }, interrupt = function(e) {
#>   NULL
#>   e
#> })
#> 	 -> tryCatchList(expr, classes, parentenv, handlers)
#> 	 -> tryCatchOne(tryCatchList(expr, names[-nh], parentenv, handlers[-nh]), names[nh],
#> parentenv, handlers[[nh]])
#> 	 -> doTryCatch(return(expr), name, parentenv, handler)
#> 	 -> tryCatchList(expr, names[-nh], parentenv, handlers[-nh])
#> 	 -> tryCatchOne(expr, names, parentenv, handlers[[1L]])
#> 	 -> doTryCatch(return(expr), name, parentenv, handler)
#> 	 -> withCallingHandlers({
#>   NULL
#>   saveRDS(do.call(do.call, c(readRDS(
#>     "/home/nat/.tmp/RtmpAJb3S4/callr-fun-2f64c25cfb0d5"), list(envir = .GlobalEnv,
#>     quote = TRUE)), envir = .GlobalEnv, quote = TRUE), file = "/home/nat/.tmp/RtmpAJb3S4/callr-res-2f64c750f9f6e")
#>   flush(stdout())
#>   flush(stderr())
#>   NULL
#>   invisible()
#> }, error = function(e) {
#>   {
#>     callr_data <- as.environment("tools:callr")$`__callr_data__`
#>     err <- callr_data$err
#>     assign(".Traceback", .traceback(4), envir = callr_data)
#>     dump.frames("__callr_dump__")
#>     assign(".Last.dump", .GlobalEnv$`__callr_dump__`, envir = callr_data)
#>     rm("__callr_dump__", envir = .GlobalEnv)
#>     e$call <- deparse(conditionCall(e), nlines = 6)
#>     e2 <- err$new_error(conditionMessage(e), call. = conditionCall(e))
#>     class(e2) <- c("callr_remote_error", class(e2))
#>     e2$error <- e
#>     calls <- sys.calls()
#>     dcframe <- which(vapply(calls, function(x) length(x) >= 1 && identical(x[[
#>       1]], quote(do.call)), logical(1)))[1]
#>     if (!is.na(dcframe)) e2$`_ignore` <- list(c(1, dcframe + 1L))
#>     e2$`_pid` <- Sys.getpid()
#>     e2$`_timestamp` <- Sys.time()
#>     if (inherits(e, "rlib_error_2_0")) e2$parent <- e$parent
#>     e2 <- err$add_trace_back(e2, embed = FALSE)
#>     saveRDS(list("error", e2), file = paste0(
#>       "/home/nat/.tmp/RtmpAJb3S4/callr-res-2f64c750f9f6e", ".error"))
#>   }
#> }, interrupt = function(e) {
#>   {
#>     callr_data <- as.environment("tools:callr")$`__callr_data__`
#>     err <- callr_data$err
#>     assign(".Traceback", .traceback(4), envir = callr_data)
#>     dump.frames("__callr_dump__")
#>     assign(".Last.dump", .GlobalEnv$`__callr_dump__`, envir = callr_data)
#>     rm("__callr_dump__", envir = .GlobalEnv)
#>     e$call <- deparse(conditionCall(e), nlines = 6)
#>     e2 <- err$new_error(conditionMessage(e), call. = conditionCall(e))
#>     class(e2) <- c("callr_remote_error", class(e2))
#>     e2$error <- e
#>     calls <- sys.calls()
#>     dcframe <- which(vapply(calls, function(x) length(x) >= 1 && identical(x[[
#>       1]], quote(do.call)), logical(1)))[1]
#>     if (!is.na(dcframe)) e2$`_ignore` <- list(c(1, dcframe + 1L))
#>     e2$`_pid` <- Sys.getpid()
#>     e2$`_timestamp` <- Sys.time()
#>     if (inherits(e, "rlib_error_2_0")) e2$parent <- e$parent
#>     e2 <- err$add_trace_back(e2, embed = FALSE)
#>     saveRDS(list("error", e2), file = paste0(
#>       "/home/nat/.tmp/RtmpAJb3S4/callr-res-2f64c750f9f6e", ".error"))
#>   }
#> }, callr_message = function(e) {
#>   try(signalCondition(e))
#> })
#> 	 -> saveRDS(do.call(do.call, c(readRDS(
#>   "/home/nat/.tmp/RtmpAJb3S4/callr-fun-2f64c25cfb0d5"), list(envir = .GlobalEnv,
#>   quote = TRUE)), envir = .GlobalEnv, quote = TRUE), file = "/home/nat/.tmp/RtmpAJb3S4/callr-res-2f64c750f9f6e")
#> 	 -> do.call(do.call, c(readRDS("/home/nat/.tmp/RtmpAJb3S4/callr-fun-2f64c25cfb0d5"),
#> list(envir = .GlobalEnv, quote = TRUE)), envir = .GlobalEnv, quote = TRUE)
#> 	 -> <function(what, args, quote = FALSE, envir = parent.frame()) {
#>   if (!is.list(args)) stop("second argument must be a list")
#>   if (quote) args <- lapply(args, enquote)
#>   .Internal(do.call(what, args, envir))
#> }>(base::quote(<function(..., crayon_enabled, crayon_colors, pkgdown_internet) {
#>   options(crayon.enabled = crayon_enabled, crayon.colors = crayon_colors,
#>     pkgdown.internet = pkgdown_internet)
#>   pkgdown::build_site(...)
#> }>), base::quote(<list: pkg = <pkgdown>, examples = TRUE, run_dont_run = FALSE,
#>   seed = 1014, lazy = FALSE, ...>), envir = base::quote(<environment>), quote = base::quote(
#>   TRUE))
#> 	 -> <function(..., crayon_enabled, crayon_colors, pkgdown_internet) {
#>   options(crayon.enabled = crayon_enabled, crayon.colors = crayon_colors,
#>     pkgdown.internet = pkgdown_internet)
#>   pkgdown::build_site(...)
#> }>(pkg = base::quote(<pkgdown>), examples = base::quote(TRUE), run_dont_run = base::quote(
#>   FALSE), seed = base::quote(1014), lazy = base::quote(FALSE), override = base::quote(
#>   <list: >), install = base::quote(FALSE), preview = base::quote(FALSE),
#> new_process = base::quote(FALSE), devel = base::quote(FALSE), crayon_enabled = base::quote(
#>   TRUE), crayon_colors = base::quote(256L), pkgdown_internet = base::quote(TRUE))
#> 	 -> pkgdown::build_site(...)
#> 	 -> build_site_local(pkg = pkg, examples = examples, run_dont_run = run_dont_run,
#>   seed = seed, lazy = lazy, override = override, preview = preview, devel = devel)
#> 	 -> build_reference(pkg, lazy = lazy, examples = examples, run_dont_run = run_dont_run,
#>   seed = seed, override = override, preview = FALSE, devel = devel)
#> 	 -> purrr::map(topics, build_reference_topic, pkg = pkg, lazy = lazy, examples_env = examples_env,
#>   run_dont_run = run_dont_run)
#> 	 -> .f(.x[[i]], ...)
#> 	 -> withCallingHandlers(data_reference_topic(topic, pkg, examples_env = examples_env,
#>   run_dont_run = run_dont_run), error = function(err) {
#>   msg <- c(paste0("Failed to parse Rd in ", topic$file_in), i = err$message)
#>   abort(msg, parent = err)
#> })
#> 	 -> data_reference_topic(topic, pkg, examples_env = examples_env, run_dont_run = run_dont_run)
#> 	 -> run_examples(tags$tag_examples[[1]], env = if (is.null(examples_env)
#> ) NULL else new.env(parent = examples_env), topic = tools::file_path_sans_ext(
#>   topic$file_in), run_dont_run = run_dont_run)
#> 	 -> highlight_examples(code, topic, env = env)
#> 	 -> downlit::evaluate_and_highlight(code, fig_save = fig_save_topic, env = child_env(
#>   env), output_handler = evaluate::new_output_handler(value = pkgdown_print))
#> 	 -> evaluate::evaluate(code, child_env(env), new_device = TRUE, output_handler = output_handler)
#> 	 -> evaluate_call(expr, parsed$src[[i]], envir = envir, enclos = enclos, debug = debug,
#> last = i == length(out), use_try = stop_on_error != 2L, keep_warning = keep_warning,
#> keep_message = keep_message, output_handler = output_handler, include_timing = include_timing)
#> 	 -> timing_fn(handle(ev <- withCallingHandlers(withVisible(eval(expr, envir, enclos)),
#> warning = wHandler, error = eHandler, message = mHandler)))
#> 	 -> handle(ev <- withCallingHandlers(withVisible(eval(expr, envir, enclos)),
#> warning = wHandler, error = eHandler, message = mHandler))
#> 	 -> try(f, silent = TRUE)
#> 	 -> tryCatch(expr, error = function(e) {
#>   call <- conditionCall(e)
#>   if (!is.null(call)) {
#>     if (identical(call[[1L]], quote(doTryCatch))) call <- sys.call(-4L)
#>     dcall <- deparse(call, nlines = 1L)
#>     prefix <- paste("Error in", dcall, ": ")
#>     LONG <- 75L
#>     sm <- strsplit(conditionMessage(e), "\n")[[1L]]
#>     w <- 14L + nchar(dcall, type = "w") + nchar(sm[1L], type = "w")
#>     if (is.na(w)) w <- 14L + nchar(dcall, type = "b") + nchar(sm[1L], type = "b")
#>     if (w > LONG) prefix <- paste0(prefix, "\n  ")
#>   } else prefix <- "Error : "
#>   msg <- paste0(prefix, conditionMessage(e), "\n")
#>   .Internal(seterrmessage(msg[1L]))
#>   if (!silent && isTRUE(getOption("show.error.messages"))) {
#>     cat(msg, file = outFile)
#>     .Internal(printDeferredWarnings())
#>   }
#>   invisible(structure(msg, class = "try-error", condition = e))
#> })
#> 	 -> tryCatchList(expr, classes, parentenv, handlers)
#> 	 -> tryCatchOne(expr, names, parentenv, handlers[[1L]])
#> 	 -> doTryCatch(return(expr), name, parentenv, handler)
#> 	 -> withCallingHandlers(withVisible(eval(expr, envir, enclos)), warning = wHandler,
#> error = eHandler, message = mHandler)
#> 	 -> withVisible(eval(expr, envir, enclos))
#> 	 -> eval(expr, envir, enclos)
#> 	 -> eval(expr, envir, enclos)
#> 	 -> humdata %hum>% (~ table(.))
#> 	
#> Error: In the call `%hum>%`(humdrumR = _), the argument humdrumR must be a humdrumR object.

humdata %hum>%
     c(by~Spine, do ~ table(.)) %hum>%
     c(mfcol ~ c(2,2), doplot~barplot(.))
#> humdrumR error in:
#> 	tryCatch(withCallingHandlers({
#>   NULL
#>   saveRDS(do.call(do.call, c(readRDS(
#>     "/home/nat/.tmp/RtmpAJb3S4/callr-fun-2f64c25cfb0d5"), list(envir = .GlobalEnv,
#>     quote = TRUE)), envir = .GlobalEnv, quote = TRUE), file = "/home/nat/.tmp/RtmpAJb3S4/callr-res-2f64c750f9f6e")
#>   flush(stdout())
#>   flush(stderr())
#>   NULL
#>   invisible()
#> }, error = function(e) {
#>   {
#>     callr_data <- as.environment("tools:callr")$`__callr_data__`
#>     err <- callr_data$err
#>     assign(".Traceback", .traceback(4), envir = callr_data)
#>     dump.frames("__callr_dump__")
#>     assign(".Last.dump", .GlobalEnv$`__callr_dump__`, envir = callr_data)
#>     rm("__callr_dump__", envir = .GlobalEnv)
#>     e$call <- deparse(conditionCall(e), nlines = 6)
#>     e2 <- err$new_error(conditionMessage(e), call. = conditionCall(e))
#>     class(e2) <- c("callr_remote_error", class(e2))
#>     e2$error <- e
#>     calls <- sys.calls()
#>     dcframe <- which(vapply(calls, function(x) length(x) >= 1 && identical(x[[
#>       1]], quote(do.call)), logical(1)))[1]
#>     if (!is.na(dcframe)) e2$`_ignore` <- list(c(1, dcframe + 1L))
#>     e2$`_pid` <- Sys.getpid()
#>     e2$`_timestamp` <- Sys.time()
#>     if (inherits(e, "rlib_error_2_0")) e2$parent <- e$parent
#>     e2 <- err$add_trace_back(e2, embed = FALSE)
#>     saveRDS(list("error", e2), file = paste0(
#>       "/home/nat/.tmp/RtmpAJb3S4/callr-res-2f64c750f9f6e", ".error"))
#>   }
#> }, interrupt = function(e) {
#>   {
#>     callr_data <- as.environment("tools:callr")$`__callr_data__`
#>     err <- callr_data$err
#>     assign(".Traceback", .traceback(4), envir = callr_data)
#>     dump.frames("__callr_dump__")
#>     assign(".Last.dump", .GlobalEnv$`__callr_dump__`, envir = callr_data)
#>     rm("__callr_dump__", envir = .GlobalEnv)
#>     e$call <- deparse(conditionCall(e), nlines = 6)
#>     e2 <- err$new_error(conditionMessage(e), call. = conditionCall(e))
#>     class(e2) <- c("callr_remote_error", class(e2))
#>     e2$error <- e
#>     calls <- sys.calls()
#>     dcframe <- which(vapply(calls, function(x) length(x) >= 1 && identical(x[[
#>       1]], quote(do.call)), logical(1)))[1]
#>     if (!is.na(dcframe)) e2$`_ignore` <- list(c(1, dcframe + 1L))
#>     e2$`_pid` <- Sys.getpid()
#>     e2$`_timestamp` <- Sys.time()
#>     if (inherits(e, "rlib_error_2_0")) e2$parent <- e$parent
#>     e2 <- err$add_trace_back(e2, embed = FALSE)
#>     saveRDS(list("error", e2), file = paste0(
#>       "/home/nat/.tmp/RtmpAJb3S4/callr-res-2f64c750f9f6e", ".error"))
#>   }
#> }, callr_message = function(e) {
#>   try(signalCondition(e))
#> }), error = function(e) {
#>   NULL
#>   try(stop(e))
#> }, interrupt = function(e) {
#>   NULL
#>   e
#> })
#> 	 -> tryCatchList(expr, classes, parentenv, handlers)
#> 	 -> tryCatchOne(tryCatchList(expr, names[-nh], parentenv, handlers[-nh]), names[nh],
#> parentenv, handlers[[nh]])
#> 	 -> doTryCatch(return(expr), name, parentenv, handler)
#> 	 -> tryCatchList(expr, names[-nh], parentenv, handlers[-nh])
#> 	 -> tryCatchOne(expr, names, parentenv, handlers[[1L]])
#> 	 -> doTryCatch(return(expr), name, parentenv, handler)
#> 	 -> withCallingHandlers({
#>   NULL
#>   saveRDS(do.call(do.call, c(readRDS(
#>     "/home/nat/.tmp/RtmpAJb3S4/callr-fun-2f64c25cfb0d5"), list(envir = .GlobalEnv,
#>     quote = TRUE)), envir = .GlobalEnv, quote = TRUE), file = "/home/nat/.tmp/RtmpAJb3S4/callr-res-2f64c750f9f6e")
#>   flush(stdout())
#>   flush(stderr())
#>   NULL
#>   invisible()
#> }, error = function(e) {
#>   {
#>     callr_data <- as.environment("tools:callr")$`__callr_data__`
#>     err <- callr_data$err
#>     assign(".Traceback", .traceback(4), envir = callr_data)
#>     dump.frames("__callr_dump__")
#>     assign(".Last.dump", .GlobalEnv$`__callr_dump__`, envir = callr_data)
#>     rm("__callr_dump__", envir = .GlobalEnv)
#>     e$call <- deparse(conditionCall(e), nlines = 6)
#>     e2 <- err$new_error(conditionMessage(e), call. = conditionCall(e))
#>     class(e2) <- c("callr_remote_error", class(e2))
#>     e2$error <- e
#>     calls <- sys.calls()
#>     dcframe <- which(vapply(calls, function(x) length(x) >= 1 && identical(x[[
#>       1]], quote(do.call)), logical(1)))[1]
#>     if (!is.na(dcframe)) e2$`_ignore` <- list(c(1, dcframe + 1L))
#>     e2$`_pid` <- Sys.getpid()
#>     e2$`_timestamp` <- Sys.time()
#>     if (inherits(e, "rlib_error_2_0")) e2$parent <- e$parent
#>     e2 <- err$add_trace_back(e2, embed = FALSE)
#>     saveRDS(list("error", e2), file = paste0(
#>       "/home/nat/.tmp/RtmpAJb3S4/callr-res-2f64c750f9f6e", ".error"))
#>   }
#> }, interrupt = function(e) {
#>   {
#>     callr_data <- as.environment("tools:callr")$`__callr_data__`
#>     err <- callr_data$err
#>     assign(".Traceback", .traceback(4), envir = callr_data)
#>     dump.frames("__callr_dump__")
#>     assign(".Last.dump", .GlobalEnv$`__callr_dump__`, envir = callr_data)
#>     rm("__callr_dump__", envir = .GlobalEnv)
#>     e$call <- deparse(conditionCall(e), nlines = 6)
#>     e2 <- err$new_error(conditionMessage(e), call. = conditionCall(e))
#>     class(e2) <- c("callr_remote_error", class(e2))
#>     e2$error <- e
#>     calls <- sys.calls()
#>     dcframe <- which(vapply(calls, function(x) length(x) >= 1 && identical(x[[
#>       1]], quote(do.call)), logical(1)))[1]
#>     if (!is.na(dcframe)) e2$`_ignore` <- list(c(1, dcframe + 1L))
#>     e2$`_pid` <- Sys.getpid()
#>     e2$`_timestamp` <- Sys.time()
#>     if (inherits(e, "rlib_error_2_0")) e2$parent <- e$parent
#>     e2 <- err$add_trace_back(e2, embed = FALSE)
#>     saveRDS(list("error", e2), file = paste0(
#>       "/home/nat/.tmp/RtmpAJb3S4/callr-res-2f64c750f9f6e", ".error"))
#>   }
#> }, callr_message = function(e) {
#>   try(signalCondition(e))
#> })
#> 	 -> saveRDS(do.call(do.call, c(readRDS(
#>   "/home/nat/.tmp/RtmpAJb3S4/callr-fun-2f64c25cfb0d5"), list(envir = .GlobalEnv,
#>   quote = TRUE)), envir = .GlobalEnv, quote = TRUE), file = "/home/nat/.tmp/RtmpAJb3S4/callr-res-2f64c750f9f6e")
#> 	 -> do.call(do.call, c(readRDS("/home/nat/.tmp/RtmpAJb3S4/callr-fun-2f64c25cfb0d5"),
#> list(envir = .GlobalEnv, quote = TRUE)), envir = .GlobalEnv, quote = TRUE)
#> 	 -> <function(what, args, quote = FALSE, envir = parent.frame()) {
#>   if (!is.list(args)) stop("second argument must be a list")
#>   if (quote) args <- lapply(args, enquote)
#>   .Internal(do.call(what, args, envir))
#> }>(base::quote(<function(..., crayon_enabled, crayon_colors, pkgdown_internet) {
#>   options(crayon.enabled = crayon_enabled, crayon.colors = crayon_colors,
#>     pkgdown.internet = pkgdown_internet)
#>   pkgdown::build_site(...)
#> }>), base::quote(<list: pkg = <pkgdown>, examples = TRUE, run_dont_run = FALSE,
#>   seed = 1014, lazy = FALSE, ...>), envir = base::quote(<environment>), quote = base::quote(
#>   TRUE))
#> 	 -> <function(..., crayon_enabled, crayon_colors, pkgdown_internet) {
#>   options(crayon.enabled = crayon_enabled, crayon.colors = crayon_colors,
#>     pkgdown.internet = pkgdown_internet)
#>   pkgdown::build_site(...)
#> }>(pkg = base::quote(<pkgdown>), examples = base::quote(TRUE), run_dont_run = base::quote(
#>   FALSE), seed = base::quote(1014), lazy = base::quote(FALSE), override = base::quote(
#>   <list: >), install = base::quote(FALSE), preview = base::quote(FALSE),
#> new_process = base::quote(FALSE), devel = base::quote(FALSE), crayon_enabled = base::quote(
#>   TRUE), crayon_colors = base::quote(256L), pkgdown_internet = base::quote(TRUE))
#> 	 -> pkgdown::build_site(...)
#> 	 -> build_site_local(pkg = pkg, examples = examples, run_dont_run = run_dont_run,
#>   seed = seed, lazy = lazy, override = override, preview = preview, devel = devel)
#> 	 -> build_reference(pkg, lazy = lazy, examples = examples, run_dont_run = run_dont_run,
#>   seed = seed, override = override, preview = FALSE, devel = devel)
#> 	 -> purrr::map(topics, build_reference_topic, pkg = pkg, lazy = lazy, examples_env = examples_env,
#>   run_dont_run = run_dont_run)
#> 	 -> .f(.x[[i]], ...)
#> 	 -> withCallingHandlers(data_reference_topic(topic, pkg, examples_env = examples_env,
#>   run_dont_run = run_dont_run), error = function(err) {
#>   msg <- c(paste0("Failed to parse Rd in ", topic$file_in), i = err$message)
#>   abort(msg, parent = err)
#> })
#> 	 -> data_reference_topic(topic, pkg, examples_env = examples_env, run_dont_run = run_dont_run)
#> 	 -> run_examples(tags$tag_examples[[1]], env = if (is.null(examples_env)
#> ) NULL else new.env(parent = examples_env), topic = tools::file_path_sans_ext(
#>   topic$file_in), run_dont_run = run_dont_run)
#> 	 -> highlight_examples(code, topic, env = env)
#> 	 -> downlit::evaluate_and_highlight(code, fig_save = fig_save_topic, env = child_env(
#>   env), output_handler = evaluate::new_output_handler(value = pkgdown_print))
#> 	 -> evaluate::evaluate(code, child_env(env), new_device = TRUE, output_handler = output_handler)
#> 	 -> evaluate_call(expr, parsed$src[[i]], envir = envir, enclos = enclos, debug = debug,
#> last = i == length(out), use_try = stop_on_error != 2L, keep_warning = keep_warning,
#> keep_message = keep_message, output_handler = output_handler, include_timing = include_timing)
#> 	 -> timing_fn(handle(ev <- withCallingHandlers(withVisible(eval(expr, envir, enclos)),
#> warning = wHandler, error = eHandler, message = mHandler)))
#> 	 -> handle(ev <- withCallingHandlers(withVisible(eval(expr, envir, enclos)),
#> warning = wHandler, error = eHandler, message = mHandler))
#> 	 -> try(f, silent = TRUE)
#> 	 -> tryCatch(expr, error = function(e) {
#>   call <- conditionCall(e)
#>   if (!is.null(call)) {
#>     if (identical(call[[1L]], quote(doTryCatch))) call <- sys.call(-4L)
#>     dcall <- deparse(call, nlines = 1L)
#>     prefix <- paste("Error in", dcall, ": ")
#>     LONG <- 75L
#>     sm <- strsplit(conditionMessage(e), "\n")[[1L]]
#>     w <- 14L + nchar(dcall, type = "w") + nchar(sm[1L], type = "w")
#>     if (is.na(w)) w <- 14L + nchar(dcall, type = "b") + nchar(sm[1L], type = "b")
#>     if (w > LONG) prefix <- paste0(prefix, "\n  ")
#>   } else prefix <- "Error : "
#>   msg <- paste0(prefix, conditionMessage(e), "\n")
#>   .Internal(seterrmessage(msg[1L]))
#>   if (!silent && isTRUE(getOption("show.error.messages"))) {
#>     cat(msg, file = outFile)
#>     .Internal(printDeferredWarnings())
#>   }
#>   invisible(structure(msg, class = "try-error", condition = e))
#> })
#> 	 -> tryCatchList(expr, classes, parentenv, handlers)
#> 	 -> tryCatchOne(expr, names, parentenv, handlers[[1L]])
#> 	 -> doTryCatch(return(expr), name, parentenv, handler)
#> 	 -> withCallingHandlers(withVisible(eval(expr, envir, enclos)), warning = wHandler,
#> error = eHandler, message = mHandler)
#> 	 -> withVisible(eval(expr, envir, enclos))
#> 	 -> eval(expr, envir, enclos)
#> 	 -> eval(expr, envir, enclos)
#> 	 -> humdata %hum>% c(by ~ Spine, do ~ table(.)) %hum>% c(mfcol ~ c(2, 2), doplot ~
#>   barplot(.))
#> 	 -> is.humdrumR((x))
#> 	 -> humdata %hum>% c(by ~ Spine, do ~ table(.))
#> 	
#> Error: In the call `%hum>%`(humdrumR = _), the argument humdrumR must be a humdrumR object.
     
humdata %hum>%
     c(by ~ Spine, do ~ table(.)) %hum<%
     (do ~ sort(.))    
#> humdrumR error in:
#> 	tryCatch(withCallingHandlers({
#>   NULL
#>   saveRDS(do.call(do.call, c(readRDS(
#>     "/home/nat/.tmp/RtmpAJb3S4/callr-fun-2f64c25cfb0d5"), list(envir = .GlobalEnv,
#>     quote = TRUE)), envir = .GlobalEnv, quote = TRUE), file = "/home/nat/.tmp/RtmpAJb3S4/callr-res-2f64c750f9f6e")
#>   flush(stdout())
#>   flush(stderr())
#>   NULL
#>   invisible()
#> }, error = function(e) {
#>   {
#>     callr_data <- as.environment("tools:callr")$`__callr_data__`
#>     err <- callr_data$err
#>     assign(".Traceback", .traceback(4), envir = callr_data)
#>     dump.frames("__callr_dump__")
#>     assign(".Last.dump", .GlobalEnv$`__callr_dump__`, envir = callr_data)
#>     rm("__callr_dump__", envir = .GlobalEnv)
#>     e$call <- deparse(conditionCall(e), nlines = 6)
#>     e2 <- err$new_error(conditionMessage(e), call. = conditionCall(e))
#>     class(e2) <- c("callr_remote_error", class(e2))
#>     e2$error <- e
#>     calls <- sys.calls()
#>     dcframe <- which(vapply(calls, function(x) length(x) >= 1 && identical(x[[
#>       1]], quote(do.call)), logical(1)))[1]
#>     if (!is.na(dcframe)) e2$`_ignore` <- list(c(1, dcframe + 1L))
#>     e2$`_pid` <- Sys.getpid()
#>     e2$`_timestamp` <- Sys.time()
#>     if (inherits(e, "rlib_error_2_0")) e2$parent <- e$parent
#>     e2 <- err$add_trace_back(e2, embed = FALSE)
#>     saveRDS(list("error", e2), file = paste0(
#>       "/home/nat/.tmp/RtmpAJb3S4/callr-res-2f64c750f9f6e", ".error"))
#>   }
#> }, interrupt = function(e) {
#>   {
#>     callr_data <- as.environment("tools:callr")$`__callr_data__`
#>     err <- callr_data$err
#>     assign(".Traceback", .traceback(4), envir = callr_data)
#>     dump.frames("__callr_dump__")
#>     assign(".Last.dump", .GlobalEnv$`__callr_dump__`, envir = callr_data)
#>     rm("__callr_dump__", envir = .GlobalEnv)
#>     e$call <- deparse(conditionCall(e), nlines = 6)
#>     e2 <- err$new_error(conditionMessage(e), call. = conditionCall(e))
#>     class(e2) <- c("callr_remote_error", class(e2))
#>     e2$error <- e
#>     calls <- sys.calls()
#>     dcframe <- which(vapply(calls, function(x) length(x) >= 1 && identical(x[[
#>       1]], quote(do.call)), logical(1)))[1]
#>     if (!is.na(dcframe)) e2$`_ignore` <- list(c(1, dcframe + 1L))
#>     e2$`_pid` <- Sys.getpid()
#>     e2$`_timestamp` <- Sys.time()
#>     if (inherits(e, "rlib_error_2_0")) e2$parent <- e$parent
#>     e2 <- err$add_trace_back(e2, embed = FALSE)
#>     saveRDS(list("error", e2), file = paste0(
#>       "/home/nat/.tmp/RtmpAJb3S4/callr-res-2f64c750f9f6e", ".error"))
#>   }
#> }, callr_message = function(e) {
#>   try(signalCondition(e))
#> }), error = function(e) {
#>   NULL
#>   try(stop(e))
#> }, interrupt = function(e) {
#>   NULL
#>   e
#> })
#> 	 -> tryCatchList(expr, classes, parentenv, handlers)
#> 	 -> tryCatchOne(tryCatchList(expr, names[-nh], parentenv, handlers[-nh]), names[nh],
#> parentenv, handlers[[nh]])
#> 	 -> doTryCatch(return(expr), name, parentenv, handler)
#> 	 -> tryCatchList(expr, names[-nh], parentenv, handlers[-nh])
#> 	 -> tryCatchOne(expr, names, parentenv, handlers[[1L]])
#> 	 -> doTryCatch(return(expr), name, parentenv, handler)
#> 	 -> withCallingHandlers({
#>   NULL
#>   saveRDS(do.call(do.call, c(readRDS(
#>     "/home/nat/.tmp/RtmpAJb3S4/callr-fun-2f64c25cfb0d5"), list(envir = .GlobalEnv,
#>     quote = TRUE)), envir = .GlobalEnv, quote = TRUE), file = "/home/nat/.tmp/RtmpAJb3S4/callr-res-2f64c750f9f6e")
#>   flush(stdout())
#>   flush(stderr())
#>   NULL
#>   invisible()
#> }, error = function(e) {
#>   {
#>     callr_data <- as.environment("tools:callr")$`__callr_data__`
#>     err <- callr_data$err
#>     assign(".Traceback", .traceback(4), envir = callr_data)
#>     dump.frames("__callr_dump__")
#>     assign(".Last.dump", .GlobalEnv$`__callr_dump__`, envir = callr_data)
#>     rm("__callr_dump__", envir = .GlobalEnv)
#>     e$call <- deparse(conditionCall(e), nlines = 6)
#>     e2 <- err$new_error(conditionMessage(e), call. = conditionCall(e))
#>     class(e2) <- c("callr_remote_error", class(e2))
#>     e2$error <- e
#>     calls <- sys.calls()
#>     dcframe <- which(vapply(calls, function(x) length(x) >= 1 && identical(x[[
#>       1]], quote(do.call)), logical(1)))[1]
#>     if (!is.na(dcframe)) e2$`_ignore` <- list(c(1, dcframe + 1L))
#>     e2$`_pid` <- Sys.getpid()
#>     e2$`_timestamp` <- Sys.time()
#>     if (inherits(e, "rlib_error_2_0")) e2$parent <- e$parent
#>     e2 <- err$add_trace_back(e2, embed = FALSE)
#>     saveRDS(list("error", e2), file = paste0(
#>       "/home/nat/.tmp/RtmpAJb3S4/callr-res-2f64c750f9f6e", ".error"))
#>   }
#> }, interrupt = function(e) {
#>   {
#>     callr_data <- as.environment("tools:callr")$`__callr_data__`
#>     err <- callr_data$err
#>     assign(".Traceback", .traceback(4), envir = callr_data)
#>     dump.frames("__callr_dump__")
#>     assign(".Last.dump", .GlobalEnv$`__callr_dump__`, envir = callr_data)
#>     rm("__callr_dump__", envir = .GlobalEnv)
#>     e$call <- deparse(conditionCall(e), nlines = 6)
#>     e2 <- err$new_error(conditionMessage(e), call. = conditionCall(e))
#>     class(e2) <- c("callr_remote_error", class(e2))
#>     e2$error <- e
#>     calls <- sys.calls()
#>     dcframe <- which(vapply(calls, function(x) length(x) >= 1 && identical(x[[
#>       1]], quote(do.call)), logical(1)))[1]
#>     if (!is.na(dcframe)) e2$`_ignore` <- list(c(1, dcframe + 1L))
#>     e2$`_pid` <- Sys.getpid()
#>     e2$`_timestamp` <- Sys.time()
#>     if (inherits(e, "rlib_error_2_0")) e2$parent <- e$parent
#>     e2 <- err$add_trace_back(e2, embed = FALSE)
#>     saveRDS(list("error", e2), file = paste0(
#>       "/home/nat/.tmp/RtmpAJb3S4/callr-res-2f64c750f9f6e", ".error"))
#>   }
#> }, callr_message = function(e) {
#>   try(signalCondition(e))
#> })
#> 	 -> saveRDS(do.call(do.call, c(readRDS(
#>   "/home/nat/.tmp/RtmpAJb3S4/callr-fun-2f64c25cfb0d5"), list(envir = .GlobalEnv,
#>   quote = TRUE)), envir = .GlobalEnv, quote = TRUE), file = "/home/nat/.tmp/RtmpAJb3S4/callr-res-2f64c750f9f6e")
#> 	 -> do.call(do.call, c(readRDS("/home/nat/.tmp/RtmpAJb3S4/callr-fun-2f64c25cfb0d5"),
#> list(envir = .GlobalEnv, quote = TRUE)), envir = .GlobalEnv, quote = TRUE)
#> 	 -> <function(what, args, quote = FALSE, envir = parent.frame()) {
#>   if (!is.list(args)) stop("second argument must be a list")
#>   if (quote) args <- lapply(args, enquote)
#>   .Internal(do.call(what, args, envir))
#> }>(base::quote(<function(..., crayon_enabled, crayon_colors, pkgdown_internet) {
#>   options(crayon.enabled = crayon_enabled, crayon.colors = crayon_colors,
#>     pkgdown.internet = pkgdown_internet)
#>   pkgdown::build_site(...)
#> }>), base::quote(<list: pkg = <pkgdown>, examples = TRUE, run_dont_run = FALSE,
#>   seed = 1014, lazy = FALSE, ...>), envir = base::quote(<environment>), quote = base::quote(
#>   TRUE))
#> 	 -> <function(..., crayon_enabled, crayon_colors, pkgdown_internet) {
#>   options(crayon.enabled = crayon_enabled, crayon.colors = crayon_colors,
#>     pkgdown.internet = pkgdown_internet)
#>   pkgdown::build_site(...)
#> }>(pkg = base::quote(<pkgdown>), examples = base::quote(TRUE), run_dont_run = base::quote(
#>   FALSE), seed = base::quote(1014), lazy = base::quote(FALSE), override = base::quote(
#>   <list: >), install = base::quote(FALSE), preview = base::quote(FALSE),
#> new_process = base::quote(FALSE), devel = base::quote(FALSE), crayon_enabled = base::quote(
#>   TRUE), crayon_colors = base::quote(256L), pkgdown_internet = base::quote(TRUE))
#> 	 -> pkgdown::build_site(...)
#> 	 -> build_site_local(pkg = pkg, examples = examples, run_dont_run = run_dont_run,
#>   seed = seed, lazy = lazy, override = override, preview = preview, devel = devel)
#> 	 -> build_reference(pkg, lazy = lazy, examples = examples, run_dont_run = run_dont_run,
#>   seed = seed, override = override, preview = FALSE, devel = devel)
#> 	 -> purrr::map(topics, build_reference_topic, pkg = pkg, lazy = lazy, examples_env = examples_env,
#>   run_dont_run = run_dont_run)
#> 	 -> .f(.x[[i]], ...)
#> 	 -> withCallingHandlers(data_reference_topic(topic, pkg, examples_env = examples_env,
#>   run_dont_run = run_dont_run), error = function(err) {
#>   msg <- c(paste0("Failed to parse Rd in ", topic$file_in), i = err$message)
#>   abort(msg, parent = err)
#> })
#> 	 -> data_reference_topic(topic, pkg, examples_env = examples_env, run_dont_run = run_dont_run)
#> 	 -> run_examples(tags$tag_examples[[1]], env = if (is.null(examples_env)
#> ) NULL else new.env(parent = examples_env), topic = tools::file_path_sans_ext(
#>   topic$file_in), run_dont_run = run_dont_run)
#> 	 -> highlight_examples(code, topic, env = env)
#> 	 -> downlit::evaluate_and_highlight(code, fig_save = fig_save_topic, env = child_env(
#>   env), output_handler = evaluate::new_output_handler(value = pkgdown_print))
#> 	 -> evaluate::evaluate(code, child_env(env), new_device = TRUE, output_handler = output_handler)
#> 	 -> evaluate_call(expr, parsed$src[[i]], envir = envir, enclos = enclos, debug = debug,
#> last = i == length(out), use_try = stop_on_error != 2L, keep_warning = keep_warning,
#> keep_message = keep_message, output_handler = output_handler, include_timing = include_timing)
#> 	 -> timing_fn(handle(ev <- withCallingHandlers(withVisible(eval(expr, envir, enclos)),
#> warning = wHandler, error = eHandler, message = mHandler)))
#> 	 -> handle(ev <- withCallingHandlers(withVisible(eval(expr, envir, enclos)),
#> warning = wHandler, error = eHandler, message = mHandler))
#> 	 -> try(f, silent = TRUE)
#> 	 -> tryCatch(expr, error = function(e) {
#>   call <- conditionCall(e)
#>   if (!is.null(call)) {
#>     if (identical(call[[1L]], quote(doTryCatch))) call <- sys.call(-4L)
#>     dcall <- deparse(call, nlines = 1L)
#>     prefix <- paste("Error in", dcall, ": ")
#>     LONG <- 75L
#>     sm <- strsplit(conditionMessage(e), "\n")[[1L]]
#>     w <- 14L + nchar(dcall, type = "w") + nchar(sm[1L], type = "w")
#>     if (is.na(w)) w <- 14L + nchar(dcall, type = "b") + nchar(sm[1L], type = "b")
#>     if (w > LONG) prefix <- paste0(prefix, "\n  ")
#>   } else prefix <- "Error : "
#>   msg <- paste0(prefix, conditionMessage(e), "\n")
#>   .Internal(seterrmessage(msg[1L]))
#>   if (!silent && isTRUE(getOption("show.error.messages"))) {
#>     cat(msg, file = outFile)
#>     .Internal(printDeferredWarnings())
#>   }
#>   invisible(structure(msg, class = "try-error", condition = e))
#> })
#> 	 -> tryCatchList(expr, classes, parentenv, handlers)
#> 	 -> tryCatchOne(expr, names, parentenv, handlers[[1L]])
#> 	 -> doTryCatch(return(expr), name, parentenv, handler)
#> 	 -> withCallingHandlers(withVisible(eval(expr, envir, enclos)), warning = wHandler,
#> error = eHandler, message = mHandler)
#> 	 -> withVisible(eval(expr, envir, enclos))
#> 	 -> eval(expr, envir, enclos)
#> 	 -> eval(expr, envir, enclos)
#> 	 -> humdata %hum>% c(by ~ Spine, do ~ table(.)) %hum<% (do ~ sort(.))
#> 	 -> is.humdrumR((x))
#> 	 -> humdata %hum>% c(by ~ Spine, do ~ table(.))
#> 	
#> Error: In the call `%hum>%`(humdrumR = _), the argument humdrumR must be a humdrumR object.