Summarize reference records in a humdrumR corpus

reference is one of humdrumR's summary functions, used to tabulate the reference records present in a humdrumR corpus. reference takes a humdrumR object and and returns a humReference table. Alternatively, reference can take a character string, which it will check against known reference codes and print a information about matching codes. For instance, reference('OTL') returns a description of the standard humdrum !!!OTL reference record (original title metadata).

Usage

reference(x)

reference('OTL')

reference(humdata)

# S3 method for humReference
[(refTable, i, j, drop = FALSE)

Details

A humReference table has one column for each reference code that appears in a humdrumR corpus. Since reference records can be long (too much to print on one screen), and humdrum files can have multiple of the same type of reference code (for instance multiple composers annotated with "!!!COM"), by default, a humReference only prints the number of each type of reference record to appear in each file. However, if only one type of reference code is present in a humReference table, the complete reference records for that code will be printed for each file. Likewise, if only one file is present in the table, all of that file's complete reference records are printed. Thus, if you want to see actualy reference records, try indexing the humReference table down to one column or row (see below).

A humReference table has one row for each file in the corpus. Rows are labeled with each file's corresponding number (from the humTables File field) and name (the Filename field). In addition, when a humReference object is printed, three different summary totals are printed for each reference code: Any indicates how many files in the corpus have at least one example of each code in them. Sum indicates the total number of each reference code to appear in the corpus, including multiple appearances in one file (like multiple "!!!COM" records). Finally, Unique tabulates the number of unique tokens in each reference code---if your corpus only hase two unique composers (encoded in "!!!COM"), the Unique total will be 2.

Indexing

humReference tables can be indexed much like base R data.frames, with two arguments: i (rows) and j (columns). If i or j are numeric, they select rows or columns respectively, ordinally. If i is a character, it is matched as a regular expression against filenames in the corpus. If j is a character, it is partially-matched against column names.

A drop argument is also available. If TRUE, a plain data.table is returned.