Summarize humdrumR corpora

Summarizes the content of a humdrumR corpus, by calling five different corpus summary functions and printing their results.

Usage

# S3 method for humdrumR
summary(object)

Details

humdrumR includes a number of separate functions for summarizing different aspects of humdrumR data objects:

census()

Tabulates the raw size of the humdrumR corpus.

reference()

Tabulates reference records (metadata) for each file.

spines()

Tabulates the number of spines and spine paths in files in the corpus.

interpretations()

Tabulates the types of exclusive and tandem interpretations in the corpus.

sections()

Tabulates any formal data (*>) in the corpus, including barlines.

Each function takes a humdrumR object and returns a data.table. The summary method for humdrumR objects simply calls all of the above functions and prints a condensed version of each.

See also

Examples

chorales <- readHumdrum(humdrumRroot, "HumdrumData/BachChorales/*.krn")
#> Finding and reading files...
#> 	REpath-pattern '/home/nat/.tmp/Rtmpiec5so/temp_libpath4863748719359/humdrumR/HumdrumData/BachChorales/*.krn' matches 10 text files in 1 directory.
#> Ten files read from disk.
#> Validating ten files...
#> all valid.
#> Parsing ten files...
#> Assembling corpus...
#> Done!
summary(chorales)
#> 		Summary of humdrumR corpus "chorales":
#> Error in checks(humdrumR, xclass("humdrumR")): object 'chorales' not found