humdrumR

humdrumR is a toolkit for the analysis of data encoded in the humdrum syntax. The humdrum syntax is an incredibly flexible, and powerful, scheme for encoding musical data. Tens of thousands of musical scores (and other musical data) have been encoded in the humdrum syntax, many available online through repositories such as KernScores.

Details

humdrumR is intended as a modernized replacement for the original humdrum toolkit, levaraging the power of R to give us enprecedented power to manipulate and analyze humdrum data using concise, expressive syntax.

Package design

The package humdrumR has XXX main components:

To represent humdrum data in R, we have the humdrumR S4 class, and it's core component the humdrum table.

To create humdrumR data, a sophisticated humdrum data parser: readHumdrum. humdrumR data can also be written back to humdrum-syntax text files using writeHumdrum.

To filter and "index" humdrumR data, we have the filterHumdrum function, which can also be called in a variety of short hands using R's standard indexing operators: [] and [[]].

To manipulate and modify humdrumR data, we have the with(in)Humdrum and humApply functions.

A set of "pipe" operators (%hum>%, %hum<%, etc.), so that humdrumR data can be manipulated and filtered in concise, bash-style pipes.

To facilate the development of functions to work with humdrum tokens---which are simple character strings packed with information---, a useful API we call our regular-expression dispatch system.

Several modules for representing and manipulating musical pitch information, including our core tonalInterval class to represent tonal pitch.

A module for representing and manipulating musical rhythm information, with a core rhythmInterval class to represent rhythms.