HumdrumR class

This S4 class is the basic unit of the humdrumR package. Each humdrumR object represents data read from one or more humdrum files. In the documentation we refer to the collection of files within a humdrumRclass object as a "corpus," and each file as a "piece." However, though humdrum data is usually encoded as one "piece" per file, this is not necessarily the case: files might represent movements within a piece, or even just a part of a score. Still, we tend to refer to them as "pieces." In coding examples, we name humdrumR objects "humdata."

Usage

is.humdrumR(x)

# S4 method for humdrumR
$(x, name)

fields(
  humdrumR,
  fieldTypes = c("Data", "Structure", "Interpretation", "Formal", "Reference")
)

# S4 method for humdrumR,vector
$(x, name) <- value

# S4 method for humdrumR,humdrumR
$(x, name) <- value

# S4 method for humdrumR,character,ANY,vector
[(x, i, j) <- value

# S4 method for humdrumR,character,ANY,humdrumR
[(x, i, j) <- value

humdata[[ , 'regex']]

humdata[[ , , ~expression]] or humdata [[ , , 'regex']] or humdata[[z = ~expression]] or humdata[[z = 'regex']]

humdata[[~expression]]

humdata[[ , ~expression]]

Details

The most imporant part of a humdrumR object is the humdrum tables it holds within it. In essence, an humdrumR object is simply a wrapper around these humdrum tables, which helps users to to visualize, index, summarize, and manipulate the table in a variety of ways.

Basic information about the size and shape of humdrumR data can be obtained with calls to nrecords, npieces, length, ncol, etc.. More detailed summary information can be obtained with the humdrumR corpus summary functions. HumdrumR data can also be coerced to more basic R data types using as.matrix, as.data.frame, etc.. A number of helpful functions are also defined to reshape humdrumR data.

The most powerful features of humdrumR are the tools it gives you to

Filter humdrum data, using filterHumdrum and the standard R indexing operators: [] and [[]].

Apply functions and arbitrary commands to humdrum data using the with(in)Humdrum routines, and their associated piping operators.

Slots

Humtable

A list of humdrum tables, each having the same fields but containing data from different types of records (e.g., interpretations, data, barlines, comments).

Files

A list of two elements. The first, "Search", contains a single character representing the pattern used in the call to readHumdrum which created this humdrumR object. The second, "Names", is a vector of strings representing all the files which matched the pattern and were read into the humdrumR object.

Fields

A list containing strings corresponding to the existing fields in the humdrumR object. The fields are divided into five categories: "Data", "Structure", "Interpretation", "Formal", and "Reference"---see the humdrum table documentation.

Active

A quosure expression which extracts data from field(s) in the humdrum table: the "active expression."

LoadTime

A POSIXct value, indicating the time at which readHumdrum was called to create this humdrumR object.

Patterns

A character vector of the original search patterns used to match files in the system.

Active field

The Active slot of a humdrumR object contains an expression which refers to fields in the internal humdrum table. Go to the dedicated active field documentation to learn more about this important slot!