humdrumR class

This S4 class is the basic unit of the humdrumR package. Each humdrumR object represents data read from one or more humdrum files. In the documentation, we refer to these objects interchangeably as "humdrumR corpora", "humdrumR objects," or humdrumR data(sets). In coding examples we name them "humData."

Usage

is.humdrumR(x)

Details

The most important part of a humdrumR object is the humdrum tables it holds within it; In essence, a humdrumR object is simply a wrapper around its humdrum table, which helps users to to visualize, index, summarize, and manipulate the table in a variety of ways.

Basic information about the size and shape of humdrumR objects can be obtained with calls to nrecord, npiece, length, ncol, etc.. More detailed summary information can be obtained with the humdrumR corpus summary functions. humdrumR data can also be coerced to more basic R data types using as.matrix, as.data.frame, etc.. A number of helpful functions are also defined to "reshape" or reorganize the data (e.g., foldHumdrum(), collapseHumdrum()), or extract the data into "normal" R data structures (e.g, as.matrix.humdrumR(), evalActive()).

The most powerful features of humdrumR are the tools it gives you to...

Print a readable view of the data in shorthand/curtailed humdrum syntax.

Filter humdrumR data, using subset.humdrumR() and the standard R indexing operators: [] and [[]].

Apply arbitrary commands to humtable fields using the with(in)Humdrum routines.

Slots

Humtable

A humdrum tables---i.e, a data.table::data.table() with particular fields.

Files

A list of two elements. The first, "Search", contains a single character representing the pattern used in the call to readHumdrum() which created this humdrumR object. The second, "Names," is a vector of strings representing all the files which matched the pattern and were read into the humdrumR object, with names() corresponding to their "subcorpora" labels (Label).

Fields

A list containing strings corresponding to the existing fields in the humdrumR object's humdrum table. The fields are divided into five categories: "Data", "Structure", "Interpretation", "Formal", and "Reference."

Active

A rlang::quosure() expression which serves the default, "active expression" for the dataset.

LoadTime

A POSIXct value, indicating the time at which readHumdrum() was called to create this humdrumR object.

See also