Sum "connected" durations

These functions are used to sum (melodically) adjacent rhythmic duration values which are not associated with new onsets/attacks. ioi() adds the duration of rests to the previous non-rest (onset) duration, to create interonset intervals (IOIs). untie() sums tied durations.

Usage

ioi(
  x,
  onsets = !grepl("r", x) & !is.na(x) & x != ".",
  ...,
  finalOnset = FALSE,
  groupby = list(),
  parseArgs = list(),
  Exclusive = NULL,
  inPlace = TRUE
)

untie(x, open = "[", close = "]", ..., groupby = list(), inPlace = TRUE)

Arguments

x

An input vector which is parsed for duration information using the rhythm parser. max, median, or mode might be reasonable alternatives.

onsets

A logical vector of the same length as x. All durations in x where onsets == FALSE are added to the previous value where onsets == TRUE.

finalOnset

(logical, length == 1) If TRUE, the last IOI is computed between the last onset and the end of the input vector. Otherwise, this last IOI is undefined (NA).

groupby

A list of vectors, of the same length as x, which are used to group x.

parseArgs

A list of arguments to pass to the rhythm parser.

inPlace

(logical, length == 1) This argument only has an effect if the input (the x argument) is character strings, and there is extra, non-pitch information in the input strings "besides" the duration information. If so, and inPlace = TRUE, the output will be placed into an output string beside the original non-duration information. If inPlace = FALSE, only the rhythm output information will be returned.

open

A character string (regular expression) to identify the beginning of ties. (May also be formula: see humWindows.)

close

A character string (regular expression) to identify the end of ties. (May also be formula: see humWindows.)

Details

Both functions return "collapsed" durations are as null data tokens. For example, untie(c('[4a', '4a]', '2g')) returns c('2a', '.', '2g'), with the second (tied) duration null (".").

For interonset intervals, the last duration in a string of durations is undefined---there is a final onset, but no next onset, so there can't really be a "interonset" interval. Thus, by default, ioi() will return NA at the location of the final duration. However, if the finalOnset argument is set to TRUE, the function will act like there is one additional onset after the end of the sequence: the last "IOI" is calculated between the last onset and this fictional "final onset." For example, if we run ioi(c('4.a','8r', '4.a','8r','2a', '2r')) the result is c("2a", ".", "2a", ".", NA, "."), with the last onset (2a) returning NA. However, if we run ioi(c('4.a','8r', '4.a','8r','2a', '2r'), finalOnset = TRUE) the result is c("2a", ".", "2a", ".", "1a", ".")---the last onset's whole duration through the end is returned!

Non-onsets (rests) that occur before the first onset are returned as null.