Humdrum key interpretation

Humdrum key interpretation

Usage

key(x, ..., Key = NULL, parseArgs = list())

Arguments

x

Input data, interpreted as diatonic keys.

Must be an atomic vector.

Key

The key used by the parser, deparser, and transposer.

Defaults to NULL.

Must be a diatonicSet or something coercable to diatonicSet; must be either length 1 or length(x)

parseArgs

An optional list of arguments passed to the key parser.

Defaults to an empty list().

Must be a list of named arguments to the key parser.

Examples

exampleToken <- c('4GG', '4G', '4E', '4F#', '4G', '4D', '4E')
results <- key(exampleToken)
results
#> [1] "*G:"  "*G:"  "*E:"  "*F#:" "*G:"  "*D:"  "*E:" 

exampleHumdrum <- readHumdrum(humdrumRroot, "HumdrumData/BeethovenVariations/B075_00_05_a.krn")
#> Finding and reading files...
#> 	REpath-pattern '/home/nat/.tmp/RtmpTnkUuG/temp_libpath43db7135c35b3/humdrumR/HumdrumData/BeethovenVariations/B075_00_05_a.krn' matches 1 text files in 1 directory.
#> One files read from disk.
#> Validating one files...
#> all valid.
#> Parsing one files...
#> Assembling corpus...
#> Done!
results <- with(exampleHumdrum[[,3:4]], key(Token))
results
#>  [1] NA     "*F:"  "*c:"  "*E:"  "*c:"  "*F:"  "*c:"  "*B:"  "*B-:" "*B:" 
#> [11] "*B-:" "*C:"  "*A:"  "*C:"  "*G:"  "*F:"  NA     "*a:"  "*c:"  "*f:" 
#> [21] "*a:"  "*c:"  "*f:"  "*f:"  "*a:"  "*a:"  "*c:"  "*a:"  "*c:"  "*g:" 
#> [31] "*b-:" "*f:"  "*a:"  "*f:"  "*d:"  "*f:"  "*d:"  "*f:"  "*c:"  "*B-:"
#> [41] "*e:"  "*A:"  "*f:"  NA