Shift data within a vector/matrix/data.frame

The lag and lead functions take input vectors, matrices, or data.frames and shifts their data by n indices. They are similar to the data.table::shift() function, but with a few additional options.

Usage

lag(x, n = 1, fill, wrap, boundaries, ...)

lead(x, n, ...)

Arguments

x

The input argument. Should be a vector (including lists), matrix, or data.frame.

n

The amount to lag/lead the data. If n == 0, x is returned unchanged.

fill

If wrap = FALSE parts of the output are padded with the fill argument. Defaults to NA.

wrap

If wrap = TRUE, data from the end (head or tail) is copied to the other end of the output, "wrapping" the data within the data structure.

boundaries

A vector or list of vectors, all of the same length as x. Each segment of x delineated by the boundaries vector(s) is treated separately.

margin

Arrays and data.frames can be lagged lead in multiple dimensions using the margin argument: margin == 1 shifts across rows while margin == 2 shifts across columns.

Details

A lagged vector has the same values as the original vector, except offset by n indices. lag moves each value to a high index (if n > 0); lead does the opposite, moving each value to a lower index (if n > 0). n can be positive or negative---negative lags are equivalent to leads, and vice versa. Values near the end/beginning are either "wrapped" to the opposite end of the vector, or replaced/padded with the value of the fill argument.

The vector a b c d e f g can be lagged by n==1 is NA a b c d e f. If we set wrap == TRUE, the "g" moved to the beginning of the output: is g a b c d e f.

Boundaries

In many cases we want to perform lagged calculations in a vector, but not across certain boundaries. For example, if your vector includes data from multiple pieces, we wouldn't want to calculate melodic intervals between pieces, only within pieces. The boundaries argument indicates one, or more, grouping vectors, which break the x (input) argument into groups. If more than boundaries vectors are given, a change in any vector indicates a boundary.

Value pairs which cross between groups are treated as if they were at the beginning. Basically, using boundaries should be essentially identical to using tapply(x, boundaries, laggedFunction, ...), except generally faster when the number of groups is large.

The most common use case in humdrum data, is looking at "melodies" within spines. For this, we want boundaries = list(File, Spine, Path ). In fact, humdrumR with(in) calls will automatically feed these three fields as boundaries arguments to certain functions: mint, hint, delta, sigma, lag, or ditto. Do any use of delta in a call to with(in), will automatically calculate the delta in a "melodic" way, within each spine path of each file. However, if you wanted, for instance, to calculate differences across spines (like harmonic intervals) you could manually set boundaries = list(File, Record).

See also