Count or measure metric position

These functions take vectors of rhythmic duration values and compute the metric position of each rhythmic onset. metlev() identifies the metric level of each onset; metcount() counts beats within a measure; metsubpos() measures the distance between an onset and the nearest metric beat. metcount() and metsubpos() parallel the more general count() and subpos() functions.

Usage

metlev(
  dur,
  meter = duple(5),
  pickup = NULL,
  value = TRUE,
  offBeats = TRUE,
  numeric = FALSE,
  deparser = recip,
  groupby = list(),
  ...,
  parseArgs = list(),
  remainderSubdivides = FALSE
)

metcount(
  dur,
  meter = duple(5),
  level = tactus(meter),
  pickup = NULL,
  ...,
  offBeats = TRUE,
  remainderSubdivides = FALSE,
  groupby = list(),
  parseArgs = list()
)

metsubpos(
  dur,
  meter = duple(5),
  pickup = NULL,
  deparser = duration,
  ...,
  remainderSubdivides = TRUE,
  groupby = list(),
  parseArgs = list()
)

Arguments

meter

A meter object, or a character string which can be parsed as a meter.

pickup

NULL, or a logical vector of same length as x.

value

(logical T/F) If TRUE, metlev() deparses and returns the (duration) value of a metric level. If FALSE, the level is output as a number.

offBeats

(logical T/F) If FALSE, any onsets that are offbeat return NA.

groupby

A list of vectors, of the same length as x, which are used to group x into. To function as a by-record timeline, the groupby list music include a named Piece and Record fields. Luckily, these are automatically passed by with(in).humdrumR, so you won't need to worry about it!

parseArgs

A list of arguments to pass to the rhythm parser.

remainderSubdivides

(logical T/F switch) Controls how off-beat onsets are associated with beats.

level

A single character string, recip() format, indicating which level in the meter to count.

Details

Watch out! These met...() functions require meter information and their output is highly dependent on how you interpret meter from scores. For a full discussion of how meter can represented, parsed, and created in humdrumR, see the meter() manual. Effective use of the meter() function is essential to use of metlev(), metcount(), and metsubpos().

Metric levels

metlev() identifies the "highest" (longest) metric level that each onset lands in/on: For example, in 4/4 time:

An onset on the downbeat is at the highest level of all, the whole-note level;

An onset on beat three of the 4/4 measure is on the half-note level;

Onsets on the backbeats (beats two and two) fall on the quarter-note level;

The next level down is the eighth-note level, in between each quarter-note beat;

etc.

The metlev() output expresses beat levels as the duration of the the level, in recip() format by default. So the whole-note level is "1" and the quarter-note level (backbeats) is "4". You can specify a different deparsing function (like duration()) using the deparser argument. (If deparser is NULL, rational() numbers are returned.)

Another option is to express the metric levels simply as natural numbers, which you can achieve with the argument value = FALSE. In this case, the top level of the meter is 1, which each next lower-level incrementing up: i.e., the quarter-note level (of 4/4) would be 3, while the sixteenth-note level would be 5.

(Full) 4/4 meter levels

1	&	2	&	3	&	4	&

Level (recip())	"1"	"8"	"4"	"8"	"2"	"8"	"4"	"8"

Level (value = FALSE)	1	4	3	4	2	4	3	4

3/4 meter levels

1	&	2	&	3	&

Level (recip())	"1"	"8"	"4"	"8"	"4"	"8"

Level (value = FALSE)	1	3	2	3	2	3

6/8 meter levels

1	&	a	2	&	a

Level (recip())	"1"	"8"	"8"	"4."	"8"	"8"

Level (value = FALSE)	1	3	3	2	3	3

Metric counts

The metcount() function counts one beat level in a metric hierarchy, within the next highest level. In a full duple meter, the counts are always simply 1, 2, 1, 2, etc. Meters with a triple level will get 1, 2, 3, etc. Why level you want to count is controlled by the level argument, which can be either a character string in recip() format or a natural number (1 is top level, 2 is next lowest level, etc.).

(Full) 4/4 meter counts:

1	&	2	&	3	&	4	&

"1" (whole)	1	1	1	1	1	1	1	1

"2" (half)	1	1	1	1	2	2	2	2

"4" (quarter)	1	1	2	2	1	1	2	2

"8" (eighth)	1	2	1	2	1	2	1	2

3/4 meter counts:

1	&	2	&	3	&

"2." (dotted-half)	1	1	1	1	1	1

"4" (quarter)	1	1	2	2	3	3

"8" (eighth)	1	2	1	2	1	2

6/8 meter counts:

1	&	a	2	&	a

"2." (dotted-half)	1	1	1	1	1	1

"4." (dotted-quarter)	1	1	1	2	2	2

"8" (eighth)	1	2	3	1	2	3

In the case of 4/4, if you want to count 1, 2, 3, 4, you'll need to make your meter() object not include a half-note level.

4/4 meter with no half-note level:

1	&	2	&	3	&	4	&

"1" (whole)	1	1	1	1	1	1	1	1

"4" (quarter)	1	1	2	2	3	3	4	4

"8" (eighth)	1	2	1	2	1	2	1	2

You can do this with meter('M4/4', fill.levels = 'below').

Metric subpositions

In some cases, onsets may occur which do not land on any beat specified in the meter. This could be very fast beat levels (e.g., 32nd notes), triplets, or other tuplets. In these cases, you might consider adding these levels to the meter(); for example, if you want to have a 32nd-note level in 4/4, you could use meter('M4/4', tick = '32'). For metlev() and metcount(), the offBeats argument can be set to FALSE to cause offbeat onsets to return NA. Another option is to use metsubpos(), which measures how far an onset is from the nearest associated beat in the meter.

By default, off-beat onsets are always associated with the closets previous position in any level in the meter. If the remainderSubdivides argument is TRUE, off-beat onsets are associated with the previous metric level which the subposition makes an even subdivision of.

Pickups

Another option is to pass the pickup argument a logical vector of the same length as the input x. Within each piece/group, any block of TRUE values at the beginning of the pickup vector indicate a pickup. The first index where the pickup logical is FALSE is used as the starting point of the timeline/count; All the earlier (pickup == TRUE) points will be negative numbers, measured backwards from the start index. In humdrumR, and datapoints before the first barline record (=) are labeled Bar == 0 in the Bar field. Thus, a common use for the pickup argument is within(humData, timeline(Token, pickup = Bar < 1), which makes the downbeat of the first complete bar 1 the starting point of the timeline---any notes in pickup bars are negative on the timeline.

See also