Translate between pitch representations.

These functions can be used to extract and "translate," or otherwise modify, data representing pitch information.

Arguments

x

(atomic vector) The x argument can be any (atomic) vector, or a tonalInterval, or NULL.

...

These arguments are passed to the pitch deparser. There are also two hidden (advanced) argumens you can specify: memoize and deparse (see the details below).

generic

(logical, length == 1) If generic = TRUE the "specific" pitch information is discarded. For tonal representations, this means no accidentals/qualities or equivalent information is printed. For atonal representations, the generic pitch is returned. As an alternative, you can specify specific as an alternative (opposite) to the generic argument: generic == !specific.

simple

(logical, length == 1) If simple = TRUE the "complex" pitch information is discarded. This means that octave information is discarded, and the resulting output is confined within the default octave. For absolute pitch representations, this is the octave above middle C. For relative pitch representations, this is the octave above a unison. (The special octave.round argument can be used to modify this behavior).

parseArgs

(list) parseArgs can be a list of arguments that are passed to the pitch parser. As a convenient syntactic sugar, instead of writing parseArgs = list(a = x, b = y, etc.), you can write parse(a = x, b = y, etc.).

transposeArgs

(list) transposeArgs can be a list of arguments that are passed to a special call to transpose. As a convenient syntactic sugar, instead of writing transposeArgs = list(a = x, b = y, etc.), you can write transpose(a = x, b = y, etc.).

inPlace

(logical, length == 1) This argument only has an effect if the input (the x argument) is character strings, and there is extral, non-pitch information in the strings "beside" the pitch information. If so, and inPlace = TRUE, the output will be placed into an output string beside the original non-pitch information. If inPlace = FALSE, only the pitch output information will be returned (details below).

Value

NULL inputs (x argument) return a NULL output. Otherwise, returns a vector of the same length as x. NA values in the input x are propagated to the output.

Details

These pitch functions all work in similar ways, with similar arguments and functionality; To read more details about each function, click on the links in the list below, or type ?func in the R command line: for example, ?kern. The functions are:

Tonal pitch representations

Absolute pitch representations

kern()

pitch()

lilypond()

helmholtz()

tonh() (German-style notation)

Relative pitch representations

interval()

solfa() (relative-do solfege)

solfg() (fixed-do solfege)

degree() (absolute scale degrees)

deg() (melodic scale degrees)

bhatk() (hindustani swara)

Partial pitch representations

step()

accidental()

quality()

octave()

Atonal pitch representations

Musical (non-tonal) pitch representations

semits()

midi()

cents()

pc()

Physical pitch/interval representations

freq()

In-place parsing

In humdrum data, character strings are often encoded with multiple pieces of musical information right besides each other: for example, **kern data might include tokens like "4.ee-[. The humdrumR parser (tonalInterval) will automatically "pull out" pitch information from within strings, if it can find any using the appropriate known regular expressions. For example, pitch('4.ee-[') returns Eb5. However, all the pitch functions (like pitch() and kern()) have an option to keep the "extra" information and return the result "in place"---i.e., embedded right where it was found in the input string. This is controlled with the inPlace argument, which is FALSE by default. So, pitch('4.ee-[', inPlace = TRUE) will return 4.Eb5[---keeping the "4." and the "[". (This obviously only works if the input is a string, not a numeric!) Note that inPlace = TRUE will force functions like semits, which normally return numeric values, to return character strings if their input is a character string.

See also