Translate between pitch representations.

These functions are used to extract and translate between different representations of pitch information. The functions can also do things like transposing and simplifying pitches.

Arguments

x

(atomic vector) The x argument can be any (atomic) vector, or a tonalInterval, or NULL.

...

These arguments are passed to the pitch deparser. There are also two hidden (advanced) arguments you can specify: memoize and deparse (see the details below).

generic

(logical, length == 1) If generic = TRUE the "specific" pitch information (accidentals and qualites) is discarded.

simple

(logical, length == 1) If simple = TRUE the "compound" pitch information (octave/contour) is discarded.

Key

(a diatonicSet or something coercable to diatonicSet, length == 1 | length == length(x)) The input Key used by the parser, deparser, and transposer.

parseArgs

(list) parseArgs can be a list of arguments that are passed to the pitch parser.

transposeArgs

(list) transposeArgs can be a list of arguments that are passed to a special call to transpose().

gamutArgs

(list) gamutArgs can be a list of arguments that are passed to a special call to gamut().

Defaults to TRUE for character-string output (like kern()) and FALSE for numeric output (like semits()).

Must be singleton logical value: an on/off switch.

inPlace

(logical, length == 1) This argument only has an effect if the input (the x argument) is character strings, and there is extra, non-pitch information in the input strings "besides" the pitch information. If so, and inPlace = TRUE, the output will be placed into an output string beside the original non-pitch information. If inPlace = FALSE, only the pitch output information will be returned (details below).

Value

NULL inputs (x argument) return a NULL output. Otherwise, returns a vector/matrix of the same length/dimension as x. NA values in the input x are propagated to the output.

Details

The full list of pitch functions is:

Tonal pitch representations

Absolute pitch representations

kern()

pitch()

lilypond()

helmholtz()

tonh() (German-style notation)

Relative pitch representations

interval()

solfa() (relative-do solfege)

solfg() (French-style fixed-do solfege)

degree() (absolute scale degrees)

deg() (melodic scale degrees)

bhatk() (hindustani swara)

Partial pitch representations

step()

accidental()

quality()

octave()

Atonal pitch representations

Musical pitch representations

semits()

midi()

cents()

pc() (pitch classes)

Physical pitch representations

freq()

These pitch functions all work in similar ways, with similar arguments and functionality. Each function takes an input pitch representation (which can be anything) and outputs its own pitch representation. For example, kern() takes any input representation and outputs **kern (pitch) data. Underneath the hood, the full processing of each function looks like this:

Input representation (e.g., **pitch or **semits) |>

Parsing (done by tonalInterval()) |>

Intermediate (tonalInterval) representation |>

Transformation (e.g., transpose()) |>

Deparsing |>

Output representation (e.g. **kern or **solfa)

To read the details of the parsing step, read this. To read the details of the "deparsing" step, read this. To read more details about each specific function, click on the links in the list above, or type ?func in the R command line: for example, ?kern.

See also