Find and read humdrum files into R

findHumdrum does the work of finding and reading valid text files into R. readHumdrum utilizes findHumdrum to read files, then parses them to create a humdrumR data object.

Usage

knownInterpretations

findHumdrum(
  ...,
  contains = NULL,
  recursive = FALSE,
  allowDuplicates = FALSE,
  verbose = FALSE
)

readHumdrum(
  ...,
  recursive = FALSE,
  contains = NULL,
  allowDuplicates = FALSE,
  verbose = FALSE,
  tandems = "known",
  reference = "all"
)

Format

Arguments

...

character: One or more regex search patterns used to identify files on the local machine. read. For details: see the "Regex file-path matching" section below.

contains

(character) If !is.null(contains), the contains argument is is treated as regular expressions: only files which contain matches to all of these regular expressions are matched.

recursive

(logical, length == 1) If TRUE, the final part of the search pattern (i.e., the file search) is searched for recursively through all sub directories.

allowDuplicates

(logical, length == 1) Indicates what should happen if multiple search patterns match the same files. If allowDuplicates = TRUE, any such files are read multiple times, grouped into their respective corpora by the Label field. If allowDuplicates = FALSE, any redundant files are only read into the corpus of the first pattern they match.

verbose

(logical, length == 1) If TRUE, the names of matching files are printed before parsing begins. This is very useful as a check to make sure you aren't reading the wrong files!

tandems

(character) This argument controls which, if any, tandem interpretations are parsed into their own fields. The default value is "known".

reference

(character) This argument controls which, if any, reference records are parsed into their own fields. The default value is "all".

Details

findHumdrum searches for matching files (details in next section) on your machine and returns a "file info" data.table. The file-info table has one row for each matched file, and six columns:

Filepath (character): The complete file path.

Pattern (character): The search pattern (details below) which matched the file.

Label (character): The "subcorpora" label identified with the search pattern. (Used as the Label field in a humdrum table.)

Files (character): The file contents themselves, in a single string.

FileLines (list of character): The file contents themselves, divided into lines.

File (integer): A integer enumerating the files. (Used as the File field in a humdrum table.) If two or more files in different directories share the same name, a unique name is created for each file by appending the names of the directories they occupy, recursively until the names are unique.

readHumdrum uses findHumdrum to identify humdrum files, then parses each file and assembles them all into a humdrum table which is in turn wrapped in a humdrumR data object (the final output). The output's Filepath, File, and Label fields are taken directly from findHumdrum's "file info" table. If a single humdrum file has multiple pieces in it---meaning that all spine paths close with *-, then open again with **---then they are parsed separately, and labeled separately in the Piece field. If there are no multi-piece files, the Piece and File will be identical.

Validity

findHumdrum and readHumdrum automatically ignore non-text files and files that contain humdrum syntax errors. When calling either command, messages will print in the terminal informing you about invalid files, if any. Validity checking is performed by the separate validateHumdrum() function. If you want to see specifically what errors occurred, call validateHumdrum() directly.

Regex file-path matching

Files on your local machine can be concisely identified using regular-expression pattern matching. To find/read humdrum files, one or more character search patterns must be specified as ... arguments; A search pattern is a file-path like string, with strings separated by zero or more delimiters (/ or \, depending on your operating system). Each delimited part of the string to the left of a delimiter (if any) is treated as a regular expression matched against directories, recursively. The right-most part of the string (or the whole string if there are no delimiters), is matched as a regular expression against files in the matched directories. The left-most directory pattern (if any) can be either relative (to R's working directory) or absolute. (Unix-like systems can use ~ for home. All systems use . for the current working directory and .. for its parent. These special keys are not treated like regular expressions.)

Let's consider some examples:

With no delimiters (relative):

readHumdrum('.*krn$') will match all files in the current working directory which match the .*krn$ regex. (This would be all files ending the krn.)

With one directory pattern (relative):

readHumdrum('[A-G]/.*krn$') will first match any subdirectories of the working directory which match the [A-G] regex. If there are any such directories, it will then search the files in all of these matching directories for matches to .*krn$.

Two directory patterns (absolute):

On Unix-like systems readHumdrum('~/[A-G]/.*krn$') will perform the same search as above, but in the home directory (~) instead of the R working directory.

With two directory patterns (relative):

readHumdrum('^Krn$/[A-G]/.*krn$') will first match the directory Krn in the working directory, search all the subdirectories of Krn for matches to [A-G]. Finally, if there are any matching directories, they will all be searched for files matching .*krn$.

Note that the pattern "^Krn$" is needed to match exactly the Krn directory, and no other directory. If used "Krn" instead, any directory containing the string Krn would be matched; for example, NoKrn would be a match.

If you provide multiple ... arguments, they are pasted together (left-to-right), separated by the system delimiter, to create a single complete regex search pattern. Thus, readHumdrum('^Krn$', '[A-G]', '.*krn$') is identical to readHumdrum('^Krn$/[A-G]/.*krn$').

If you don't specify any ... arguments, findHumdrum (or readHumdrum) will default to "", the empty regex pattern "" which matches any file. Thus, readHumdrum() will read any humdrum files in the working directory.

Multiple patterns

If you want to search for more than one pattern, you can input them as a character vector: For instance, readHumdrum(c("Mozart", "Beethoven")) will search for files (in the working directory) matching the regexes Mozart OR Beethoven. This works for directories too: readHumdrum(c('Mozart/.*krn', 'Beethoven/.*krn')) will search for files matching .*krn in directories matching Mozart or Beethoven. For this sort of call, it can be convenient to break the search across arguments: readHumdrum(c("Mozart", "Beethoven"), ".*krn") will give the same result. When the multiple ... arguments are appended, every possible combination of arguments is used to create search patterns. Thus, a call like readHumdrum(c('[A-G]', '[a-g]), c('Mozart', 'Beethoven'), '.*krn') will search all the following patterns:

[A-G]/Mozart/.*krn

[a-g]/Mozart/.*krn

[A-G]/Beethoven/.*krn

[a-g]/Beethoven/.*krn

When reading humdrum data, we refer to files matched from different regex patterns to be "subcorpora" of the total corpus. The last search above would create a dataset with four subcorpora, which are tracked in the humdrumR object's Label field. By default, subcorpora are simply numbered: "_1", "_2", etc. However, if pattern arguments are named, these names will show up in the Label field instead.

Filter files by content

The contains argument can be a character vector of regular expressions to search in the contents of matched files (as opposed to matching the file name/path). Is contains is not NULL (which is the default), humdrumR searches all files matched by the regex search pattern(s) for any matches to the contains regexes. Files which match all the contains regexes are read, others are discarded.

For example, the command readHumdrum(./.*krn$, contains = c('[Ee]-', '[Aa]-')) will read all krn-extension files in the working directory which contain E-flat and a A-flat tokens.

Tandem Interpretations

All tandem interpretations in a humdrum dataset are summarized in the humdrum table's Tandem field, which is described in detail here. In addition, certain "known" tandem interpretations are parsed into their own fields automatically. For example, *clefG4 and "*clefF2 are parsed into a Clef field, while *k[b-] is parsed into a KeySignature field. The "known" tandem interpretations that humdrumR recognizes are encoded in a built-in table called knownInterpretations, which contains five character columns:

Name: A code name humdrumR uses (e.g., Clef or TimeSignatue).

Exclusive: Zero or more (comma-separated) exclusive interpretations associated with the tandem interpretation.

RE: A regular expression which matches instances of the tandem interpretation.

Pretty: A readable shorthand of the tandem interpretation's pattern.

Type: One of three types: "Tandem", "Exclusive", or "Atomic".

The tandems argument to readHumdrum controls which tandem interpretations are parsed into their own fields. This can be helpful to either save processing time and memory by not parsing interpretations you won't need, or to parse interpretations that humdrumR doesn't recognize. The default value for the tandems argument is "known". If the tandems argument contains "known" all tandem interpretations in the built-in knownInterpretations table are parsed. If you want to limit the number of interpretations that are parsed, you may explicitly indicate known interpretations to parse by giving their Name (matching knownInterprations$Name) in the tandems argument. For instance, if you specify tandems = c('Clef', 'TimeSignature'), only clef (e.g., "*clefG2"), and time signature (e.g., "*M3/4") interpretations will be parsed.

If you want to parse novel interpreations that aren't known to humdrumR, you can alternatively specify an arbitrary regular expressions in the tandem argument. (So long as they don't exactly match a Name from knownInterpretations.) These regular expressions are matched against the Tandem field and used to generate new fields for the humdrum table. If any values in tandems are named, these names will be used for the resulting fields. If no matches to an given interpretation are found, no field is created for that interpretation. If tandems = NULL, then no tandem interpretations are parsed.

Reference Records

By default (reference = "all"), humdrumR reads all reference records in the data. The reference code for each record (e.g, the "OTL", in "!!!OTL: xxx") is used as the name of an associated field. If a reference record has no reference code (i.e., it lacks a colon), the field is called "Unkeyed." If there are more than one reference records with the same reference code, either explicitly numbered (e.g., "!!!COM1:", "!!!COM2:") or not, a single field is created (COM in this case) with the multiple values separated by ";".

Controlling reference parsing

The reference argument can be used to control which reference codes are parsed by readHumdrum, and/or rename their resulting fields. Unless reference == "all", only referene codes specifically indicated in reference are parsed; For example, readHumdrum(_, reference = "OTL") will only parse OTL reference records. By naming the values of reference, the resulting field name can be controled: If we specify reference = c(Title = 'OTL'), the OTL reference records will be placed in a field called Title. If reference = NULL, no reference records are parsed.

In large datasets with many reference records, the reference data can actually make up a large portion of the humdrum table, using up a lot of memory. In these cases, simply not reading reference records (reference = NULL) or selectively reading only the reference records that we are planning to use in our analyses can be quick and easy way to reduce memory consumption.

Spines and Paths

At first glance, humdrum data looks like a simple two-dimensional, tab-delineated table structure. However, the humdrum syntax is bit more complicated than that. Wheres as two-dimensional table has stable "columns", humdrum data has "spines": "Spines" can be a single column, or they may (at any time) split into multiple "spine paths", which can in turn split again, using the "*^" interpretation token. The reverse can happen as well, with two or more paths merging into a single spine, using the "v" token. Thus the humdrum syntax is actually a flexible tree structure; As spines split and merge, the total number of "columns" can change during a piece. A related issue is that a "corpus" (collection) of humdrum files may have varying numbers of spines/columns between pieces. Finally, global comment records are also a special case, as that are always a single value, even though they are interspersed with multi-spine local records.)

To make parsing and manipulating humdrum data data simpler, readHumdrum assumes a slightly restricted version of the humdrum syntax: namely, that spines which appear at the beginning of a file (headed with exclusive interpretations like "**kern") can never merge into each other. Spine merges ("*v") can only happen within spine paths that originally split off the same spine, which means that a humdrum file read into humdrumR must not end with fewer columns than it starts. This extra-strict specification of spine paths in the humdrum syntax is, fortunately, something that has been informally followed in most humdrum data sets.

Our strict spine-path definition makes parsing humdrum data into a uniform structure possible: Within a piece, the spines which appear at the beginning of the piece are the "true" spines throughout the piece, numbered from left to right, starting from 1. For each local token, the value in the Spine field is an integer indicating which of these "true" spines it belongs to---global tokens have a NA value in their Spine field, because they do not belong to any spine. Any spine path splits ("*^") from the main spines form spine paths. Every spine's paths are numbered in the Path field, from right to left, starting from 0. A spine with no splits will have all 0s in its Path field.

Examples

readHumdrum() # loads all valid humdrum files in the current directory.
#> Finding and reading files...
#> 	REpath-pattern '' matches 101 text files in 1 directory.
#> 101 files read from disk.
#> Validating 101 files...
#> 22,798 errors in 101 files...
#> zero valid files.
#> NULL

readHumdrum(".*krn$") # loads all files ending with "krn" in the currect directory
#> Finding and reading files...
#> 	REpath-pattern '.*krn$' matches 0 text files in 1 directory.
#> Zero files read from disk.
#> No files to validate.
#> NULL

readHumdrum("^Composers$/^Be|^Mo/.*/^Joined$/.*krn$") 
#> Finding and reading files...
#> 	REpath-pattern '^Composers$/^Be|^Mo/.*/^Joined$/.*krn$' matches 0 text files in 0 directories.
#> Zero files read from disk.
#> No files to validate.
#> NULL
# Goes inside the directory "Composers".
# Inside "Composers" looks for directories that start with "Be" or "Mo".
# If there are any "Be|Mo" matching directories within "Composers", matches all directories within them.
# Within these directories, looks for directories called "Joined".
# If there are any directories called "Joined", loads all files (if any) that end with "krn".

readHumdrum("^Composers$", "^Be|^Mo", ".*", "^Joined$", ".*krn$")
#> Finding and reading files...
#> 	REpath-pattern '^Composers$/^Be|^Mo/.*/^Joined$/.*krn$' matches 0 text files in 0 directories.
#> Zero files read from disk.
#> No files to validate.
#> NULL
# exactly the same as the previous!

readHumdrum("^Composers$", c(Beethoven = "^Be", Mozart = "^Mo"), ".*", "^Joined$", ".*krn$") 
#> Finding and reading files...
#> 	REpath-pattern Beethoven = '^Composers$/^Be/.*/^Joined$/.*krn$' matches 0 text files in 0 directories.	REpath-pattern Mozart = '^Composers$/^Mo/.*/^Joined$/.*krn$' matches 0 text files in 0 directories.
#> Zero files read from disk.
#> No files to validate.
#> NULL
# exactly the same as the previous, except now the two matching patterns ("^Be", or "^Mo") will be grouped
# in the Label field as "Beethoven" and "Mozart" respectively.