Roman Numeral

Roman numerals can be calculated for diatonicSets (keys) and for tertian sets (chords). The later case is the standard meaning of "roman numeral." However, the former case is used as well, for instance to represent modulation schemes in analyses of classical music. For instance, modulate from I-V, the to vi/V. More importantly, many "roman numerals" in harmonic analyses implicitely combine tertian and diatonic roman numerals: in "applied" roman numerals. Given a roman numeral like "V65/V", the "V65" represents a chord while the "/V" represents a key.