Cumulative sum of numeric vector

Calculate sequential cummulative sum of values in numeric vectors.

Usage

sigma(x, lag, skip = is.na, init, boundaries = list(), ...)

# S3 method for default
sigma(x, lag = 1, skip = is.na, init = 0, boundaries = list(), ...)

# S3 method for matrix
sigma(x, margin = 2L, ...)

Arguments

x

(Any numeric vector.) NULL values are returned NULL.

lag

(Non-zero integer.) Which lag to use. (See Great lags section, below.)

skip

(function.) This must be a function which can be applied to x and returns a logical vector of the same length. And TRUE values are skipped over in the calculations. By default, the skip function is is.na, so NA values in the input (x argument) are skipped. The skipped values are returned as is in the output vector.

init

(Atomic value of same class as x, with length(init) <= lag.) NA values at the beginning (or end of right == TRUE) are filled with these values before summing.

boundaries

(vector of same length as x, or a list of such vectors) Differences are not calculated across groups indicated by the boundaries vector(s).

right

(single logical value) Should the init padding be at the "right" (end of the vector)? By default, right == FALSE so the init padding is at the beginning of the output.

Details

sigma is very similar base-R cumsum(). However, sigma should be favored in humdrumR use because:

It has a boundaries argument, which is automatically used by humdrumR with(in) commands to constrain the differences within files/spines/paths of humdrum data. The boundaries approach (details below) is generally faster than applying the commands within groupby groups.

They (can) automatically skip NA (or other) values.

sigma also has a init argument which can be used to ensure full invertability with delta(). See the "Invertability" section below.

If applied to a matrix, sigma is applied separately to each column, unless margin is set to 1 (rows) or, if you have a higher-dimensional array, a higher value.

Invertability

The sigma and delta functions are inverses of each other, meaning that with the right arguments set, sigma(delta(x)) == x and delta(sigma(x)) == x. In other words, the two functions "reverse" each other. The key is that the init argument needs to be set to 0, and all other arguments (lag, skip, boundaries, etc.) need to match. So actually, sigma(delta(x, init = 0, ...)) == x and delta(sigma(x), init = 0)) == x.

When we take the differences between values (delta(x)), the resulting differences can't tell us fully how to reconstruct the original unless we know where to "start" (a constant offset). For example,

delta(c(5, 7, 5, 6)) == c(NA, 2, -2, 1)

We know our input goes up 2, back down 2, then up 1, but the starting value (the first 5) is lost. If we call sigma on this, we'll get:

sigma(c(NA, 2, -2, 1)) == c(0, 2,0, 1)

We get the right contour, but we're offset by that constant 5.

If we call delta(x, init = 0) the necessary constant (the first value) is kept at the beginning of the vector

delta(c(5, 7, 5, 6), init = 0) == c(5, 2, -2, 1)

so sigma gets what we want, full invertability:

sigma(delta(c(5, 7, 5, 6), init = 0)) == c(5, 7, 5, 6)

Alternatively, we could specify the necessary constant as the init argument of sigma:

sigma(delta(c(5, 7, 5, 6)), init = 5) == c(5, 7, 5, 6)

so the init arguments of the two functions are complementary.

Currently, the right argument of delta has no complement in sigma, so invertability only holds true if right = FALSE (the default).

Greater lags

The behavior of sigma when abs(lag) > 1 is easiest to understand as the inverse of the behavior of delta(abs(lag) > 1), which is more intuitive. (sigma is the inverse of delta(), see the Invertability section above).

Generally, if abs(lag) > 1, x is grouped by its indices modulo lag, and the cumulative sum is calculated separately for each set of modulo indices. For example, consider lag == 2 for the following input:

x	index	index modulo 2

1	1	1

3	2	0

2	3	1

2	4	0

5	2	1

The cumulative sum of the 1 and 0 modulo-index groups are:

Index 1: cumsum(c(1,2,5)) == c(1, 3, 8).

Index 0: cumsum(c(3,2)) == c(3, 5)

Interleaved back into order, the result is c(1,3,3,5,8). This may not be very clear, but sure enough delta(c(1, 3, 3, 5, 8), lag = 2, init = 0) returns the original c(1,3,2,2,5) vector! Again, understanding delta(..., lag = n) is easier than sigma(..., lag = n) (see the Invtertability section below.)

Negative lag

If lag is negative, the output is the same as the equivalent positive lag, except the sign is reversed (output * -1). This behavior is easiest to understand as the inverse of the behavior of delta(lag < 0), which is more intuitive. (sigma is the inverse of delta(), see the Invertability section above).

Boundaries

In many cases we want to perform lagged calculations in a vector, but not across certain boundaries. For example, if your vector includes data from multiple pieces, we wouldn't want to calculate melodic intervals between pieces, only within pieces. The boundaries argument indicates one, or more, grouping vectors, which break the x (input) argument into groups. If more than boundaries vectors are given, a change in any vector indicates a boundary.

Value pairs which cross between groups are treated as if they were at the beginning. Basically, using boundaries should be essentially identical to using tapply(x, boundaries, laggedFunction, ...), except generally faster when the number of groups is large.

The most common use case in humdrum data, is looking at "melodies" within spines. For this, we want boundaries = list(File, Spine, Path ). In fact, humdrumR with(in) calls will automatically feed these three fields as boundaries arguments to certain functions: mint, delta, sigma, lag, or ditto. Do any use of delta in a call to with(in), will automatically calculate the delta in a "melodic" way, within each spine path of each file. However, if you wanted, for instance, to calculate differences across spines (like harmonic intervals) you could manually set boundaries = list(File, Record).

See also