Filter humdrum data

subset.humdrumR is a command used to filter a humdrumR corpus. The standard indexing operators ([] and [[]]) actually work by calling subset under-the-hood; you can read about these indexing options here. However, using subset directly can accomplish much more sophisticated filtering commands than the indexing methods.

Usage

# S3 method for humdrumR
subset(x, ...)

clearFilter(humdrumR)

removeEmptyFiles(humdrumR)

removeEmptySpines(humdrumR)

removeEmptyPaths(humdrumR)

removeEmptyRecords(humdrumR)

removeEmptyStops(humdrumR)

Arguments

x

A humdrumR data object.

...

Arbitrary expressions passed to with(in)---the "within" expression(s) must evaluate to full-length logical values.

humdrumR

A humdrumR data object.

Details

subset.humdrumR is used in a similar manner to withinHumdrum, taking any number of "within expressions" as arguments. In fact, expression arguments are passed directly to an internal call to withinHumdrum, and other control expressions (like by or where) can be used as well. The only requirement is that the expressions/functions fed to subset.humdrumR must be predicate expressions which return a logical (TRUE/FALSE) vector. The returned vector must also be the same length as the input data (the number of rows in the humdrum table). (You can use a fill expression if you want to "expand" shorter outputs for filtering purposes.)

Filter field

When using subset, humdrumR doesn't necessarily delete the data you filter out. Rather, there is a logical field in the humdrum table called Filter. subset.humdrumR updates the humdrum table's Filter field using an logical OR (|) between the existing Filter field and the negation of your predicate: Filter | !Predicate. HumdrumR functions (mostly) ignore all data points where Filter == TRUE, treating them like NULL data. For example, when you print a filtered humdrumR you'll see all the filtered data points turned to null data (.), and any calls to withinHumdrum will ignore the filtered data. This means that you can, recover the filtered data by calling clearFilter on your dataset.

In some cases you might filter out large parts of your data, which will leave a bunch of empty null data points ("."). If you want to remove these filtered data points, you can call removeEmptyFiles, removeEmptySpines, removeEmptyPaths, removeEmptyRecords, or removeEmptyStops. These functions go through each piece/spine/path/record and check if all the data in that region is null or filtered (i.e., Null == TRUE | Filter == TRUE); if so, that data will be removed. You can only remove the data if all of it is null (within a region) because otherwise the humdrum syntax is broken.

By default, subset.humdrumR automatically calls removeEmptyFiles at the end. However, you can stop this by specifying removeEmptyFiles = FALSE.

Renumbering

If filtered files are removed from a corpus (using removeEmptyFiles or removeEmptySpines, in combination with subset) the File and/or Spine fields are renumbered to represented the remaining regions, starting from 1. For example, if you have a corpus of 10 files and remove the first file (File == 1), the remaining files are renumbered from 2:10 to 1:9. Spine renumbering works the same, except it is done independently within each file (columns are also renumbered accordingly).

See also