Find common denominator of beats

In humdrumR, we define a tatum as the greatest common denominator of a set of durations. In other words, given a set of durations, the largest duration that divides all the given beats is the tatum---a common unit which can measure any of the other durations

Usage

tatum(x, ...)

# S3 method for meter
tatum(x, deparser = recip)

# S3 method for character
tatum(x, deparser = recip)

# S3 method for numeric
tatum(x, deparser = duration)

# S3 method for rational
tatum(x)

# S3 method for `NULL`
tatum(x)

Arguments

x

The input to compute the tatum of.

Must be a meter() object, a singleton character value, or vector of either character, numeric, or rational() values.

For character input, valuest that match the regular expression "^\*?M" are parsed as a time signature using meter(), while other strings are parsed as durations using rhythmInterval(). numeric input is also parsed using rhythmInterval(); parse failures result in errors.

deparser

What output format is desired?

For character or meter input, the default is recip(); for numeric input, the default is duration().

Must be a rhythm function or NULL.

Details

tatum() is a generic function; it can read any input which can be parsed by the rhythm parser. If can also take a meter() object or character string of the form "MX/Y".

The tatum of a meter() is the tatum of all that meters metric levels. If meters and durations are provided---like tatum(c('M4/4', '6')---, the tatum of all the meters' levels and all the durations is computed.

The deparser argument is a rhythm function which controls the output format. If deparser is NULL, the tatum is returned as a rational() value.

See also

Examples

tatum(c("4", "8", "12"))
#> **recip
#> [1] 24

tatum(c("M4/4"))
#> [1] "16"
#> attr(,"Exclusive")
#> [1] "recip"
#> attr(,"deparseArgs")
#> named list()
#> attr(,"factorizer")
#> function (token) 
#> {
#>     deparseArgs <- attr(token, "deparseArgs")
#>     levels <- do.call(makeRamut, c(list(reference = token, deparseArgs = deparseArgs, 
#>         deparser = attr(token, "deparser"))))
#>     factor(token, levels = levels)
#> }
#> <bytecode: 0x55609f2a65a0>
#> <environment: namespace:humdrumR>
#> attr(,"deparser")
#> function (x, sep = "%", ...) 
#> {
#>     num <- as.numeric(x@Numerator)
#>     den <- as.numeric(x@Denominator)
#>     SIGN <- c("-", "", "")[2 + sign(num)]
#>     num <- abs(num)
#>     den[num == 0L] <- 1
#>     while (any(doubles <- den <= 1L & den >= 0.25 & num > 2L & 
#>         num < 16L & num%%2 == 0 & num%%5 != 0)) {
#>         den <- .ifelse(doubles, den/2, den)
#>         num <- .ifelse(doubles, num/2, num)
#>     }
#>     dots <- log(num + 1L, base = 2L)
#>     dots <- dots - 1L
#>     dots[!(dots %in% 1:100 & (log(den, 2)%%1) == 0) | (num/den) > 
#>         16] <- 0L
#>     num[dots != 0L] <- 1L
#>     den[dots != 0L] <- den[dots != 0L]/(2^dots[dots != 0L])
#>     den[(num/den) %in% c(2, 4, 8)] <- strrep("0", log(num/den, 
#>         2)[(num/den) %in% c(2, 4, 8)])
#>     num[den %in% c("0", "00", "000")] <- "1"
#>     output <- .ifelse(num == 1L, den, .paste(den, sep, num))
#>     .paste(SIGN, output, strrep(".", dots))
#> }
#> <bytecode: 0x55609f2053e0>
#> <environment: namespace:humdrumR>

tatum("M4/4", '6')
#> Error in deparser(result): could not find function "deparser"