with(in)Humdrum

Apply arbitrary expressions to fields within humdrumR data.

Usage

# S3 method for humdrumR
with(data, ..., dataTypes = "D", drop = TRUE, variables = list())

# S3 method for humdrumR
within(data, ..., dataTypes = "D", variables = list())

Arguments

...

Any number of do expressions or evaluation modifying expressions.

dataTypes

A string or vector of characters drawn from c("D", "d", "I", "L", "M","G"). These characters correspond to types of humdrum records: Data, null data, Interpretations, Measures, Local comments, and Global comments respectively. The expression is only evaluated on data drawn from the specified record types (defaults to "D").

drop

This argument is conceptually similar to the drop argument in R matrices. If drop = TRUE, the output of with.humdrumR is simplified as much as possible (trying to return the "raw" vector, list, table, etc. within it). If drop = FALSE, the result is always a data.table. The default value (drop = TRUE) is usually what we want because it is more intuitive, but in more complex code, it can be helpful to set drop = FALSE so that the output is consistent.

humdrumR

A [humdrumR][humdrumRclass] data object.

Value

From within.humdrumR a new humdrumR data object. From with.humdrumR, whatever value is returned by the expression or, if drop = TRUE, a data.table.

Overview

These functions are the primary means of working with humdrumR data. They are analogous to the base functions with and within methods for data.frames. Specifically they allow you to evaluate arbitrary expressions involving fields in a humdrumR data object. They also includes a number of special evaluation options:

Evaluate an expression in a subset of the data.

Evaluate the same expression separately in different subsets of the data.

Evaluate an expression across windows in the data (e.g., ngrams, rolling windows).

Evaluate an expression which produces a plot, with particular plotting parameters set using graphics::par().

The difference between with.humdrumR and within.humdrumR is analogous to the difference between base::with() and base::within(). with.humdrumR evaluates your expression(s) and then simply returns the result of the evaluation. within.humdrumR evaluates your expression(s) and then inserts the results back into the humdrumR object (if possible), generating new fields called ResultX (see details).

Expression evaluation

An "expression" is a legal bit of R code, like 2 + 2 or x - mean(x). Each call to with/within.humdrumR must have at least one expression to evaluate, or "do" which we call "do" expressions. These expressions are evaluated (executed) within the humdrumR object's humdrum table. which means the expression can, refer to any field in the humdrumR object (Record, Token, File, etc.). Since all the fields in a data.table are all vectors of the same length, expressions should usually be vectorized.

A number of special syntactic sugars can be used in the "do" expressions. The most basic, is that a . anywhere in the expression, will be interpreted as the humdrumR object's current active expression. More syntactic sugars are described below.

If multiple do expressions are provided, each expression is evaluated in order, and can refer to results of the previous do expression as . (variables assigned in previous expressions can also be used.) Other syntactic sugars are described in sections below.

humdata <- readHumdrum(humdrumRroot, 'HumdrumData/BachChorales/.*krn') # read some data

within(humdata, pitch(Token))
within(humdata, pitch(.)) # Same as previous (unless `Active` field has been changed))

with(humdata, table(solfa(Token, Key = Key)))
# Assumes that the Key field was parsed during the call to `[readHumdrum][readHumdrum]`

within(humdata, semits(Token) - mean(semits(Token)))

Unnamed arguments (or formulae without a left-hand side) are interpreted as do expressions. Do expressions can also be explicitly labeled by naming an argument "do" or with do on the left side of a formula:

with(chorales, table(Token))
with(chorales, do = table(Token))
with(chorales, do ~ table(Token))
# These all return the same result

Special Do

A few special versions of the do expression can be used.

doplot:

The expression is evaluated, but the original humdrumR input if returned unchanged. This can be used for achieving a side effect (like making a plot) without saving the result.

dofill:

The result is evaluated, and the result is recycled to the length of the input, making sure the result is the same length as the input. Basically, do fill is the equivalent of rep(do ~ f(x), length.out = length(x)).

Special Evaluation Keywords

with.humdrumR and within.humdrumR can be provided with additional keyword expressions which modify how the main "do" expressions are evaluated. The complete list of options are:

by (group by)

where (apply to subset)

windows

ngrams

pre and post

These special expressions can be specified either as named arguments (with(data, do(x), by = group)) or as formula, with the keyword on the left side (with(data, do(x), by ~ group). Any unnamed argument (or formula without a left-side) is interpreted as a do expression, not a special expression.

A by expression is used to break the data into groups, with the do expression(s) evaluated separately in each group (see "Group by", below).

A where expression indicates a subset of the data in which to evaluate the do expression (see "Partitioning", below).

ngrams A positive number n. The expression is evaluated across overlapping length-n windows. +. pre An expression to evaluate once before evaluating the do expression(s). Useful, for instance, for taking logs or opening a graphing window. The pre expression is evaluated in the global environment. +. post An expression evaluate once after evaluating the do expression(s). Always evaluated in the global environment.

Group by

A by expression is used to break the data into subsets, with the do expression(s) evaluated separately within each subset. This works the similarly to the by argument in [data.table][data.table]s, the INDEX argument of [base][tapply], or the INDICES argument of [base][by]. Each by expression must evaluate, within the humdrumR data object, to a vector (or a list of vectors of equal length) of categories to group the data by.

Most commonly, the by expression(s) are simply field(s) in the data: for instance,

with(humdata,
     do = table(Token),
     by = File)

will apply the function [base][table] to the Token field separately for each file in the humdrumR data. However, we can also use more complex expressions like

with(humdata,
     do = table(Token), 
     by = Spine > 3 | Record \%\% 2 == 0)

which will evaluate the do expression in two groups, one where either the spine number is three or less or the record number is even, and another group where the opposite is true.

If the by expression evaluates to a list of grouping vectors, the do expressions are evaluated across every combination of categories in all the vectors. Thus,

with(humdata, 
     do = table(Token),
     by = list(File, Spine))

will apply table to Token across each spine in each file.

As some syntactic sugar, if the by expression is input as a formulae, lists of grouping expressions can be created by separating each expression by a ~. Thus, by = list(File ~ Spine) can be written by = File ~ Spine or by ~ File ~ Spine.

Apply where TRUE

A where expression is used to identify a subset of the data and evaluate the do expression(s) only in that subset. where expressions must evaluate, within the humdrumR data object, to a single logical vector. The do expression(s) are only evaluated where this logical vector is TRUE. kept unchanged.

Advanced partitioning

If multiple by or where expressions, or combinations of the two, are specified, each is evaluated recursively, in order from left to right. If where is specified after by, the where expression is evaluated within each by group If by is specified after where, the grouping by expression is evaluated only where where == TRUE. Thus, if you specify

within(humdata,
         do = sd(Semits),
         by = File, 
         where = Semits > mean(Semits))
```

the standard deviation of the `semits` field will be calculated in each file,
but only where the `semits` field is greater than the mean `semits` value
*within that file*. Contrast this with this call:

```
within(humdata,
         do = sd(Semits)
         where = Semits > mean(Semits), 
         by = File) 
```

wherein the standard deviation of `semits` is, again, calculated for each file,
but this time wherever the `semits` field is greater than the mean value *across all the data*.

Plotting

The doplot keyword behaves exactly like the do keyword, except that the result of the evaluation is ignored. This is useful for plotting as well as other side-effects (like writing to a file). If doplot is used with with.humdrumR, the function simply returns NULL (after executing the doplot expression. If doplot is used with within.humdrumR, the function simply returns the unaltered humdrumR argument.

within.humdrumR also allows you to specify plotting options inline, without having to make a separate call to par(). Any par() argument can be specified by providing a named list to the graphics keyword. For example, we can set the plot margins with the mar argument:

within(data, 
       doplot = plot(sort(table(Token))), 
       graphics = list(mar = c(4, 4, 4, 4)))

The best part is within.humdrumR will reset par to it's previous state after its done.

You can also use the syntactic sugar, graphics(parargs = ...):

within(data,
       doplot = plot(sort(nchar(Token))),
       graphics(mar = c(4, 4, 4, 4)))

Tandem interpretations

The function [readHumdrum][readHumdrum] automatically parses tandem interpretations (that it recognizes) into their own fields in the resulting [humdrumR][humdrumRclass] data. For instance, data with a '*clefF4' will show up as a Clef field. However, users might read humdrum data with their own custom tandem interpretations that are not built into humdrumR. humdrumR includes the function [getTandem][getTandem] to help us extract arbitrary tandem intrpretation data. Luckily, within.humdrumR knows some syntactic sugar which makes it easy to do this anywhere in our expressions, simply by putting a named object beginning with the symbol *. Of course, R doesn't normally allow names to begin with symbols like *, but you can force it by placing grave symbols around the name *name. If you do this in a within.humdrumR expression, within.humdrumR will treat this name as a regular expression and substitute a call getTandem(Tandem, 'regular expression') in the expression. This means you can could do something like within(humdata, do ~ myFunction(Token, *mytandempattern)) and myFunction will be called with the first argument being the Token field, and the second argument being tandem interpretations which match 'mytandempattern' (extracted from the Tandem field).

Splatting

("Splatting" refers to feeding a function a list/vector of arguments.) Sometimes we want to divide our data into pieces (a l\'a partition option), but rather than applying the same expression to each piece, we want to feed the separate pieces as separate arguments to the same function. In with(in).humdrumR you can use some syntactic sugar to do just this. We can index any field in our call with a splat argument, which must be a Field %in% x. For example,

within(humdata, list(Token[splat = Spine %in% 1:2]))

In this call, the Token field will be divided into two groups, one where Spine == 1 and the other where Spine == 2; the first group (Spine == 1) will be used as the first argument to list, and the second group (Spine == 2) as the second argument. Thus, wihtin.humdrumR translates the previous expression to this:

within(humdata,
       list(Token[Spine == 1], Token[Spine == 2]))

Argument interpolation

The variables argument is a list of named arguments which are [humdrumR:interpolateArguments][interpolated] into the do expressions. This is useful if you've already created a list of formulas that you like, but would like to make small changes to a function call within the do expressions, without starting from scratch. Examples:

mycommand <- c(do ~ mean(., na.rm = x), by ~ Spine ~ File)
within(humdata,
              mycommand,
              variables(x = TRUE))

N grams

Results

The difference between with.humdrumR and within.humdrumR is in what they do with the results of the evaluated do expression(s).

With

For calls to with.humdrumR, the result is simply returned as is. This is what you want when you want to get out of thue humdrumR object and drop back into "normal" R, often in the last stages of an analysis. However, you may optionally specify drop = FALSE, in which case the result is returned in a data.table().

Within

For calls to within.humdrumR, the result of each do expression is inserted back into the [humtable][humdrum table]. Usually, do expressions should evaluate to atomic output which is the same length as the input: in other words, every data point in the input field(s) correspond to a single data point in the new field. However, if the results are shorter than input fields humtable, they are padded with null tokens to match the full (original) length.

Naming results

If you don't explicitely name the results, they are put into new field(s) labeled Result1, ResultX, ..., ResultN. You can name the new fields in one of two ways:

normal R assignment, using <- or -> anywhere in a do expression.

For example, within(data, Semits <- semits(Token)).

end the expression with a named list.

For example, within(data, list(Semits = semits(Token)))

Examples

humdata <- readHumdrum('directorywithdata/*.krn')
#> Finding and reading files...
#> 	REpath-pattern 'directorywithdata/*.krn' matches 0 text files in 0 directories.
#> Zero files read from disk.
#> No files to validate.

within(humdata, nchar(.)) # counts characters in each data token.
#> Error in UseMethod("within"): no applicable method for 'within' applied to an object of class "NULL"
within(humdata, table(.), by ~ Spine) # Tabulates data tokens in each Spine.
#> Error in UseMethod("within"): no applicable method for 'within' applied to an object of class "NULL"