with(in)Humdrum

Evaluate arbitrary expressions using the fields within humdrumR data, while employing split/apply/combine, windowing, and other analysis options.

Usage

# S3 method for humdrumR
with(
  data,
  ...,
  dataTypes = "D",
  expandPaths = FALSE,
  drop = TRUE,
  variables = list()
)

# S3 method for humdrumR
within(data, ..., dataTypes = "D", expandPaths = FALSE, variables = list())

Arguments

data

HumdrumR data.

Must be a humdrumR data object.

...

Any number of expressions to evaluate.

Unnamed expressions are interpreted as the "main" within-expressions. Possible evaluation control arguments include by, subset, and windows. Other evaluation options can be achieved with recycle or side arguments.

dataTypes

Which types of humdrum records to include.

Defaults to "D".

Must be a single character string. Legal values are 'G', 'L', 'I', 'M', 'D', 'd' or any combination of these (e.g., "LIM"). (See the humdrum table documentation Fields section for explanation.)

drop

Whether to return a simplified data structure.

Defaults to TRUE.

Must be a singleton logical value: an on/off switch.

This argument is conceptually similar to the drop argument in R matrices. If drop = TRUE, the output of with.humdrumR is simplified as much as possible (trying to return the "raw" vector, list, table, etc. within it). If drop = FALSE, the result is always a data.table. The default value (drop = TRUE) is usually what we want because it is more intuitive, but in more complex code, it can be helpful to set drop = FALSE so that the output is consistent.

variables

A named list of values, which are interpolated into the within-expression(s) wherever a variable name matches a named from the list.

Defaults to list().

Must be list.

Value

From within.humdrumR a new humdrumR data object. From with.humdrumR, whatever value is returned by the expression or, if drop = TRUE, a data.table.

Overview

These functions are the primary means of working with humdrumR data. They are analogous to the base-R with and within methods for data.frames. Specifically they allow you to evaluate arbitrary expressions using the fields of a humdrumR data object. This means that you can drop "inside" your humdrumR data object and run whatever commands you want using the fields of the humdrum data table, while keeping the data safely "encapsulated" within the humdrum object---at least, until you want to pull it out.

The difference between with.humdrumR and within.humdrumR is analogous to the difference between base::with() and base::within(). with.humdrumR evaluates your expression(s) and then simply returns the result of the evaluation. within.humdrumR evaluates your expression(s) and then inserts the results back into the humdrumR object, generating new fields called ResultX (see details).

In addition, with and within offer a number of powerful options that make working with humdrum data easier: evaluation control arguments can be used to control how your expressions are evaluated. You can evaluate expressions...

In a subset of the data using subset...

either ignoring the rest of the data or evaluating a different expression in the other part.

Separately in different subsets of the data, which are then recombined (split-apply-combine) using by.

Across windows in the data (e.g., ngrams, rolling windows).

Which produces a plot, with particular plotting parameters, and/or without returning anything using sidefx.

"Fill" short results to match the original field size using fill.

Only in certain record types (defaulting only data records) using dataTypes.

These arguments are specified as named arguments to with/within calls. Even though they aren't formal arguments, they are partially matched, so if write grou instead of groupby, you won't get an error! In some cases, you can specify more than one of the same type of control argument (details below).

Expression evaluation

An "expression" is a legal bit of R code, like 2 + 2 or x - mean(x). Each call to with/within.humdrumR must have at least one expression to evaluate. We will refer to these as "within-expressions." These expressions are passed to with/within.humdrumR as unnamed arguments: for example, with(humData, myExpressionHere).

Within expressions are evaluated within the humdrumR object's humdrum table, which means the expression can refer fields in the humdrumR object by name (Record, Token, File, etc.) just like any other variables. Since all the fields in a humdrum object are vectors of the same length, within expressions are easily (and generally should be) vectorized. Note that the within-expression value is only evaluated over data-points/records that match the type indicated in the dataTypes argument. By default, only non-null data tokens ("D") are used.

If multiple within-expressions are provided, each expression is evaluated in order (left to right). Each expression can refer to the results of the last expression (as .), or to variables defined in previous expressions.

A number of special syntactic sugars can be used in within expressions.

The . placeholder.

Side effects

Recycled ("filled") results

Lagged vectors

etc.

Each of these is explained below.

The . placeholder

The . variable can be used as a special placeholder in within expressions. In the first within expression, . is interpreted as the humdrumR object's current active expression. If multiple within expressions are given, beyond the first expression, . refers to result of the previous expression. For example, if Token is the active expression, then:

with(humData, nchar(.), mean(.), .^2)

would return the same result as:

with(humData, mean(nchar(Token))^2)

Side effects:

In some cases, you want to evaluate a within-expression for its "side effect"; This means that the expression does something you want (the "side effect") but doesn't actually evaluate to (return) a result that you want. The most common "side effect" is creating a plot. Other examples might be printing text to the console using base::cat() or base::print(), or writing to a file.

Side effects can be achieved by naming your expression sidefx or fx---as usual, these arguments can be partially matched, so side also works, and is commonly used. Side-effect expressions are executed, but their result (if any) is ignored. This means that if you call something like newData <- within(humData, side = plot(x)), the plot is made but the result (newData) is identical to humData.

Side-effects can also be used in combination with other within expressions. Their result is ignored, and not fed to the next expression as .. For example the command

with(humData, nchar(Token), side = hist(.), mean(.))

creates a histogram of nchar(Token) and also returns the mean of nchar(Token). (Note that variables explicitly assigned in a side call are visible in later calls, which is confusing, so don't do it!)

Recycling ("filling") results:

The result of your within expression may be shorter than the input vectors (humtable fields). However, in some calls to within.humdrumR in particular, you'd like to return a single number and recycle it to "fill" the original data field. In other words, you'd like the output (result) of your expression to be repeated until it matches the length of the input field(s). You could do this manually be using the base::rep() function, but with/within.humdrumR provide a syntactic sugar for this: You can name your expression recycle or fill (partially matched), which will cause the result to be recycled. All this does is take yourExpression(field, ...) and wrap it in rep(yourExpression(field, ...), length.out = length(field)).

Lagged vectors

We very often want to work with "lagged" vectors of data. For example, we want to look at the relationship between a vector and the previous values of the same vector---e.g., the vector offset or "lagged" by one index. The humdrumR lag() function is useful for this, as it gives us several options for lagging vectors, always keeping them the same length so vectorization is never hindered. with and within.humdrumR give us a very convenient short cut to using lag. In a within-expression, any vector can be indexed with an integer argument named lag (case insensitive), causing it to be lagged by that integer. (A vector indexed with lag = 0 returns the unchanged vector.) For example, the following two calls are the same:

with(humData, Token[lag = 1])
with(humData, lag(Token, 1))

If the lag index has multiple values and the indexed object appears within a higher function call, each lag is inserted as a separate argument to that call. Thus, these two calls are also the same:

with(humData, table(Token[lag = 1:2])

with(humData, table(lag(Token, 1), lag(Token, 2))

lag() is a function with a groupby argument, which with/within.humdrumR will automatically feed the fields list(File, Spine, Path). This is the default "melodic" behavior in most music. If you'd like to turn this off, you need to override it by adding your own groupby argument to the lagged index, like Token[lag = 1, groupby = list(...)].

Using lagged vectors, since they are vectorized, is the fastest (computationally) and easiest way of working with n-grams. For example, if you want to create character-string 5-grams of your data, you could call:

with(humData, paste(Token[lag = 0:5], sep = '-'))

Note that, since with/within.humdrumR passes groupby = list(File, Spine, Path) to lag(), these are true "melodic" n-grams, only created within spine-paths within each file.

Results

The difference between with.humdrumR and within.humdrumR is in what they do with the results of the evaluated within-expression(s).

Within

For calls to within.humdrumR, the result of the final within-expression is inserted back into the [humtable][humdrum table]. Most results should be atomic vectors, though lists can also be placed as fields. If results are shorter than the input fields of the humtable they are padded with null values to fill the same length as the input. Non-vector results (objects) and tables are put into lists, and treated like any other list. In the result, data/record-types that are not indicated by the dataTypes argument are also returned padded as null values.

If you don't explicitly name the result of your within-expression, it is put into a new field labeled ResultX, where X is the lowest number of "ResultX" fields that is not already taken. You can explicitly name the result field with a top-level assignment in the expression. For example, within(data, Semits <- semits(Token)) will put the result of semits(Token) into a new field called Semits. In fact, if you assign the results of multiple within-expressions, each result that is the same length as the last result will be put into a new field. Thus, you can create multiple new fields in a single call: for example,

within(humData, 
       Semits <- semits(Token),
       Recip <- recip(Token))

creates two new fields, Semits and Recip. (As explained above, the second within-expression could also refer to the Semits variable created in the previous expression.)

With

For calls to with.humdrumR, the result is simply returned as is. This is what you want when you want to get out of the humdrumR object and drop back into "normal" R, often in the last stages of an analysis.

with.humdrumR has a drop argument, which defaults to TRUE. If drop = FALSE, the result is returned in a data.table(). The column names of this results table are generated as described in the previous section: i.e., defaulting to ResultX but allowing explicit naming with through one or more explicit assignments in the within-expressions. In addition, if subset or groupby arguments are used, columns are included which indicate the value of the evaluated subset/groupby factor for each row.

Partitioning data

groupby (e.g., by) and subset expression control arguments all you to evaluate your within-expressions within specific subsets of the data. A subset argument can be used to evaluate your within-expression only within a subset of the data. A groupby argument breaks the data into groups, evaluating the within-expression(s) separately within each group. The results of the grouped evaluations are then returned in a list (with) or recombined into the original data within---this is a form of the "split-apply-combine" routine that is key to R data analysis.

subset and groupby arguments are themselves arbitrary expressions which are evaluated within the [humdrum table][humTable], so they can (and usually do) refer to fields in the table. Any with/within.humdrumR call can include zero, one, or more subset and/or groupby arguments, including combinations of both. If more than one subset/groupy by argument is included, they are evaluated in order (left to right), recursively: each one evaluated within the partition(s) established by the previous expression. The normal within expression(s) are then, all evaluated within the partition(s) established by the last subset/groupby argument. The "Advanced" partitioning section below explores this in more detail.

Apply in subset:

A subset argument is an arbitrary expression which identifies a subset of the humdrum data. subset expressions must evaluate to a single logical vector, The subset result, if short, will be automatically recycled to the full length of the [humdrum table][humTable]. The within expression(s) are only evaluated where the subset argument(s) return TRUE.

In a call to with, only the result evaluated where subset == TRUE is returned. However, in a call to within, we must decide what to do with rest of the data: the complement of the subset. By default, within pads the returned values with null data where ever subset == FALSE (in the complement). So if you, for example, run the command within(humData, kern(Token), subset = Spine == 1) the new field created by within will be filled with kern data where Spine == 1, but the remaining spines (if any) will all by null.

If you want to explicitly control what is put into the complement part of a new field, you can specifying alternate within-expression(s) to evaluate where subset == FALSE. These must be named complement, or the aliases rest or otherwise (these are all [partially matched][partialMatching]). A complement expression can only be specified in combination with a subset argument, and must be in addition to a normal within-expression. The idea is that you evaluate the "normal" within-expression where subset == TRUE, or else you specify evaluate the complement expression. The results of Complement expressions are always recycled to fill the whole complement (see the "recycle")

A common use case for a complement expression is to use the within expression to change the data in one spine but return the data unchanged in other spines. For example, we could specify within(humData, kern(Token), subset = Spine == 1, complement = Token). Spine 1 will (as before) have kern applied to it. However, instead of return a new field with null values in the other spine(s), this call will return the original (unaltered) values from the Token field in the other spines.

Group by:

A groupby expression (use by for short) partitions your data exhaustively into (possibly non-contiguous) groups, and evaluates your within-expression(s) separately within each group. This works the similarly to the by argument in [data.table][data.table]s, the INDEX argument of [base][tapply], or the INDICES argument of [base][by]. Each groupby expression must evaluate, within the humdrumR data object, to a vector (or a list of vectors of equal length) to group the data by. Each unique combination of values across these vectors becomes one group.

Most commonly, the groupby expression(s) are simply field(s) in the data: for instance,

with(humdata,
     table(Token),
     by = File)

will apply the function [base][table] to the Token field separately for each file in the humdrumR data. However, we can also use more complex expressions like

with(humdata,
     table(Token), 
     by = Spine > 3 | Record \%\% 2 == 0)

which will evaluate the do expression in two groups, one where either the spine number is three or less or the record number is even, and another group where the opposite is true.

If the groupby expression evaluates to a list of grouping vectors, the within expressions are evaluated across every combination of categories across all the vectors. Thus,

with(humdata, 
     table(Token),
     by = list(File, Spine))

will apply table to Token across each spine in each file.

Advanced partitioning:

If multiple groupby or subset expressions, or combinations of the two, are specified, each is evaluated recursively, in order from left to right. If subset is specified after groupby, the subset expression is evaluated within each groupby group If groupby is specified after subset, the grouping by expression is evaluated only where subset == TRUE. Thus, if you specify

within(humdata,
         sd(Semits),
         by = File, 
         subset = Semits > mean(Semits))
```

the standard deviation of the `semits` field will be calculated in each file,
but only where the `semits` field is greater than the mean `semits` value
*within that file*. Contrast this with this call:

```
within(humdata,
         sd(Semits)
         subset = Semits > mean(Semits), 
         by = File) 
```

wherein the standard deviation of `semits` is, again, calculated for each file,
but this time wherever the `semits` field is greater than the mean value *across all the data*.

[humTable]: R:humTable
[humTable]: R:humTable
[complement]: R:complement
[partialMatching]: R:partialMatching
[data.table]: R:data.table
[tapply]: R:tapply
[by]: R:by
[table]: R:table

Windowing data

XXXX

Plotting parameters

As mentioned above, plots in within-expressions should (often) be called using the sidefx argument name. When plotting, with/within.humdrumR also allows you to specify plotting options inline, without having to make a separate call to par(). Any par() argument can be specified by providing a named list to the graphics keyword. For example, we can set the plot margins with the mar argument:

within(data, 
       side = plot(sort(table(Token))), 
       graphics = list(mar = c(4, 4, 4, 4)))

The best part is with/within.humdrumR will reset par to it's previous state after its done.

You can also use the syntactic sugar, graphics(parargs = ...):

within(data,
       side = plot(sort(nchar(Token))),
       graphics(mar = c(4, 4, 4, 4)))

Splatting

("Splatting" refers to feeding a function a list/vector of arguments.) Sometimes we want to divide our data into pieces (a l\'a partition option), but rather than applying the same expression to each piece, we want to feed the separate pieces as separate arguments to the same function. In with/within.humdrumR you can use some syntactic sugar to do just this. We can index any field in our call with a splat argument, which must be a Field %in% x. For example,

within(humdata, list(Token[splat = Spine %in% 1:2]))

In this call, the Token field will be divided into two groups, one where Spine == 1 and the other where Spine == 2; the first group (Spine == 1) will be used as the first argument to list, and the second group (Spine == 2) as the second argument. Thus, within translates the previous expression to this:

within(humdata,
       list(Token[Spine == 1], Token[Spine == 2]))

N grams

XXXX

Advanced scripting

with.humdrumR and within.humdrumR use non-standard evaluation of their expressions. This is very useful on the command line or in a script running one command at a time. However, if you want to more advanced scripting non-standard evaluation can be a problem. For example, if you to loop through a list of within-expressions or reuse a common combination coor arguments many times.

Fortunately, R has formula, which are a way of capturing ("quoting") expressions into a concrete object that you can manipulate. Better yet, with/within.humdrumR will interpret formulae passed to them as arguments. Basically, the right-hand side of any formula is interpreted as an expression to evaluate. If a formula is passed as a unnamed argument to with/within, the left-hand side of the formula (if any) is treated as the argument name. If you name a formula argument that has a left-hand side, the left-hand side is ignored.

with(humdrumR, table(Token), by = Spine)
with(humdrumR, ~table(Token), by ~ Spine)

This approach would allows us to, same ~table(Token) or by ~ Spine as variables, allowing us to do things like:

tabler <- ~table(.)
byspine <- by ~ Spine

with(humData, tabler, byspine)

We can even make lists and loop through them:

bys <- list(~ Spine, ~ File, ~ COM)

for (b in bys) with(humData, table(.), by = b)

Variables

Another useful scripting option is to include free variables in your within-expressions, which can then be fed in using the variables argument. variables must be a named list. Variables in the within-expression which match a name of from the variables list have that value interpolated. This allows us, for example, to run commands like:

nums <- c(2, 3, 4, 5)
expression <- ~.^N

for (n in nums) with(humData, expression, variables = list(N = n))

Each time with is called, the N in .^N is replaced by the current value of n.