Write humdrumR data to humdrum files

writeHumdrum writes humdrumR data into humdrum-syntax text files. The current active expression is evaluated to generate the humdrum output data. The written output should match the printout if printing the data in the R terminal.

Usage

writeHumdrum(
  humdrumR,
  prefix = "humdrumR_",
  renamer = force,
  affix = "",
  extension = NULL,
  directory = NULL,
  overwrite = FALSE,
  verbose = FALSE,
  EMD = paste0("Edited in humdrumR ", packageVersion("humdrumR"), " on ", Sys.Date())
)

Arguments

humdrumR

A humdrumR data object.

prefix

(character, length == 1L) A character string prepended to the file name.

renamer

A function. Must accept a character vector and return a new atomic vector of the same length.

affix

(character, length == 1L) A character string appended to the file name (before the extension).

directory

(character, length == 1L) A directory to write the files into. If NULL, files are written to the same directory (or directories) they were read from.

overwrite

(logical, length == 1) If FALSE, writeHumdrum will refuse to overwrite any files. If TRUE, writeHumdrum will overwrite files, but only after an additional prompt from the user.

verbose

(logical, length == 1) If TRUE, each new output file name is printed on the console as the writing happens.

EMD

(character, length == 1) A string to write to a new !!!EMD: record in each file. If NULL, not appended.

File names

The main option to control with writeHumdrum is what files to write to. writeHumdrum uses the original names of the files, as read by readHumdrum, as the basis for new file names. By default, writeHumdrum will refuse to overwrite the original files---overwriting will only be allowed if you specify overwrite == TRUE and respond with "y" to a prompt.

writeHumdrum generates new file names by modifying the original read file names. The renamer argument must be a function which takes the original names as an input character vector (excluding the directory path and the file extension) and returns a new character vector of the same length (the default is R's identity function force). After running renamer, the character-string affix and prefix arguments are appended/prepended to the renamed names. (The affix is affixed before the extension.) Finally, the extension argument can be used to specify a different file extension.

The directory argument indicates a file path where to write the files. If the directory doesn't exist, it is created. If directory is NULL, files are written to their original input directory (or directories).

The EMD argument specifies a character string to put into a new !!!EMD reference record at the end of each file. EMD records keep track of modifications to humdrum data. The default behavior is to print a string indicating the humdrumR version number and date. If EMD is set to NULL, it is not appended to the files.